Ouroboros.Network.Block

Description

Abstract view over blocks

The network layer does not make any concrete assumptions about what blocks look like.

Documentation

newtype SlotNo

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

newtype BlockNo

Constructors

BlockNo

Fields

unBlockNo ∷ Word64

type family HeaderHash b ∷ Type

Header hash

data HeaderFields b

Header fields we expect to be present in a block

These fields are lazy because they are extracted from a block or block header; this type is not intended for storage.

Constructors

HeaderFields

Fields

headerFieldSlot ∷ SlotNo

headerFieldBlockNo ∷ BlockNo

headerFieldHash ∷ HeaderHash b

NOTE: this field is last so that the derived Eq and Ord instances first compare the slot and block numbers, which is cheaper than comparing hashes.

castHeaderFields ∷ HeaderHash b ~ HeaderHash b' ⇒ HeaderFields b → HeaderFields b'

class (StandardHash b, Typeable b) ⇒ HasHeader b where

Abstract over the shape of blocks (or indeed just block headers)

Methods

getHeaderFields ∷ b → HeaderFields b

blockNo ∷ HasHeader b ⇒ b → BlockNo

blockSlot ∷ HasHeader b ⇒ b → SlotNo

blockHash ∷ HasHeader b ⇒ b → HeaderHash b

class HasHeader b ⇒ HasFullHeader b where

Extension of HasHeader with some additional information

Used in tests and assertions only.

Methods

blockPrevHash ∷ b → ChainHash b

blockInvariant ∷ b → Bool

class (Eq (HeaderHash b), Ord (HeaderHash b), Show (HeaderHash b), Typeable (HeaderHash b), NoThunks (HeaderHash b)) ⇒ StandardHash b

StandardHash summarises the constraints we want header hashes to have

Without this class we would need to write

deriving instance Eq (HeaderHash block) => Eq (ChainHash block)`

That requires UndecidableInstances; not a problem by itself, but it also means that we can then not use deriving Eq anywhere else for datatypes that reference Hash, which is very frustrating; see

https://downloads.haskell.org/~ghc/latest/docs/html/users_guide/glasgow_exts.html#inferred-context-for-deriving-clauses

Introducing the StandardHash class avoids this problem.

Having these constraints directly as part of the HasHeader class is possible but libraries that use the networking layer may wish to be able to talk about StandardHash independently of HasHeader since the latter may impose yet further constraints.

data ChainHash b

Constructors

GenesisHash

BlockHash !(HeaderHash b)

castHash ∷ Coercible (HeaderHash b) (HeaderHash b') ⇒ ChainHash b → ChainHash b'

newtype Point block

A point on the chain is identified by its Slot and HeaderHash.

The Slot tells us where to look and the HeaderHash either simply serves as a check, or in some contexts it disambiguates blocks from different forks that were in the same slot.

It's a newtype rather than a type synonym, because using a type synonym would lead to ambiguity, since HeaderHash is a non-injective type family.

Constructors

Point

Fields

getPoint ∷ WithOrigin (Block SlotNo (HeaderHash block))

pointSlot ∷ Point block → WithOrigin SlotNo

pointHash ∷ Point block → ChainHash block

castPoint ∷ Coercible (HeaderHash b) (HeaderHash b') ⇒ Point b → Point b'

blockPoint ∷ HasHeader block ⇒ block → Point block

pattern GenesisPoint ∷ Point block

pattern BlockPoint ∷ SlotNo → HeaderHash block → Point block

atSlot ∷ Point block → SlotNo

withHash ∷ Point block → HeaderHash block

data Tip b

Used in chain-sync protocol to advertise the tip of the server's chain.

Constructors

TipGenesis
The tip is genesis

Tip !SlotNo !(HeaderHash b) !BlockNo
The tip is not genesis

castTip ∷ HeaderHash a ~ HeaderHash b ⇒ Tip a → Tip b

The equivalent of castPoint for Tip

getTipPoint ∷ Tip b → Point b

getTipBlockNo ∷ Tip b → WithOrigin BlockNo

getTipSlotNo ∷ Tip b → WithOrigin SlotNo

getLegacyTipBlockNo ∷ Tip b → BlockNo

Deprecated: Use getTipBlockNo

Get the block number associated with a Tip, or genesisBlockNo otherwise

TODO: This is wrong. There is no block number if we are at genesis (genesisBlockNo is the block number of the first block on the chain). Usage of this function should be phased out.

tipFromHeader ∷ HasHeader a ⇒ a → Tip a

legacyTip ∷ Point b → BlockNo → Tip b

Deprecated: Use tipFromHeader instead

Inverse of toLegacyTip

TODO: This should be phased out, since it makes no sense to have a BlockNo for the genesis point.

toLegacyTip ∷ Tip b → (Point b, BlockNo)

Deprecated: Use getTipPoint and getTipBlockNo

Translate to the format it was before (to maintain binary compatibility)

encodeTip ∷ (HeaderHash blk → Encoding) → Tip blk → Encoding

encodedTipSize ∷ (Proxy (HeaderHash blk) → Size) → Proxy (Tip blk) → Size

decodeTip ∷ ∀ blk. (∀ s. Decoder s (HeaderHash blk)) → ∀ s. Decoder s (Tip blk)

data ChainUpdate block a

A representation of two actions to update a chain: add a block or roll back to a previous point.

The type parameter a is there to allow a Functor instance. Typically, it will be instantiated with block itself.

Constructors

AddBlock a

RollBack (Point block)

data MaxSlotNo

The highest slot number seen.

Constructors

NoMaxSlotNo
No block/header has been seen yet, so we don't have a highest slot number.

MaxSlotNo !SlotNo
The highest slot number seen.

maxSlotNoFromMaybe ∷ Maybe SlotNo → MaxSlotNo

maxSlotNoToMaybe ∷ MaxSlotNo → Maybe SlotNo

maxSlotNoFromWithOrigin ∷ WithOrigin SlotNo → MaxSlotNo

genesisPoint ∷ Point block

Serialisation

encodePoint ∷ (HeaderHash block → Encoding) → Point block → Encoding

encodedPointSize ∷ (Proxy (HeaderHash block) → Size) → Proxy (Point block) → Size

encodeChainHash ∷ (HeaderHash block → Encoding) → ChainHash block → Encoding

decodePoint ∷ (∀ s. Decoder s (HeaderHash block)) → ∀ s. Decoder s (Point block)

decodeChainHash ∷ (∀ s. Decoder s (HeaderHash block)) → ∀ s. Decoder s (ChainHash block)

Serialised block/header

newtype Serialised a

An already serialised value

When streaming blocks/header from disk to the network, there is often no need to deserialise them, as we'll just end up serialising them again when putting them on the wire.

Constructors

Serialised

Fields

unSerialised ∷ ByteString

wrapCBORinCBOR ∷ (a → Encoding) → a → Encoding

Wrap CBOR-in-CBOR

This is primarily useful for the decoder; see unwrapCBORinCBOR

unwrapCBORinCBOR ∷ (∀ s. Decoder s (ByteString → a)) → ∀ s. Decoder s a

Unwrap CBOR-in-CBOR

The CBOR-in-CBOR encoding gives us the ByteString we need in order to to construct annotations.

mkSerialised ∷ (a → Encoding) → a → Serialised a

Construct Serialised value from an unserialised value

fromSerialised ∷ (∀ s. Decoder s (ByteString → a)) → Serialised a → ∀ s. Decoder s a

Decode a Serialised value

Unlike a regular Decoder, which has an implicit input stream, fromSerialised takes the Serialised value as an argument.