Ouroboros.Network.BlockFetch.Client

Block fetch protocol client implementation

blockFetchClient ∷ ∀ header block m void. (MonadSTM m, MonadTime m, MonadThrow m, HasHeader header, HasHeader block, HeaderHash header ~ HeaderHash block) ⇒ FetchClientContext header block m → PeerPipelined (BlockFetch block) AsClient BFIdle m void

The implementation of the client side of block fetch protocol designed to work in conjunction with our fetch logic.

type BlockFetchClient header block m a = FetchClientContext header block m → PeerPipelined (BlockFetch block) AsClient BFIdle m a

TODO: use a fetch client wrapper type rather than the raw PeerPipelined, and eliminate this alias. It is only here to avoid large types leaking into the consensus layer.

data FetchClientContext header block m

The context that is passed into the block fetch protocol client when it is started.

data TraceFetchClientState header

Tracing types for the various events that change the state (i.e. FetchClientStateVars) for a block fetch client.

Note that while these are all state changes, the AddedFetchRequest occurs in the decision thread while the other state changes occur in the block fetch client threads.

newtype FetchRequest header

Constructors

FetchRequest

Fields

fetchRequestFragments ∷ [ChainFragment header]

data FetchClientStateVars m header

A set of variables shared between the block fetch logic thread and each thread executing the client side of the block fetch protocol. That is, these are the shared variables per peer. The FetchClientRegistry contains the mapping of these for all peers.

The variables are used for communicating from the protocol thread to the decision making thread the status of things with that peer. And in the other direction one shared variable is for providing new fetch requests.

Exception types

data BlockFetchProtocolFailure