Ouroboros.Network.BlockFetch.ClientRegistry

Registry of block fetch clients

data FetchClientRegistry peer header block m

A registry for the threads that are executing the client side of the BlockFetch protocol to communicate with our peers.

The registry contains the shared variables we use to communicate with these threads, both to track their status and to provide instructions.

The threads add/remove themselves to/from this registry when they start up and shut down.

Constructors

FetchClientRegistry

Fields

fcrCtxVar :: StrictTMVar m (Tracer m (TraceLabelPeer peer (TraceFetchClientState header)), WhetherReceivingTentativeBlocks -> STM m (FetchClientPolicy header block m))

fcrFetchRegistry :: StrictTVar m (Map peer (FetchClientStateVars m header))

fcrSyncRegistry :: StrictTVar m (Map peer (ThreadId m, StrictTMVar m (), StrictTMVar m ()))

fcrDqRegistry :: StrictTVar m (Map peer PeerGSV)

fcrKeepRegistry :: StrictTVar m (Map peer (ThreadId m, StrictTMVar m ()))

fcrDying :: StrictTVar m (Set peer)

newFetchClientRegistry :: MonadSTM m => m (FetchClientRegistry peer header block m)

bracketFetchClient

Arguments

:: forall m a peer header block version. (MonadSTM m, MonadFork m, MonadMask m, Ord peer)

=> FetchClientRegistry peer header block m

-> version

-> (version -> WhetherReceivingTentativeBlocks)
is pipelining enabled function

-> peer

-> (FetchClientContext header block m -> m a)

-> m a

This is needed to start a block fetch client. It provides the required FetchClientContext. It registers and unregisters the fetch client on start and end.

It also manages synchronisation with the corresponding chain sync client.

bracketKeepAliveClient :: forall m a peer header block. (MonadSTM m, MonadFork m, MonadMask m, Ord peer) => FetchClientRegistry peer header block m -> peer -> (StrictTVar m (Map peer PeerGSV) -> m a) -> m a

bracketSyncWithFetchClient :: forall m a peer header block. (MonadSTM m, MonadFork m, MonadCatch m, Ord peer) => FetchClientRegistry peer header block m -> peer -> m a -> m a

The block fetch and chain sync clients for each peer need to synchronise their startup and shutdown. This bracket operation provides that synchronisation for the chain sync client.

This must be used for the chain sync client outside of its own state registration and deregistration.

setFetchClientContext :: MonadSTM m => FetchClientRegistry peer header block m -> Tracer m (TraceLabelPeer peer (TraceFetchClientState header)) -> (WhetherReceivingTentativeBlocks -> STM m (FetchClientPolicy header block m)) -> m ()

data FetchClientPolicy header block m

The policy used by the fetch clients. It is set by the central block fetch logic, and passed to them via the FetchClientRegistry.

Constructors

FetchClientPolicy

Fields

blockFetchSize :: header -> SizeInBytes

blockMatchesHeader :: header -> block -> Bool

addFetchedBlock :: Point block -> block -> m ()

blockForgeUTCTime :: FromConsensus block -> STM m UTCTime

readFetchClientsStatus :: MonadSTM m => FetchClientRegistry peer header block m -> STM m (Map peer (PeerFetchStatus header))

A read-only STM action to get the current PeerFetchStatus for all fetch clients in the FetchClientRegistry.

readFetchClientsStateVars :: MonadSTM m => FetchClientRegistry peer header block m -> STM m (Map peer (FetchClientStateVars m header))

A read-only STM action to get the FetchClientStateVars for all fetch clients in the FetchClientRegistry.

readPeerGSVs :: forall block header m peer. (MonadSTM m, Ord peer) => FetchClientRegistry peer header block m -> STM m (Map peer PeerGSV)

A read-only STM action to get the PeerGSVs for all fetch clients in the FetchClientRegistry.