Ouroboros.Network.ChainFragment

ChainFragment type and fundamental operations

newtype ChainFragment block

A fragment of the chain.

The chain grows to the right. The oldest block is the left-most block and the newest block is the right-most block.

Invariant: a chain fragment should never contain the origin point. That is the point reserved for genesis.

A fragment is represented by a finger tree for efficient searching based on the SlotNo (or Point) of a block.

Constructors

ChainFragment (StrictFingerTree BlockMeasure block)

Bundled Patterns

pattern Empty ∷ HasHeader block ⇒ ChainFragment block

pattern (:>) ∷ (HasHeader block, HasCallStack) ⇒ ChainFragment block → block → ChainFragment block infixl 5
\( O(1) \). Add a block to the right of the chain fragment.

pattern (:<) ∷ HasHeader block ⇒ block → ChainFragment block → ChainFragment block infixl 5
\( O(1) \). Add a block to the left of the chain fragment.

valid ∷ HasHeader block ⇒ ChainFragment block → Bool

\( O(n) \).

validExtension ∷ (HasHeader block, HasCallStack) ⇒ ChainFragment block → block → Bool

\( O(1) \).

validExtension' ∷ (HasHeader block, HasCallStack) ⇒ ChainFragment block → block → Either String ()

Variation on validExtension that provides more information

isValidSuccessorOf

Arguments

∷ (HasCallStack, HasHeader block)

⇒ block
bSucc

→ block
b

→ Bool

Checks whether the first block bSucc is a valid successor of the second block b.

The blockPrevHash of the bSucc must match that of b.

The blockSlot of bSucc must be strictly larger than that of b.

The blockNo of bSucc must be 1 greater than that of b.

This function does not check whether any of the two blocks satisfy blockInvariant.

isValidSuccessorOf'

Arguments

∷ (HasCallStack, HasHeader block)

⇒ block
bSucc

→ block
b

→ Either String ()

Variation on isValidSuccessorOf that provides more information

foldChainFragment ∷ HasHeader block ⇒ (a → block → a) → a → ChainFragment block → a

\( O(n) \). Fold a ChainFragment.

Implemented as a strict left fold.

mapChainFragment ∷ (HasHeader block1, HasHeader block2) ⇒ (block1 → block2) → ChainFragment block1 → ChainFragment block2

\( O(n) \). Maps over the chain blocks. This is not allowed to change the block Points, or it would create an invalid chain.

Block re-exports

class (StandardHash b, Measured BlockMeasure b, Typeable b) ⇒ HasHeader b where

Abstract over the shape of blocks (or indeed just block headers)

Methods

blockHash ∷ b → HeaderHash b

blockPrevHash ∷ b → ChainHash b

blockSlot ∷ b → SlotNo

blockNo ∷ b → BlockNo

blockInvariant ∷ b → Bool

newtype Point block

A point on the chain is identified by its Slot and HeaderHash.

The Slot tells us where to look and the HeaderHash either simply serves as a check, or in some contexts it disambiguates blocks from different forks that were in the same slot.

It's a newtype rather than a type synonym, because using a type synonym would lead to ambiguity, since HeaderHash is a non-injective type family.

Constructors

Point

Fields

getPoint ∷ WithOrigin (Block SlotNo (HeaderHash block))

castPoint ∷ HeaderHash a ~ HeaderHash b ⇒ Point a → Point b

blockPoint ∷ HasHeader block ⇒ block → Point block

ChainFragment construction and inspection

Head inspection

headPoint ∷ HasHeader block ⇒ ChainFragment block → Maybe (Point block)

\( O(1) \).

headSlot ∷ HasHeader block ⇒ ChainFragment block → Maybe SlotNo

\( O(1) \).

headHash ∷ HasHeader block ⇒ ChainFragment block → Maybe (ChainHash block)

\( O(1) \).

headBlockNo ∷ HasHeader block ⇒ ChainFragment block → Maybe BlockNo

\( O(1) \).

Basic operations

head ∷ HasHeader block ⇒ ChainFragment block → Maybe block

\( O(1) \).

last ∷ HasHeader block ⇒ ChainFragment block → Maybe block

\( O(1) \).

lastPoint ∷ HasHeader block ⇒ ChainFragment block → Maybe (Point block)

\( O(1) \).

lastSlot ∷ HasHeader block ⇒ ChainFragment block → Maybe SlotNo

\( O(1) \).

toNewestFirst ∷ HasHeader block ⇒ ChainFragment block → [block]

TODO. Make a list of blocks from a ChainFragment, in newest-to-oldest order.

toOldestFirst ∷ HasHeader block ⇒ ChainFragment block → [block]

\( O(n) \). Make a list of blocks from a ChainFragment, in oldest-to-newest order.

fromNewestFirst ∷ HasHeader block ⇒ [block] → ChainFragment block

\( O(n) \). Make a ChainFragment from a list of blocks in newest-to-oldest order.

fromOldestFirst ∷ HasHeader block ⇒ [block] → ChainFragment block

\( O(n) \). Make a ChainFragment from a list of blocks in oldest-to-newest order.

dropNewest

Arguments

∷ HasHeader block

⇒ Int
n

→ ChainFragment block

→ ChainFragment block

\( O(\log(\min(i,n-i)) \). Drop the newest n blocks from the ChainFragment.

dropOldest

Arguments

∷ HasHeader block

⇒ Int
n

→ ChainFragment block

→ ChainFragment block

\( O(\log(\min(i,n-i)) \). Drop the oldest n blocks from the ChainFragment.

takeNewest

Arguments

∷ HasHeader block

⇒ Int
n

→ ChainFragment block

→ ChainFragment block

\( O(\log(\min(i,n-i)) \). Take the newest n blocks from the ChainFragment.

takeOldest

Arguments

∷ HasHeader block

⇒ Int
n

→ ChainFragment block

→ ChainFragment block

\( O(\log(\min(i,n-i)) \). Take the oldest n blocks from the ChainFragment.

takeWhileNewest ∷ HasHeader block ⇒ (block → Bool) → ChainFragment block → ChainFragment block

\( O(n) \). Select the newest blocks that satisfy the predicate.

dropWhileNewest ∷ HasHeader block ⇒ (block → Bool) → ChainFragment block → ChainFragment block

\( O(n) \). Drop the newest blocks that satisfy the predicate, keeping the remainder.

takeWhileOldest ∷ HasHeader block ⇒ (block → Bool) → ChainFragment block → ChainFragment block

length ∷ HasHeader block ⇒ ChainFragment block → Int

\( O(1) \).

null ∷ ChainFragment block → Bool

\( O(1) \).

Update type and operations

data ChainUpdate block a

A representation of two actions to update a chain: add a block or roll back to a previous point.

The type parameter a is there to allow a Functor instance. Typically, it will be instantiated with block itself.

Constructors

AddBlock a

RollBack (Point block)

addBlock ∷ HasHeader block ⇒ block → ChainFragment block → ChainFragment block

\( O(1) \). Add a block to the right of the chain fragment.

Synonym for :>.

rollback ∷ HasHeader block ⇒ Point block → ChainFragment block → Maybe (ChainFragment block)

\( O(\log(\min(i,n-i)) \). If the Point is in the ChainFragment, roll back to a ChainFragment such that its last Point is the given Point.

In other words, remove blocks from the end of the ChainFragment until the given Point is the last block. If the given Point is not part of the ChainFragment, return Nothing.

applyChainUpdate ∷ HasHeader block ⇒ ChainUpdate block block → ChainFragment block → Maybe (ChainFragment block)

applyChainUpdates ∷ HasHeader block ⇒ [ChainUpdate block block] → ChainFragment block → Maybe (ChainFragment block)

Special operations

pointOnChainFragment ∷ HasHeader block ⇒ Point block → ChainFragment block → Bool

\( O(\log(\min(i,n-i)) + s \) where s is the number of blocks with the same slot.

successorBlock ∷ HasHeader block ⇒ Point block → ChainFragment block → Maybe block

\( O(\log(\min(i,n-i)) \). Find the block after the given point.

splitAfterPoint ∷ (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ ChainFragment block1 → Point block2 → Maybe (ChainFragment block1, ChainFragment block1)

\( O(\log(\min(i,n-i)) + s \) where s is the number of blocks with the same slot. Split the ChainFragment after the block at the given Point. Return Nothing if the ChainFragment does not contain a block at the given Point.

If the chain fragment contained a block at the given Point, it will be the (newest/rightmost) block of the first returned chain.

splitBeforePoint ∷ (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ ChainFragment block1 → Point block2 → Maybe (ChainFragment block1, ChainFragment block1)

\( O(\log(\min(i,n-i)) + s \) where s is the number of blocks with the same slot. Split the ChainFragment before the block at the given Point. Return Nothing if the ChainFragment does not contain a block at the given Point.

If the chain fragment contained a block at the given Point, it will be the (oldest/leftmost) block of the second returned chain.

sliceRange ∷ HasHeader block ⇒ ChainFragment block → Point block → Point block → Maybe (ChainFragment block)

Select a slice of a chain fragment between two points, inclusive.

Both points must exist on the chain, in order, or the result is Nothing.

lookupByIndexFromEnd ∷ HasHeader block ⇒ ChainFragment block → Int → SearchResult BlockMeasure block

\( O(\log(\min(i,n-i)) \). Look up a block in the ChainFragment based on the given index, i.e. the offset starting from the newest/rightmost block.

Note that search used to contain a bug, but this has been fixed in version 0.1.4.2 of the fingertree library. See https://hub.darcs.net/ross/fingertree/issue/8.

data SearchResult v a

A result of search, attempting to find a point where a predicate on splits of the sequence changes from False to True.

Since: cardano-prelude-0.1.2.0

Constructors

Position !(StrictFingerTree v a) !a !(StrictFingerTree v a)
A tree opened at a particular element: the prefix to the left, the element, and the suffix to the right.

OnLeft
A position to the left of the sequence, indicating that the predicate is True at both ends.

OnRight
A position to the right of the sequence, indicating that the predicate is False at both ends.

Nowhere
No position in the tree, returned if the predicate is True at the left end and False at the right end. This will not occur if the predicate in monotonic on the tree.

filter ∷ HasHeader block ⇒ (block → Bool) → ChainFragment block → [ChainFragment block]

\( O\(n\) ). Filter the chain based on a predicate. As filtering removes blocks the result is a sequence of disconnected fragments. The fragments are in the original order and are of maximum size.

filterWithStop

Arguments

∷ HasHeader block

⇒ (block → Bool)
Filtering predicate

→ (block → Bool)
Stop condition

→ ChainFragment block

→ [ChainFragment block]

\( O\(n\) ). Same as filter, but as soon as the stop condition is true, the filtering stops and the remaining fragment (starting with the first element for which the stop condition is true) is the final fragment in the returned list.

The stop condition wins from the filtering predicate: if the stop condition is true for an element, but the filter predicate not, then the element still ends up in final fragment.

For example, given the fragment containing [1, 2, 3, 4, 5, 6]:

filter         odd        -> [[1], [3], [5]]
filterWithStop odd (>= 4) -> [[1], [3], [4, 5, 6]]

selectPoints ∷ HasHeader block ⇒ [Int] → ChainFragment block → [Point block]

\( O(o \log(\min(i,n-i))) \). Select a bunch of Points based on offsets from the head of the chain fragment. This is used in the chain consumer protocol as part of finding the intersection between a local and remote chain.

The list of offsets must be increasing monotonically.

The typical pattern is to use a selection of offsets covering the last K blocks, biased towards more recent blocks. For example:

selectPoints (0 : [ fib n | n <- [1 .. 17] ])

Only for offsets within the bounds of the chain fragment, will there be points in the returned list.

findFirstPoint ∷ HasHeader block ⇒ [Point block] → ChainFragment block → Maybe (Point block)

\( O(p (\log(\min(i,n-i)) \). Find the first Point in the list of points that is on the given ChainFragment. Return Nothing if none of them are on the ChainFragment. TODO test?

intersectChainFragments ∷ (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ ChainFragment block1 → ChainFragment block2 → Maybe (ChainFragment block1, ChainFragment block2, ChainFragment block1, ChainFragment block2)

\( O(n_2 \log(n_1)) \). Look for the intersection of the two ChainFragments c1 and c2.

If they intersect, i.e., share a common Point, then return a tuple of:

l1: the prefix of the first chain fragment

l2: The prefix of the second chain fragment

r1: The suffix of the first chain fragment

r2: The suffix of the second chain fragment

l1 and l2 will have same last block, i.e. head. The original chain c1 can be obtained by putting r1 after l1, similarly for c2: by putting r2 after l2:

Just c1 = joinChainFragments l1 r1
Just c1 = joinChainFragments l1 r1

Chains that intersect will always have the exact same common prefix, but chain fragments might have prefixes that differ in length. The blocks they have in common will be the same, but not each fragment might contain all the blocks the other fragment contains.

Take for example the following two chain fragments that share blocks 4 and 5. The two chain fragments are fragments of the same chain, but don't contain all blocks of the original chain. The missing history of the fragments is indicated with an asterisk (*). The -A and -B suffixes denote that blocks are part of a fork of the chain.

    ┆ * ┆
    ├───┤
    │ 2 │     ┆ * ┆
    ├───┤     ├───┤
    │ 4 │     │ 4 │
    ├───┤     ├───┤
    │ 5 │     │ 5 │
────┼───┼─────┼───┼───
    │ 6A│     │ 6B│
    └───┘     ├───┤
              │ 8B│
              └───┘
      c1        c2

The intersection of c1 and c2 is block 5 (the last Point the two fragments have in common) and we return the following fragments:

    ┆ * ┆
    ├───┤
    │ 2 │     ┆ * ┆
    ├───┤     ├───┤
    │ 4 │     │ 4 │
    ├───┤     ├───┤
    │ 5 │     │ 5 │
────┴───┴─────┴───┴──────┬───┬─────┬───┬───
                         │ 6A│     │ 6B│
                         └───┘     ├───┤
                                   │ 8B│
                                   └───┘
Just (l1,       l2,        r1,       r2)

isPrefixOf ∷ (HasHeader block, Eq block) ⇒ ChainFragment block → ChainFragment block → Bool

\( O(\max(n_1, n_2)) \). Check whether the first chain fragment is a prefix of the second.

joinChainFragments ∷ HasHeader block ⇒ ChainFragment block → ChainFragment block → Maybe (ChainFragment block)

\( O(\log(\min(n_1, n_2))) \). Join two ChainFragments if the first (oldest) block of the second fragment is the successor of the last (newest) block of the first fragment.

Helper functions

prettyPrintChainFragment ∷ HasHeader block ⇒ String → (block → String) → ChainFragment block → String

Reference implementations for testing

foldChainFragmentSpec ∷ HasHeader block ⇒ (a → block → a) → a → ChainFragment block → a

\( O(n) \). Specification of foldChainFragment.

Use foldChainFragment, as it should be faster.

This function is used to verify whether foldChainFragment behaves as expected.

pointOnChainFragmentSpec ∷ HasHeader block ⇒ Point block → ChainFragment block → Bool

\( O(n) \). Specification of pointOnChainFragment.

Use pointOnChainFragment, as it should be faster.

This function is used to verify whether pointOnChainFragment behaves as expected.

selectPointsSpec ∷ HasHeader block ⇒ [Int] → ChainFragment block → [Point block]

\( O(o * n) \). Specification of selectPoints.

Use selectPoints, as it should be faster.

This function is used to verify whether selectPoints behaves as expected.