Ouroboros.Network.Diffusion

Common API

data P2P

Promoted data types.

Constructors

P2P

NonP2P

data DiffusionTracer ntnAddr ntcAddr

The DiffusionTracer logs

diffusion initialisation messages

terminal errors thrown by diffusion

Constructors

RunServer (NonEmpty ntnAddr)

RunLocalServer ntcAddr

UsingSystemdSocket ntcAddr

CreateSystemdSocketForSnocketPath ntcAddr

CreatedLocalSocket ntcAddr

ConfiguringLocalSocket ntcAddr FileDescriptor

ListeningLocalSocket ntcAddr FileDescriptor

LocalSocketUp ntcAddr FileDescriptor

CreatingServerSocket ntnAddr

ConfiguringServerSocket ntnAddr

ListeningServerSocket ntnAddr

ServerSocketUp ntnAddr

UnsupportedLocalSystemdSocket ntnAddr

UnsupportedReadySocketCase

DiffusionErrored SomeException

SystemdSocketConfiguration SystemdSocketTracer

data Tracers ntnAddr ntnVersion ntcAddr ntcVersion m

Common DiffusionTracers interface between P2P and NonP2P

Constructors

Tracers

Fields

dtMuxTracer :: Tracer m (WithMuxBearer (ConnectionId ntnAddr) MuxTrace)

Mux tracer

dtHandshakeTracer :: Tracer m (HandshakeTr ntnAddr ntnVersion)

Handshake protocol tracer

dtLocalMuxTracer :: Tracer m (WithMuxBearer (ConnectionId ntcAddr) MuxTrace)

Mux tracer for local clients

dtLocalHandshakeTracer :: Tracer m (HandshakeTr ntcAddr ntcVersion)

Handshake protocol tracer for local clients

dtDiffusionTracer :: Tracer m (DiffusionTracer ntnAddr ntcAddr)

Diffusion initialisation tracer

nullTracers :: Applicative m => Tracers ntnAddr ntnVersion ntcAddr ntcVersion m

data ExtraTracers (p2p :: P2P) where

Tracers which depend on p2p mode.

Constructors

P2PTracers :: TracersExtra RemoteAddress NodeToNodeVersion NodeToNodeVersionData LocalAddress NodeToClientVersion NodeToClientVersionData IOException IO -> ExtraTracers 'P2P

NonP2PTracers :: TracersExtra -> ExtraTracers 'NonP2P

data Failure where

Constructors

UnsupportedReadySocket :: Failure

UnexpectedIPv4Address :: forall ntnAddr. (Show ntnAddr, Typeable ntnAddr) => ntnAddr -> Failure

UnexpectedIPv6Address :: forall ntnAddr. (Show ntnAddr, Typeable ntnAddr) => ntnAddr -> Failure

NoSocket :: Failure

DiffusionError :: SomeException -> Failure

data Arguments ntnFd ntnAddr ntcFd ntcAddr

Common DiffusionArguments interface between P2P and NonP2P

Constructors

Arguments

Fields

daIPv4Address :: Maybe (Either ntnFd ntnAddr)

an IPv4 socket ready to accept connections or an IPv4 addresses

daIPv6Address :: Maybe (Either ntnFd ntnAddr)

an IPv6 socket ready to accept connections or an IPv6 addresses

daLocalAddress :: Maybe (Either ntcFd ntcAddr)

an AF_UNIX socket ready to accept connections or an AF_UNIX socket path

daAcceptedConnectionsLimit :: AcceptedConnectionsLimit

parameters for limiting number of accepted connections

daMode :: DiffusionMode

run in initiator only mode

data ExtraArguments (p2p :: P2P) m where

Diffusion arguments which depend on p2p mode.

Constructors

P2PArguments :: ArgumentsExtra m -> ExtraArguments 'P2P m

NonP2PArguments :: ArgumentsExtra -> ExtraArguments 'NonP2P m

data Applications ntnAddr ntnVersion ntnVersionData ntcAddr ntcVersion ntcVersionData m a

Versioned mini-protocol bundles run on a negotiated connection.

Constructors

Applications

Fields

daApplicationInitiatorMode :: Versions ntnVersion ntnVersionData (OuroborosBundleWithExpandedCtx InitiatorMode ntnAddr ByteString m a Void)

NodeToNode initiator applications for initiator only mode.

TODO: we should accept one or the other, but not both: daApplicationInitiatorMode, daApplicationInitiatorResponderMode.

Even in non-p2p mode we use p2p apps.

daApplicationInitiatorResponderMode :: (PeerSharingAmount -> m [ntnAddr]) -> Versions ntnVersion ntnVersionData (OuroborosBundleWithExpandedCtx InitiatorResponderMode ntnAddr ByteString m a ())

NodeToNode initiator & responder applications for bidirectional mode.

daLocalResponderApplication :: Versions ntcVersion ntcVersionData (OuroborosApplicationWithMinimalCtx ResponderMode ntcAddr ByteString m Void ())

NodeToClient responder application (server role)

Because p2p mode does not infect local connections we we use non-p2p apps.

daLedgerPeersCtx :: LedgerPeersConsensusInterface m

Interface used to get peers from the current ledger.

TODO: it should be in InterfaceExtra

data ExtraApplications (p2p :: P2P) ntnAddr m a where

Application data which depend on p2p mode.

Constructors

P2PApplications :: ApplicationsExtra ntnAddr m a -> ExtraApplications 'P2P ntnAddr m a

NonP2PApplications :: ApplicationsExtra -> ExtraApplications 'NonP2P ntnAddr m a

Run data diffusion

run :: forall (p2p :: P2P) a. Tracers RemoteAddress NodeToNodeVersion LocalAddress NodeToClientVersion IO -> ExtraTracers p2p -> Arguments Socket RemoteAddress LocalSocket LocalAddress -> ExtraArguments p2p IO -> Applications RemoteAddress NodeToNodeVersion NodeToNodeVersionData LocalAddress NodeToClientVersion NodeToClientVersionData IO a -> ExtraApplications p2p RemoteAddress IO a -> IO ()

Run data diffusion in either P2P or NonP2P mode.

Re-exports

type AbstractTransitionTrace peerAddr = TransitionTrace' peerAddr AbstractState