Ouroboros.Network.ExitPolicy

Documentation

newtype RepromoteDelay

After demoting a peer to Warm or Cold, we use a delay to re-promote it back.

Constructors

RepromoteDelay

Fields

repromoteDelay :: DiffTime

type ReconnectDelay = RepromoteDelay

Deprecated: Use RepromoteDelay instead

Deprecated in `ouroboros-network-0.11.0.0`

data ExitPolicy a

ReturnPolicy allows to compute reconnection delay from value return by a mini-protocol. If a mini-protocol returned with an error epErrorDelay is used.

Constructors

ExitPolicy

Fields

epReturnDelay :: ReturnPolicy a

Compute RepromoteDelay from a return value.

epErrorDelay :: RepromoteDelay

The delay when a mini-protocol returned with an error.

stdExitPolicy :: ReturnPolicy a -> ExitPolicy a

ExitPolicy with 10s error delay.

type ReturnPolicy a = a -> RepromoteDelay

alwaysCleanReturnPolicy

Arguments

:: RepromoteDelay
re-promote delay on error

-> ExitPolicy a