Ouroboros.Network.MockChain.Chain

Description

Reference implementation of a representation of a block chain

Chain type and fundamental operations

data Chain block

Constructors

Genesis

(Chain block) :> block infixl 5

valid ∷ HasHeader block ⇒ Chain block → Bool

validExtension ∷ (HasCallStack, HasHeader block) ⇒ Chain block → block → Bool

foldChain ∷ (a → b → a) → a → Chain b → a

chainToList ∷ Chain block → [block]

Make a list from a Chain, in newest-to-oldest order.

Block re-exports

class (StandardHash b, Measured BlockMeasure b, Typeable b) ⇒ HasHeader b where

Abstract over the shape of blocks (or indeed just block headers)

Methods

blockHash ∷ b → HeaderHash b

blockPrevHash ∷ b → ChainHash b

blockSlot ∷ b → SlotNo

blockNo ∷ b → BlockNo

blockInvariant ∷ b → Bool

type family HeaderHash b ∷ ★

Header hash

Point type

newtype Point block

A point on the chain is identified by its Slot and HeaderHash.

The Slot tells us where to look and the HeaderHash either simply serves as a check, or in some contexts it disambiguates blocks from different forks that were in the same slot.

It's a newtype rather than a type synonym, because using a type synonym would lead to ambiguity, since HeaderHash is a non-injective type family.

Constructors

Point

Fields

getPoint ∷ WithOrigin (Block SlotNo (HeaderHash block))

blockPoint ∷ HasHeader block ⇒ block → Point block

Chain construction and inspection

Genesis

genesis ∷ Chain b

Head inspection

headPoint ∷ HasHeader block ⇒ Chain block → Point block

headSlot ∷ HasHeader block ⇒ Chain block → WithOrigin SlotNo

headHash ∷ HasHeader block ⇒ Chain block → ChainHash block

headTip ∷ HasHeader block ⇒ Chain block → Tip block

headBlockNo ∷ HasHeader block ⇒ Chain block → WithOrigin BlockNo

legacyHeadBlockNo ∷ HasHeader block ⇒ Chain block → BlockNo

TODO: This is wrong. There is no block number if we are at genesis (genesisBlockNo is the block number of the first block on the chain). Usage of this function should be phased out.

Basic operations

head ∷ Chain b → Maybe b

toNewestFirst ∷ Chain block → [block]

Produce the list of blocks, from most recent back to genesis

toOldestFirst ∷ Chain block → [block]

Produce the list of blocks, from genesis to the most recent

fromNewestFirst ∷ HasHeader block ⇒ [block] → Chain block

Make a chain from a list of blocks. The head of the list is the head of the chain.

fromOldestFirst ∷ HasHeader block ⇒ [block] → Chain block

Construct chain from list of blocks from oldest to newest

drop ∷ Int → Chain block → Chain block

length ∷ Chain block → Int

null ∷ Chain block → Bool

Update type and operations

data ChainUpdate block a

A representation of two actions to update a chain: add a block or roll back to a previous point.

The type parameter a is there to allow a Functor instance. Typically, it will be instantiated with block itself.

Constructors

AddBlock a

RollBack (Point block)

addBlock ∷ (HasCallStack, HasHeader block) ⇒ block → Chain block → Chain block

rollback ∷ HasHeader block ⇒ Point block → Chain block → Maybe (Chain block)

applyChainUpdate ∷ HasHeader block ⇒ ChainUpdate block block → Chain block → Maybe (Chain block)

applyChainUpdates ∷ HasHeader block ⇒ [ChainUpdate block block] → Chain block → Maybe (Chain block)

Special operations

pointOnChain ∷ HasHeader block ⇒ Point block → Chain block → Bool

pointIsAfter ∷ HasHeader block ⇒ Point block → Point block → Chain block → Bool

Check whether the first point is after the second point on the chain. Usually, this can simply be checked using the SlotNos, but some blocks may have the same SlotNo.

When the first point equals the second point, the answer will be False.

PRECONDITION: both points are on the chain.

successorBlock ∷ HasHeader block ⇒ Point block → Chain block → Maybe block

selectChain ∷ HasHeader block ⇒ Chain block → Chain block → Chain block

selectPoints ∷ HasHeader block ⇒ [Int] → Chain block → [Point block]

Select a bunch of Points based on offsets from the head of the chain. This is used in the chain consumer protocol as part of finding the intersection between a local and remote chain.

The typical pattern is to use a selection of offsets covering the last K blocks, biased towards more recent blocks. For example:

selectPoints (0 : [ fib n | n <- [1 .. 17] ])

findBlock ∷ (block → Bool) → Chain block → Maybe block

selectBlockRange ∷ HasHeader block ⇒ Chain block → Point block → Point block → Maybe [block]

findFirstPoint ∷ HasHeader block ⇒ [Point block] → Chain block → Maybe (Point block)

intersectChains ∷ HasHeader block ⇒ Chain block → Chain block → Maybe (Point block)

isPrefixOf ∷ Eq block ⇒ Chain block → Chain block → Bool

Conversion to/from ChainFragment

toChainFragment ∷ HasHeader block ⇒ Chain block → ChainFragment block

Convert a Chain to a ChainFragment.

fromChainFragment ∷ HasHeader block ⇒ ChainFragment block → Chain block

Convert a ChainFragment to a Chain.

unvalidatedToChainFragment ∷ HasHeader block ⇒ Chain block → ChainFragment block

Variant of toChainFragment that assumes a valid chain and will not validate the constructed ChainFragment.

Conversion to/from AnchoredFragment

fromAnchoredFragment ∷ HasHeader block ⇒ AnchoredFragment block → Maybe (Chain block)

Convert an AnchoredFragment to a Chain.

The anchor of the fragment must be genesisPoint, otherwise Nothing is returned.

toAnchoredFragment ∷ HasHeader block ⇒ Chain block → AnchoredFragment block

Convert a Chain to an AnchoredFragment.

The anchor of the fragment will be genesisPoint.

Helper functions

prettyPrintChain ∷ String → (block → String) → Chain block → String