Ouroboros.Network.NodeToClient

Description

This is the starting point for a module that will bring together the overall node to client protocol, as a collection of mini-protocols.

Documentation

nodeToClientProtocols :: NodeToClientProtocols appType addr bytes m a b -> NodeToClientVersion -> OuroborosApplicationWithMinimalCtx appType addr bytes m a b

Make an OuroborosApplication for the bundle of mini-protocols that make up the overall node-to-client protocol.

This function specifies the wire format protocol numbers as well as the protocols that run for each NodeToClientVersion.

They are chosen to not overlap with the node to node protocol numbers. This is not essential for correctness, but is helpful to allow a single shared implementation of tools that can analyse both protocols, e.g. wireshark plugins.

data NodeToClientProtocols appType ntcAddr bytes m a b

Record of node-to-client mini protocols.

Constructors

NodeToClientProtocols

Fields

localChainSyncProtocol :: RunMiniProtocolWithMinimalCtx appType ntcAddr bytes m a b

local chain-sync mini-protocol

localTxSubmissionProtocol :: RunMiniProtocolWithMinimalCtx appType ntcAddr bytes m a b

local tx-submission mini-protocol

localStateQueryProtocol :: RunMiniProtocolWithMinimalCtx appType ntcAddr bytes m a b

local state-query mini-protocol

localTxMonitorProtocol :: RunMiniProtocolWithMinimalCtx appType ntcAddr bytes m a b

local tx-monitor mini-protocol

data NodeToClientVersion

Constructors

NodeToClientV_9

NodeToClientV_10

NodeToClientV_11

NodeToClientV_12

NodeToClientV_13

NodeToClientV_14

NodeToClientV_15

NodeToClientV_16

data NodeToClientVersionData

Constructors

NodeToClientVersionData

Fields

networkMagic :: !NetworkMagic

query :: !Bool

data NetworkConnectTracers addr vNumber

Constructors

NetworkConnectTracers

Fields

nctMuxTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

nctHandshakeTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

nullNetworkConnectTracers :: NetworkConnectTracers addr vNumber

connectTo

Arguments

:: LocalSnocket
callback constructed by withIOManager

-> NetworkConnectTracers LocalAddress NodeToClientVersion

-> Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplicationWithMinimalCtx InitiatorMode LocalAddress ByteString IO a b)
A dictionary of protocol versions & applications to run on an established connection. The application to run will be chosen by initial handshake protocol (the highest shared version will be chosen).

-> FilePath
path of the unix socket or named pipe

-> IO ()

A specialised version of connectToNode. It is a general purpose function which can connect using any version of the protocol. This is mostly useful for future enhancements.

data NetworkServerTracers addr vNumber

Constructors

NetworkServerTracers

Fields

nstMuxTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

nstHandshakeTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

nstErrorPolicyTracer :: Tracer IO (WithAddr addr ErrorPolicyTrace)

nstAcceptPolicyTracer :: Tracer IO AcceptConnectionsPolicyTrace

nullNetworkServerTracers :: NetworkServerTracers addr vNumber

data NetworkMutableState addr

Constructors

NetworkMutableState

Fields

nmsConnectionTable :: ConnectionTable IO addr

nmsPeerStates :: StrictTVar IO (PeerStates IO addr)

newNetworkMutableState :: IO (NetworkMutableState addr)

newNetworkMutableStateSTM :: STM (NetworkMutableState addr)

cleanNetworkMutableState :: NetworkMutableState addr -> IO ()

withServer :: LocalSnocket -> NetworkServerTracers LocalAddress NodeToClientVersion -> NetworkMutableState LocalAddress -> LocalSocket -> Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplicationWithMinimalCtx ResponderMode LocalAddress ByteString IO a b) -> ErrorPolicies -> IO Void

A specialised version of withServerNode.

Comments to withServer apply here as well.

type NetworkClientSubcriptionTracers = NetworkSubscriptionTracers Identity LocalAddress NodeToClientVersion

data NetworkSubscriptionTracers withIPList addr vNumber

IP subscription tracers.

Constructors

NetworkSubscriptionTracers

Fields

nsMuxTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nsHandshakeTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nsErrorPolicyTracer :: Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nsSubscriptionTracer :: Tracer IO (withIPList (SubscriptionTrace addr))

subscription tracers; it is infrequent it should not be nullTracer by default.

data ClientSubscriptionParams a

Constructors

ClientSubscriptionParams

Fields

cspAddress :: !LocalAddress

cspConnectionAttemptDelay :: !(Maybe DiffTime)

cspErrorPolicies :: !ErrorPolicies

ncSubscriptionWorker :: forall mode x y. HasInitiator mode ~ True => LocalSnocket -> NetworkClientSubcriptionTracers -> NetworkMutableState LocalAddress -> ClientSubscriptionParams () -> Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplicationWithMinimalCtx mode LocalAddress ByteString IO x y) -> IO Void

ncSubscriptionWorker which starts given application versions on each established connection.

Null Protocol Peers

chainSyncPeerNull :: forall (header :: Type) (point :: Type) (tip :: Type) m a. MonadDelay m => Peer (ChainSync header point tip) AsClient StIdle m a

localStateQueryPeerNull :: forall (block :: Type) (point :: Type) (query :: Type -> Type) m a. MonadDelay m => Peer (LocalStateQuery block point query) AsClient StIdle m a

localTxSubmissionPeerNull :: forall (tx :: Type) (reject :: Type) m a. MonadDelay m => Peer (LocalTxSubmission tx reject) AsClient StIdle m a

localTxMonitorPeerNull :: forall (txid :: Type) (tx :: Type) (slot :: Type) m a. MonadDelay m => Peer (LocalTxMonitor txid tx slot) AsClient StIdle m a

Re-exported network interface

newtype IOManager

Constructors

IOManager

Fields

associateWithIOManager :: forall hole. hole -> IO ()

type AssociateWithIOCP = IOManager

withIOManager :: WithIOManager

type LocalSnocket = Snocket IO LocalSocket LocalAddress

localSnocket :: IOManager -> LocalSnocket

newtype LocalSocket

Constructors

LocalSocket

Fields

getLocalHandle :: LocalHandle

newtype LocalAddress

Constructors

LocalAddress

Fields

getFilePath :: FilePath

Versions

newtype Versions vNum vData r

Constructors

Versions

Fields

getVersions :: Map vNum (Version vData r)

versionedNodeToClientProtocols :: NodeToClientVersion -> NodeToClientVersionData -> NodeToClientProtocols appType LocalAddress bytes m a b -> Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplicationWithMinimalCtx appType LocalAddress bytes m a b)

Versions containing a single version of nodeToClientProtocols.

simpleSingletonVersions :: vNum -> vData -> r -> Versions vNum vData r

foldMapVersions :: (Ord vNum, Foldable f, HasCallStack) => (x -> Versions vNum extra r) -> f x -> Versions vNum extra r

combineVersions :: (Ord vNum, Foldable f, HasCallStack) => f (Versions vNum extra r) -> Versions vNum extra r

Codecs

nodeToClientHandshakeCodec :: forall (m :: Type -> Type). MonadST m => Codec (Handshake NodeToClientVersion Term) DeserialiseFailure m ByteString

nodeToClientVersionCodec :: CodecCBORTerm (Text, Maybe Int) NodeToClientVersion

nodeToClientCodecCBORTerm :: NodeToClientVersion -> CodecCBORTerm Text NodeToClientVersionData

Re-exports

data ConnectionId addr

Constructors

ConnectionId

Fields

localAddress :: !addr

remoteAddress :: !addr

newtype MinimalInitiatorContext addr

Constructors

MinimalInitiatorContext

Fields

micConnectionId :: ConnectionId addr

newtype ResponderContext addr

Constructors

ResponderContext

Fields

rcConnectionId :: ConnectionId addr

type LocalConnectionId = ConnectionId LocalAddress

data ErrorPolicies

Constructors

ErrorPolicies

Fields

epAppErrorPolicies :: [ErrorPolicy]

epConErrorPolicies :: [ErrorPolicy]

networkErrorPolicies :: ErrorPolicies

ErrorPolicies for client application. Additional rules can be added by means of a Semigroup instance of ErrorPolicies.

This error policies will try to preserve subscriptionWorker, e.g. if the connect function throws an IOException we will suspend it for a shortDelay, and try to re-connect.

This allows to recover from a situation where a node temporarily shutsdown, or running a client application which is subscribed two more than one node (possibly over network).

nullErrorPolicies :: ErrorPolicies

data ErrorPolicy where

Constructors

ErrorPolicy :: forall e. Exception e => (e -> Maybe (SuspendDecision DiffTime)) -> ErrorPolicy

data ErrorPolicyTrace

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) DiffTime DiffTime

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) DiffTime

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)

ErrorPolicyResumePeer

ErrorPolicyKeepSuspended

ErrorPolicyResumeConsumer

ErrorPolicyResumeProducer

ErrorPolicyUnhandledApplicationException SomeException

ErrorPolicyUnhandledConnectionException SomeException

ErrorPolicyAcceptException IOException

data WithAddr addr a

Constructors

WithAddr

Fields

wiaAddr :: addr

wiaEvent :: a

data SuspendDecision t

Constructors

SuspendPeer !t !t

SuspendConsumer !t

Throw

data TraceSendRecv ps where

Constructors

TraceSendMsg :: forall ps. AnyMessageAndAgency ps -> TraceSendRecv ps

TraceRecvMsg :: forall ps. AnyMessageAndAgency ps -> TraceSendRecv ps

data ProtocolLimitFailure

data Handshake (vNumber :: k) (vParams :: k1)

data LocalAddresses addr

Constructors

LocalAddresses

Fields

laIpv4 :: Maybe addr

laIpv6 :: Maybe addr

laUnix :: Maybe addr

data SubscriptionTrace addr

Constructors

SubscriptionTraceConnectStart addr

SubscriptionTraceConnectEnd addr ConnectResult

Exception e => SubscriptionTraceSocketAllocationException addr e

Exception e => SubscriptionTraceConnectException addr e

Exception e => SubscriptionTraceApplicationException addr e

SubscriptionTraceTryConnectToPeer addr

SubscriptionTraceSkippingPeer addr

SubscriptionTraceSubscriptionRunning

SubscriptionTraceSubscriptionWaiting Int

SubscriptionTraceSubscriptionFailed

SubscriptionTraceSubscriptionWaitingNewConnection DiffTime

SubscriptionTraceStart Int

SubscriptionTraceRestart DiffTime Int Int

SubscriptionTraceConnectionExist addr

SubscriptionTraceUnsupportedRemoteAddr addr

SubscriptionTraceMissingLocalAddress

SubscriptionTraceAllocateSocket addr

SubscriptionTraceCloseSocket addr

type HandshakeTr ntcAddr ntcVersion = WithMuxBearer (ConnectionId ntcAddr) (TraceSendRecv (Handshake ntcVersion Term))