Ouroboros.Network.NodeToClient.Version

Documentation

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_1

NodeToClientV_2
added local-query mini-protocol

NodeToClientV_3
enabled CardanoNodeToClientVersion2

NodeToClientV_4
enabled CardanoNodeToClientVersion3, adding more queries

NodeToClientV_5
enabled CardanoNodeToClientVersion4, i.e., Allegra

NodeToClientV_6
enabled CardanoNodeToClientVersion5, i.e., Mary

NodeToClientV_7
enabled CardanoNodeToClientVersion6, adding a query

NodeToClientV_8
LocalStateQuery protocol codec change, allows to acquire tip point.

NodeToClientV_9
enabled CardanoNodeToClientVersion7, i.e., Alonzo

NodeToClientV_10
added GetChainBlockNo and GetChainPoint queries

NodeToClientV_11
added GetRewardInfoPools Block query

NodeToClientV_12
added LocalTxMonitor mini-protocol

newtype NodeToClientVersionData

Version data for NodeToClient protocol v1

Constructors

NodeToClientVersionData

Fields

networkMagic ∷ NetworkMagic

nodeToClientVersionCodec ∷ CodecCBORTerm (Text, Maybe Int) NodeToClientVersion

We set 16ths bit to distinguish NodeToNodeVersion and NodeToClientVersion. This way connecting wrong protocol suite will fail during Handshake negotiation

This is done in backward compatible way, so NodeToClientV_1 encoding is not changed.

nodeToClientCodecCBORTerm ∷ NodeToClientVersion → CodecCBORTerm Text NodeToClientVersionData