Ouroboros.Network.NodeToNode

Description

This is the starting point for a module that will bring together the overall node to node protocol, as a collection of mini-protocols.

Documentation

nodeToNodeProtocols :: MiniProtocolParameters -> (ConnectionId addr -> STM m ControlMessage -> NodeToNodeProtocols muxMode bytes m a b) -> NodeToNodeVersion -> OuroborosBundle muxMode addr bytes m a b

Make an OuroborosApplication for the bundle of mini-protocols that make up the overall node-to-node protocol.

This function specifies the wire format protocol numbers.

The application specific protocol numbers start from 2. The MiniProtocolNum 0 is reserved for the Handshake protocol, while MiniProtocolNum 1 is reserved for DeltaQ messages. Handshake protocol is not included in NodeToNodeProtocols as it runs before mux is started but it reusing MuxBearer to send and receive messages. Only when the handshake protocol succeeds, we will know which protocols to run / multiplex.

These are chosen to not overlap with the node to client protocol numbers (and the handshake protocol number). This is not essential for correctness, but is helpful to allow a single shared implementation of tools that can analyse both protocols, e.g. wireshark plugins.

data NodeToNodeProtocols appType bytes m a b

Constructors

NodeToNodeProtocols

Fields

chainSyncProtocol :: RunMiniProtocol appType bytes m a b

chain-sync mini-protocol

blockFetchProtocol :: RunMiniProtocol appType bytes m a b

block-fetch mini-protocol

txSubmissionProtocol :: RunMiniProtocol appType bytes m a b

tx-submission mini-protocol

keepAliveProtocol :: RunMiniProtocol appType bytes m a b

keep-alive mini-protocol

data MiniProtocolParameters

Constructors

MiniProtocolParameters

Fields

chainSyncPipeliningHighMark :: !Word32

high threshold for pipelining (we will never exceed that many messages pipelined).

chainSyncPipeliningLowMark :: !Word32

low threshold: if we hit the chainSyncPipeliningHighMark we will listen for responses until there are at most chainSyncPipeliningLowMark pipelined message

Must be smaller than chainSyncPipeliningHighMark.

Note: chainSyncPipeliningLowMark and chainSyncPipeliningLowMark are passed to pipelineDecisionLowHighMark.

blockFetchPipeliningMax :: !Word

maximal number of pipelined messages in 'block-fetch' mini-protocol.

txSubmissionMaxUnacked :: !Word16

maximal number of unacked tx (pipelining is bounded by twice this number)

chainSyncProtocolLimits :: MiniProtocolParameters -> MiniProtocolLimits

blockFetchProtocolLimits :: MiniProtocolParameters -> MiniProtocolLimits

txSubmissionProtocolLimits :: MiniProtocolParameters -> MiniProtocolLimits

keepAliveProtocolLimits :: MiniProtocolParameters -> MiniProtocolLimits

defaultMiniProtocolParameters :: MiniProtocolParameters

data NodeToNodeVersion

Enumeration of node to node protocol versions.

Constructors

NodeToNodeV_7
Changes:

new KeepAlive codec

Enable CardanoNodeToNodeVersion5, i.e., Alonzo

NodeToNodeV_8
Changes:

Enable block diffusion pipelining in ChainSync and BlockFetch logic.

NodeToNodeV_9
Changes:

Enable CardanoNodeToNodeVersion6, i.e., Babbage

NodeToNodeV_10
Changes:

Enable full duplex connections. NOTE: This is an experimental protocol version, which is not yet released. Until initial P2P version it must be kept as the last version, which allows us to keep it as an experimental version.

data NodeToNodeVersionData

Version data for NodeToNode protocol

Constructors

NodeToNodeVersionData

Fields

networkMagic :: !NetworkMagic

diffusionMode :: !DiffusionMode

data NetworkConnectTracers addr vNumber

Tracer used by connectToNode (and derivatives, like connectTo or 'Ouroboros.Network.NodeToClient.connectTo).

Constructors

NetworkConnectTracers

Fields

nctMuxTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nctHandshakeTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nullNetworkConnectTracers :: NetworkConnectTracers addr vNumber

connectTo :: Snocket IO Socket SockAddr -> NetworkConnectTracers SockAddr NodeToNodeVersion -> Versions NodeToNodeVersion NodeToNodeVersionData (OuroborosApplication InitiatorMode SockAddr ByteString IO a b) -> Maybe SockAddr -> SockAddr -> IO ()

A specialised version of connectToNode.

data NetworkServerTracers addr vNumber

Tracers required by a server which handles inbound connections.

Constructors

NetworkServerTracers

Fields

nstMuxTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nstHandshakeTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nstErrorPolicyTracer :: Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nstAcceptPolicyTracer :: Tracer IO AcceptConnectionsPolicyTrace

tracing rate limiting of accepting connections.

nullNetworkServerTracers :: NetworkServerTracers addr vNumber

data NetworkMutableState addr

Mutable state maintained by the network component.

Constructors

NetworkMutableState

Fields

nmsConnectionTable :: ConnectionTable IO addr

ConnectionTable which maintains information about current upstream and downstream connections.

nmsPeerStates :: StrictTVar IO (PeerStates IO addr)

PeerStates which maintains state of each downstream / upstream peer that errored, misbehaved or was not interesting to us.

data AcceptedConnectionsLimit

Policy which governs how to limit the number of accepted connections.

Constructors

AcceptedConnectionsLimit

Fields

acceptedConnectionsHardLimit :: !Word32

Hard limit of accepted connections.

acceptedConnectionsSoftLimit :: !Word32

Soft limit of accepted connections. If we are above this threshold, we will start rate limiting.

acceptedConnectionsDelay :: !DiffTime

Max delay for limiting accepted connections. We use linear regression starting from 0 at the soft limit up to acceptedConnectionDelay at the hard limit.

newNetworkMutableState :: IO (NetworkMutableState addr)

newNetworkMutableStateSTM :: STM (NetworkMutableState addr)

cleanNetworkMutableState :: NetworkMutableState addr -> IO ()

Clean PeerStates within NetworkMutableState every 200s

withServer :: SocketSnocket -> NetworkServerTracers SockAddr NodeToNodeVersion -> NetworkMutableState SockAddr -> AcceptedConnectionsLimit -> Socket -> Versions NodeToNodeVersion NodeToNodeVersionData (OuroborosApplication ResponderMode SockAddr ByteString IO a b) -> ErrorPolicies -> IO Void

A specialised version of withServerNode. It forks a thread which runs an accept loop (server thread):

when the server thread throws an exception the main thread rethrows it (by wait)

when an async exception is thrown to kill the main thread the server thread will be cancelled as well (by withAsync)

P2P Governor

data DomainAccessPoint

A product of a Domain and PortNumber. After resolving the domain we will use the PortNumber to form SockAddr.

Constructors

DomainAccessPoint

Fields

dapDomain :: !Domain

dapPortNumber :: !PortNumber

data PeerAdvertise

Should this peer be advertised to other peers asking for known peers? For certain peers specified by configuration it would be an appropriate policy to keep them private.

Constructors

DoNotAdvertisePeer

DoAdvertisePeer

data PeerSelectionTargets

Adjustable targets for the peer selection mechanism.

These are used by the peer selection governor as targets. They are used by the peer churn governor loop as knobs to adjust, to influence the peer selection governor.

The known, established and active peer targets are targets both from below and from above: the governor will attempt to grow or shrink the sets to hit these targets.

Unlike the other targets, the root peer target is "one sided", it is only a target from below. The governor does not try to shrink the root set to hit it, it simply stops looking for more.

There is also an implicit target that enough local root peers are selected as active. This comes from the configuration for local roots, and is not an independently adjustable target.

Constructors

PeerSelectionTargets

Fields

targetNumberOfRootPeers :: !Int

targetNumberOfKnownPeers :: !Int

targetNumberOfEstablishedPeers :: !Int

targetNumberOfActivePeers :: !Int

Subscription Workers

IP subscription worker

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps :: ![SockAddr]

List of destinations to possibly connect to

ispValency :: !Int

Number of parallel connections to keep actice.

type NetworkIPSubscriptionTracers addr vNumber = NetworkSubscriptionTracers WithIPList addr vNumber

data NetworkSubscriptionTracers withIPList addr vNumber

IP subscription tracers.

Constructors

NetworkSubscriptionTracers

Fields

nsMuxTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nsHandshakeTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nsErrorPolicyTracer :: Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nsSubscriptionTracer :: Tracer IO (withIPList (SubscriptionTrace addr))

subscription tracers; it is infrequent it should not be nullTracer by default.

nullNetworkSubscriptionTracers :: NetworkSubscriptionTracers withIPList addr vNumber

data SubscriptionParams a target

ipSubscriptionWorker and dnsSubscriptionWorker parameters

Constructors

SubscriptionParams

Fields

spLocalAddresses :: LocalAddresses SockAddr

spConnectionAttemptDelay :: SockAddr -> Maybe DiffTime

should return expected delay for the given address

spErrorPolicies :: ErrorPolicies

spSubscriptionTarget :: target

type IPSubscriptionParams a = SubscriptionParams a IPSubscriptionTarget

ipSubscriptionWorker :: forall mode x y. HasInitiator mode ~ True => SocketSnocket -> NetworkIPSubscriptionTracers SockAddr NodeToNodeVersion -> NetworkMutableState SockAddr -> IPSubscriptionParams () -> Versions NodeToNodeVersion NodeToNodeVersionData (OuroborosApplication mode SockAddr ByteString IO x y) -> IO Void

ipSubscriptionWorker which starts given application versions on each established connection.

DNS subscription worker

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain :: !Domain

dstPort :: !PortNumber

dstValency :: !Int

type DnsSubscriptionParams a = SubscriptionParams a DnsSubscriptionTarget

data NetworkDNSSubscriptionTracers vNumber addr

DNS subscription tracers.

Constructors

NetworkDNSSubscriptionTracers

Fields

ndstMuxTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

ndstHandshakeTracer :: Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

ndstErrorPolicyTracer :: Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

ndstSubscriptionTracer :: Tracer IO (WithDomainName (SubscriptionTrace addr))

subscription tracer; it is infrequent it should not be nullTracer by default.

ndstDnsTracer :: Tracer IO (WithDomainName DnsTrace)

dns resolver tracer; it is infrequent it should not be nullTracer by default.

nullNetworkDNSSubscriptionTracers :: NetworkDNSSubscriptionTracers vNumber peerid

dnsSubscriptionWorker :: forall mode x y. HasInitiator mode ~ True => SocketSnocket -> NetworkDNSSubscriptionTracers NodeToNodeVersion SockAddr -> NetworkMutableState SockAddr -> DnsSubscriptionParams () -> Versions NodeToNodeVersion NodeToNodeVersionData (OuroborosApplication mode SockAddr ByteString IO x y) -> IO Void

dnsSubscriptionWorker which starts given application versions on each established connection.

Versions

newtype Versions vNum vData r

The version map supported by the local agent keyed on the version identifier.

Each Version contains a function which takes negotiated version data and returns negotiated application (the r type variable).

If one needs to combine multiple versions the simplest way is to use one of the combinators: foldMapVersions, combineVersions or the Semigroup instance directly:

fold $ (simpleSingletonVersions ...)
      :| [ (simpleSingletonVersions ...)
         , (simpleSingletonVersions ...)
         , ...
         ]

Constructors

Versions

Fields

getVersions :: Map vNum (Version vData r)

data DiffusionMode

The flag which indicates wheather the node runs only initiator or both initiator or responder node. It does not however specify weather the node is using duplex connections, this is implicit see NodeToNodeV_4

Constructors

InitiatorOnlyDiffusionMode

InitiatorAndResponderDiffusionMode

simpleSingletonVersions :: vNum -> vData -> r -> Versions vNum vData r

Singleton smart constructor for Versions.

foldMapVersions :: (Ord vNum, Foldable f, HasCallStack) => (x -> Versions vNum extra r) -> f x -> Versions vNum extra r

Useful for folding multiple Versions.

A foldMap restricted to the Versions Semigroup.

PRECONDITION: f x is non-empty.

combineVersions :: (Ord vNum, Foldable f, HasCallStack) => f (Versions vNum extra r) -> Versions vNum extra r

Codecs

nodeToNodeHandshakeCodec :: MonadST m => Codec (Handshake NodeToNodeVersion Term) DeserialiseFailure m ByteString

Handshake codec for the node-to-node protocol suite.

nodeToNodeVersionCodec :: CodecCBORTerm (Text, Maybe Int) NodeToNodeVersion

nodeToNodeCodecCBORTerm :: NodeToNodeVersion -> CodecCBORTerm Text NodeToNodeVersionData

Re-exports

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress :: !addr

remoteAddress :: !addr

type RemoteAddress = SockAddr

type RemoteConnectionId = ConnectionId RemoteAddress

data ProtocolLimitFailure

data Handshake (vNumber :: k) (vParams :: k1)

The handshake mini-protocol is used initially to agree the version and associated parameters of the protocol to use for all subsequent communication.

data LocalAddresses addr

Constructors

LocalAddresses

Fields

laIpv4 :: Maybe addr

Local IPv4 address to use, Nothing indicates don't use IPv4

laIpv6 :: Maybe addr

Local IPv6 address to use, Nothing indicates don't use IPv6

laUnix :: Maybe addr

Local Unix address to use, Nothing indicates don't use Unix sockets

data Socket

Basic type for a socket.

isPipeliningEnabled :: NodeToNodeVersion -> Bool

Check whether a version enabling diffusion pipelining has been negotiated.

Error Policies and Peer state

data ErrorPolicies

List of error policies for exception handling and a policy for handing application return values.

Constructors

ErrorPolicies

Fields

epAppErrorPolicies :: [ErrorPolicy]

Application Error Policies

epConErrorPolicies :: [ErrorPolicy]

connect Error Policies

remoteNetworkErrorPolicy :: ErrorPolicies

A minimal error policy for remote peers, which only handles exceptions raised by `ouroboros-network`.

localNetworkErrorPolicy :: ErrorPolicies

Error policy for local clients. This is equivalent to nullErrorPolicies, but explicit in the errors which can be caught.

We are very permissive here, and very strict in the networkErrorPolicy. After any failure the client will be killed and not penalised by this policy. This allows to restart the local client without a delay.

nullErrorPolicies :: ErrorPolicies

data ErrorPolicy where

Constructors

ErrorPolicy

Fields

:: forall e. Exception e

=> (e -> Maybe (SuspendDecision DiffTime))

Nothing means no decision. It is equivalent to not having the policy at all. In evalErrorPolicies this will fall-through and match against the remaining policies.

-> ErrorPolicy

data SuspendDecision t

Semigroup of commands which acts on PeerState. The t variable might be initiated to DiffTime or Time m.

This semigroup allows to either suspend both consumer and producer or just the consumer part.

Constructors

SuspendPeer !t !t
peer is suspend; The first t is the time until which a local producer is suspended, the second one is the time until which a local consumer is suspended.

SuspendConsumer !t
suspend local consumer / initiator side until t (this mean we are not allowing to communicate with the producer / responder of a remote peer).

Throw
throw an error from the main thread.

Traces

data AcceptConnectionsPolicyTrace

Trace for the AcceptConnectionsLimit policy.

Constructors

ServerTraceAcceptConnectionRateLimiting DiffTime Int

ServerTraceAcceptConnectionHardLimit Word32

ServerTraceAcceptConnectionResume Int

data TraceSendRecv ps where

Structured Tracer output for runPeer and derivitives.

Constructors

TraceSendMsg :: forall ps. AnyMessageAndAgency ps -> TraceSendRecv ps

TraceRecvMsg :: forall ps. AnyMessageAndAgency ps -> TraceSendRecv ps

data SubscriptionTrace addr

Constructors

SubscriptionTraceConnectStart addr

SubscriptionTraceConnectEnd addr ConnectResult

Exception e => SubscriptionTraceSocketAllocationException addr e

Exception e => SubscriptionTraceConnectException addr e

Exception e => SubscriptionTraceApplicationException addr e

SubscriptionTraceTryConnectToPeer addr

SubscriptionTraceSkippingPeer addr

SubscriptionTraceSubscriptionRunning

SubscriptionTraceSubscriptionWaiting Int

SubscriptionTraceSubscriptionFailed

SubscriptionTraceSubscriptionWaitingNewConnection DiffTime

SubscriptionTraceStart Int

SubscriptionTraceRestart DiffTime Int Int

SubscriptionTraceConnectionExist addr

SubscriptionTraceUnsupportedRemoteAddr addr

SubscriptionTraceMissingLocalAddress

SubscriptionTraceAllocateSocket addr

SubscriptionTraceCloseSocket addr

data DnsTrace

Constructors

DnsTraceLookupException SomeException

DnsTraceLookupAError DNSError

DnsTraceLookupAAAAError DNSError

DnsTraceLookupIPv6First

DnsTraceLookupIPv4First

DnsTraceLookupAResult [SockAddr]

DnsTraceLookupAAAAResult [SockAddr]

data ErrorPolicyTrace

Trace data for error policies

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) DiffTime DiffTime
suspending peer with a given exception until

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) DiffTime
suspending consumer until

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)
caught a local exception

ErrorPolicyResumePeer
resume a peer (both consumer and producer)

ErrorPolicyKeepSuspended
consumer was suspended until producer will resume

ErrorPolicyResumeConsumer
resume consumer

ErrorPolicyResumeProducer
resume producer

ErrorPolicyUnhandledApplicationException SomeException
an application throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyUnhandledConnectionException SomeException
connect throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyAcceptException IOException
accept throwed an exception

data WithIPList a

Constructors

WithIPList

Fields

wilSrc :: LocalAddresses SockAddr

wilDsts :: [SockAddr]

wilEvent :: a

data WithDomainName a

Constructors

WithDomainName

Fields

wdnDomain :: Domain

wdnEvent :: a

data WithAddr addr a

Constructors

WithAddr

Fields

wiaAddr :: addr

wiaEvent :: a

type HandshakeTr ntnAddr ntnVersion = WithMuxBearer (ConnectionId ntnAddr) (TraceSendRecv (Handshake ntnVersion Term))

For Consensus ThreadNet Tests

chainSyncMiniProtocolNum :: MiniProtocolNum

blockFetchMiniProtocolNum :: MiniProtocolNum

txSubmissionMiniProtocolNum :: MiniProtocolNum

keepAliveMiniProtocolNum :: MiniProtocolNum