Ouroboros.Network.NodeToNode.Version

Documentation

data NodeToNodeVersion

Enumeration of node to node protocol versions.

Constructors

NodeToNodeV_1

NodeToNodeV_2
Changes:

Enable block size hints for Byron headers in ChainSync

Enable CardanoNodeToNodeVersion2, i.e., Shelley

NodeToNodeV_3
Changes:

Enable KeepAlive miniprotocol

NodeToNodeV_4
Changes:

Added DiffusionMode Handshake argument.

Enable CardanoNodeToNodeVersion3, i.e., Allegra

NodeToNodeV_5
Changes:

Enable CardanoNodeToNodeVersion4, i.e., Mary

NodeToNodeV_6
Changes:

Replace TxSubmision with Txsubmission2 protocol.

NodeToNodeV_7
Changes:

new KeepAlive codec

Enable CardanoNodeToNodeVersion5, i.e., Alonzo

NodeToNodeV_8
Changes:

Enable full duplex connections.

data NodeToNodeVersionData

Version data for NodeToNode protocol

Constructors

NodeToNodeVersionData

Fields

networkMagic ∷ !NetworkMagic

diffusionMode ∷ !DiffusionMode

data DiffusionMode

The flag which indicates wheather the node runs only initiator or both initiator or responder node. It does not however specify weather the node is using duplex connections, this is implicit see NodeToNodeV_4

Constructors

InitiatorOnlyDiffusionMode

InitiatorAndResponderDiffusionMode

data ConnectionMode

Constructors

UnidirectionalMode

DuplexMode

nodeToNodeVersionCodec ∷ CodecCBORTerm (Text, Maybe Int) NodeToNodeVersion

nodeToNodeCodecCBORTerm ∷ NodeToNodeVersion → CodecCBORTerm Text NodeToNodeVersionData