Ouroboros.Network.PeerSelection.Governor.Types

P2P governor policies

data PeerSelectionPolicy peeraddr m

Constructors

PeerSelectionPolicy

Fields

policyPickKnownPeersForGossip ∷ PickPolicy peeraddr m

policyPickColdPeersToPromote ∷ PickPolicy peeraddr m

policyPickWarmPeersToPromote ∷ PickPolicy peeraddr m

policyPickHotPeersToDemote ∷ PickPolicy peeraddr m

policyPickWarmPeersToDemote ∷ PickPolicy peeraddr m

policyPickColdPeersToForget ∷ PickPolicy peeraddr m

policyFindPublicRootTimeout ∷ !DiffTime

policyMaxInProgressGossipReqs ∷ !Int

policyGossipRetryTime ∷ !DiffTime

policyGossipBatchWaitTime ∷ !DiffTime

policyGossipOverallTimeout ∷ !DiffTime

data PeerSelectionTargets

Adjustable targets for the peer selection mechanism.

These are used by the peer selection governor as targets. They are used by the peer churn governor loop as knobs to adjust, to influence the peer selection governor.

The known, established and active peer targets are targets both from below and from above: the governor will attempt to grow or shrink the sets to hit these targets.

Unlike the other targets, the root peer target is "one sided", it is only a target from below. The governor does not try to shrink the root set to hit it, it simply stops looking for more.

There is also an implicit target that enough local root peers are selected as active. This comes from the configuration for local roots, and is not an independently adjustable target.

Constructors

PeerSelectionTargets

Fields

targetNumberOfRootPeers ∷ !Int

targetNumberOfKnownPeers ∷ !Int

targetNumberOfEstablishedPeers ∷ !Int

targetNumberOfActivePeers ∷ !Int

nullPeerSelectionTargets ∷ PeerSelectionTargets

sanePeerSelectionTargets ∷ PeerSelectionTargets → Bool

type PickPolicy peeraddr m = (peeraddr → PeerSource) → (peeraddr → Int) → (peeraddr → Bool) → Set peeraddr → Int → STM m (Set peeraddr)

A peer pick policy is an action that picks a subset of elements from a map of peers.

The pre-condition is that the map of available choices will be non-empty, and the requested number to pick will be strictly positive.

The post-condition is that the picked set is non-empty but must not be bigger than the requested number.

pickPeers ∷ (Ord peeraddr, Functor m) ⇒ PeerSelectionState peeraddr peerconn → ((peeraddr → PeerSource) → (peeraddr → Int) → (peeraddr → Bool) → Set peeraddr → Int → m (Set peeraddr)) → Set peeraddr → Int → m (Set peeraddr)

Check pre-conditions and post-conditions on the pick policies, and supply additional peer atributes from the current state.

P2P governor low level API

data PeerStateActions peeraddr peerconn m

Callbacks which are performed to change peer state.

Constructors

PeerStateActions

Fields

monitorPeerConnection ∷ peerconn → STM m PeerStatus

Monitor peer state.

establishPeerConnection ∷ peeraddr → m peerconn

Establish new connection.

activatePeerConnection ∷ peerconn → m ()

Activate a connection: warm to hot promotion.

deactivatePeerConnection ∷ peerconn → m ()

Deactive a peer: hot to warm demotion.

closePeerConnection ∷ peerconn → m ()

Close a connection: warm to cold transition.

data PeerSelectionActions peeraddr peerconn m

Actions performed by the peer selection governor.

These being pluggable allows:

choice of known peer root sets

running both in simulation and for real

Constructors

PeerSelectionActions

Fields

readPeerSelectionTargets ∷ STM m PeerSelectionTargets

readLocalRootPeers ∷ STM m [(Int, Map peeraddr PeerAdvertise)]

Read the current set of locally or privately known root peers.

In general this is expected to be updated asynchronously by some other thread. It is intended to cover the use case of peers from local configuration. It could be dynamic due to DNS resolution, or due to dynamic configuration updates.

It is structured as a collection of (non-overlapping) groups of peers where we are supposed to select n from each group.

requestPublicRootPeers ∷ Int → m (Set peeraddr, DiffTime)

Request a sample of public root peers.

It is intended to cover use cases including:

federated relays from a DNS pool

stake pool relays published in the blockchain

a pre-distributed snapshot of stake pool relays from the blockchain

requestPeerGossip ∷ peeraddr → m [peeraddr]

The action to contact a known peer and request a sample of its known peers.

This is synchronous, but it should expect to be interrupted by a timeout asynchronous exception. Failures are throw as exceptions.

peerStateActions ∷ PeerStateActions peeraddr peerconn m

Core actions run by the governor to change PeerStatus.

data ChurnMode

Constructors

ChurnModeBulkSync

ChurnModeNormal

P2P govnernor internals

data PeerSelectionState peeraddr peerconn

The internal state used by the peerSelectionGovernor.

The local and public root sets are disjoint, and their union is the overall root set.

Constructors

PeerSelectionState

Fields

targets ∷ !PeerSelectionTargets

localRootPeers ∷ !(LocalRootPeers peeraddr)

The current set of local root peers. This is structured as a bunch of groups, with a target for each group. This gives us a set of n-of-m choices, e.g. "pick 2 from this group and 1 from this group".

The targets must of course be achievable, and to keep things simple, the groups must be disjoint.

publicRootPeers ∷ !(Set peeraddr)

knownPeers ∷ !(KnownPeers peeraddr)

establishedPeers ∷ !(EstablishedPeers peeraddr peerconn)

activePeers ∷ !(Set peeraddr)

publicRootBackoffs ∷ !Int

A counter to manage the exponential backoff strategy for when to retry querying for more public root peers. It is negative for retry counts after failure, and positive for retry counts that are successful but make no progress.

publicRootRetryTime ∷ !Time

The earliest time we would be prepared to request more public root peers. This is used with the publicRootBackoffs to manage the exponential backoff.

inProgressPublicRootsReq ∷ !Bool

inProgressGossipReqs ∷ !Int

inProgressPromoteCold ∷ !(Set peeraddr)

inProgressPromoteWarm ∷ !(Set peeraddr)

inProgressDemoteWarm ∷ !(Set peeraddr)

inProgressDemoteHot ∷ !(Set peeraddr)

fuzzRng ∷ !StdGen

Rng for fuzzy delay

countersCache ∷ !(Cache PeerSelectionCounters)

PeerSelectionCounters counters cache. Allows to only trace values when necessary.

emptyPeerSelectionState ∷ StdGen → PeerSelectionState peeraddr peerconn

assertPeerSelectionState ∷ Ord peeraddr ⇒ PeerSelectionState peeraddr peerconn → a → a

establishedPeersStatus ∷ Ord peeraddr ⇒ PeerSelectionState peeraddr peerconn → Map peeraddr PeerStatus

A view of the status of each established peer, for testing and debugging.

data Guarded m a

The governor is using Guarded m (Decision m peeraddr peerconn) where m is an STM monad, to drive its progress.

Constructors

GuardedSkip !(Maybe (Min Time))
GuardedSkip is used to instruct that there is no action to be made by the governor.

Let us note that the combined value which is computed by guardedDecisions term in peerSelectionGovernorLoop will never return it: this is bacause there are monitoring decisions which never return this constructor, e.g. targetPeers, jobs, connections, and thus the governor has always something to do.

Bundled Patterns

pattern Guarded ∷ Maybe (Min Time) → m a → Guarded m a
Guarded constructor which provides an action, possibly with a timeout, to the governor main loop. It hides the use of FirstToFinish synchronisation.

data Decision m peeraddr peerconn

Constructors

Decision

Fields

decisionTrace ∷ TracePeerSelection peeraddr

A trace event to classify the decision and action

decisionState ∷ PeerSelectionState peeraddr peerconn

An updated state to use immediately

decisionJobs ∷ [Job () m (Completion m peeraddr peerconn)]

An optional Job to execute asynchronously. This job leads to a further Decision. This gives a state update to apply upon completion, but also allows chaining further job actions.

type TimedDecision m peeraddr peerconn = Time → Decision m peeraddr peerconn

Decision which has access to the current time, rather than the time when the governor's loop blocked to make a decision.

type MkGuardedDecision peeraddr peerconn m = PeerSelectionPolicy peeraddr m → PeerSelectionState peeraddr peerconn → Guarded (STM m) (TimedDecision m peeraddr peerconn)

Type alias for function types which are used to create governor decisions. Allmost all decisions are following this pattern.

newtype Completion m peeraddr peerconn

Constructors

Completion (PeerSelectionState peeraddr peerconn → Time → Decision m peeraddr peerconn)

data PeerSelectionCounters

Constructors

PeerSelectionCounters

Fields

coldPeers ∷ Int

warmPeers ∷ Int

hotPeers ∷ Int

peerStateToCounters ∷ Ord peeraddr ⇒ PeerSelectionState peeraddr peerconn → PeerSelectionCounters

Traces

data TracePeerSelection peeraddr

Constructors

TraceLocalRootPeersChanged (LocalRootPeers peeraddr) (LocalRootPeers peeraddr)

TraceTargetsChanged PeerSelectionTargets PeerSelectionTargets

TracePublicRootsRequest Int Int

TracePublicRootsResults (Set peeraddr) Int DiffTime

TracePublicRootsFailure SomeException Int DiffTime

TraceGossipRequests Int Int (Set peeraddr) (Set peeraddr)
target known peers, actual known peers, peers available for gossip, peers selected for gossip

TraceGossipResults [(peeraddr, Either SomeException [peeraddr])]

TraceForgetColdPeers Int Int (Set peeraddr)
target known peers, actual known peers, selected peer

TracePromoteColdPeers Int Int (Set peeraddr)
target established, actual established, selected peers

TracePromoteColdLocalPeers Int Int (Set peeraddr)
target local established, actual local established, selected peers

TracePromoteColdFailed Int Int peeraddr DiffTime SomeException
target established, actual established, peer, delay until next promotion, reason

TracePromoteColdDone Int Int peeraddr
target established, actual established, peer

TracePromoteWarmPeers Int Int (Set peeraddr)
target active, actual active, selected peers

TracePromoteWarmLocalPeers [(Int, Int)] (Set peeraddr)
local per-group (target active, actual active), selected peers

TracePromoteWarmFailed Int Int peeraddr SomeException
target active, actual active, peer, reason

TracePromoteWarmDone Int Int peeraddr
target active, actual active, peer

TracePromoteWarmAborted Int Int peeraddr
aborted promotion of a warm peer; likely it was asynchronously demoted in the meantime.

target active, actual active, peer

TraceDemoteWarmPeers Int Int (Set peeraddr)
target established, actual established, selected peers

TraceDemoteWarmFailed Int Int peeraddr SomeException
target established, actual established, peer, reason

TraceDemoteWarmDone Int Int peeraddr
target established, actual established, peer

TraceDemoteHotPeers Int Int (Set peeraddr)
target active, actual active, selected peers

TraceDemoteLocalHotPeers [(Int, Int)] (Set peeraddr)
local per-group (target active, actual active), selected peers

TraceDemoteHotFailed Int Int peeraddr SomeException
target active, actual active, peer, reason

TraceDemoteHotDone Int Int peeraddr
target active, actual active, peer

TraceDemoteAsynchronous (Map peeraddr PeerStatus)

TraceGovernorWakeup

TraceChurnWait DiffTime

TraceChurnMode ChurnMode

data DebugPeerSelection peeraddr peerconn

Constructors

TraceGovernorState Time (Maybe DiffTime) (PeerSelectionState peeraddr peerconn)