Ouroboros.Network.PeerSelection.KnownPeers

Types

data KnownPeers peeraddr

The set of known peers. To a first approximation it can be thought of as a Set of peeraddr.

It has two special features:

It tracks which peers we are permitted to gossip with now, or for peers we cannot gossip with now the time at which we would next be allowed to do so.

It tracks the subset of peers that we are happy to publish in reply to gossip requests to our node. It supports random sampling from this set.

invariant ∷ Ord peeraddr ⇒ KnownPeers peeraddr → Bool

Basic container operations

empty ∷ KnownPeers peeraddr

size ∷ KnownPeers peeraddr → Int

insert ∷ Ord peeraddr ⇒ Set peeraddr → KnownPeers peeraddr → KnownPeers peeraddr

delete ∷ Ord peeraddr ⇒ Set peeraddr → KnownPeers peeraddr → KnownPeers peeraddr

toSet ∷ KnownPeers peeraddr → Set peeraddr

O(n)

member ∷ Ord peeraddr ⇒ peeraddr → KnownPeers peeraddr → Bool

Special operations

setCurrentTime ∷ Ord peeraddr ⇒ Time → KnownPeers peeraddr → KnownPeers peeraddr

incrementFailCount ∷ Ord peeraddr ⇒ peeraddr → KnownPeers peeraddr → (Int, KnownPeers peeraddr)

resetFailCount ∷ Ord peeraddr ⇒ peeraddr → KnownPeers peeraddr → KnownPeers peeraddr

lookupFailCount ∷ Ord peeraddr ⇒ peeraddr → KnownPeers peeraddr → Maybe Int

lookupTepidFlag ∷ Ord peeraddr ⇒ peeraddr → KnownPeers peeraddr → Maybe Bool

setTepidFlag ∷ Ord peeraddr ⇒ peeraddr → KnownPeers peeraddr → KnownPeers peeraddr

clearTepidFlag ∷ Ord peeraddr ⇒ peeraddr → KnownPeers peeraddr → KnownPeers peeraddr

Tracking when we can gossip

minGossipTime ∷ Ord peeraddr ⇒ KnownPeers peeraddr → Maybe Time

The first time that a peer will become available for gossip. If peers are already available for gossip, or there are no known peers at all then the result is Nothing.

setGossipTime ∷ Ord peeraddr ⇒ Set peeraddr → Time → KnownPeers peeraddr → KnownPeers peeraddr

availableForGossip ∷ KnownPeers peeraddr → Set peeraddr

The subset of known peers that we would be allowed to gossip with now. This is because we have not gossiped with them recently.

Tracking when we can (re)connect

minConnectTime ∷ Ord peeraddr ⇒ KnownPeers peeraddr → Maybe Time

setConnectTime ∷ Ord peeraddr ⇒ Set peeraddr → Time → KnownPeers peeraddr → KnownPeers peeraddr

availableToConnect ∷ KnownPeers peeraddr → Set peeraddr

The subset of known peers that we would be allowed to try to establish a connection to now. This is because we have not connected with them before or because any failure backoff time has expired.