Ouroboros.Network.PeerSelection.PeerMetric

Peer metrics

data PeerMetrics m p

Mutable peer metrics state accessible via STM.

newtype PeerMetricsConfiguration

Constructors

PeerMetricsConfiguration

Fields

maxEntriesToTrack :: Int

The maximum numbers of slots we will store data for. On some chains sometimes this corresponds to 1h worth of metrics *sighs*.

this number MUST correspond to number of headers / blocks which are produced in one hour.

newPeerMetric :: (MonadLabelledSTM m, NoThunks p, NFData p) => PeerMetricsConfiguration -> m (PeerMetrics m p)

Metric calculations

joinedPeerMetricAt :: forall p m. MonadSTM m => Ord p => PeerMetrics m p -> STM m (Map p SlotNo)

upstreamyness :: forall p m. MonadSTM m => Ord p => PeerMetrics m p -> STM m (Map p Int)

Returns a Map which counts the number of times a given peer was the first to present us with a block/header.

fetchynessBytes :: forall p m. MonadSTM m => Ord p => PeerMetrics m p -> STM m (Map p Int)

Returns a Map which counts the number of bytes downloaded for a given peer.

fetchynessBlocks :: forall p m. MonadSTM m => Ord p => PeerMetrics m p -> STM m (Map p Int)

Returns a Map which counts the number of times a given peer was the first we downloaded a block from.

Tracers

headerMetricTracer :: forall m p. (MonadSTM m, Ord p) => PeerMetricsConfiguration -> PeerMetrics m p -> Tracer (STM m) (TraceLabelPeer (ConnectionId p) (SlotNo, Time))

Tracer which updates header metrics (upstreameness) and inserts new peers into peerRegistry.

fetchedMetricTracer :: forall m p. (MonadSTM m, Ord p) => PeerMetricsConfiguration -> PeerMetrics m p -> Tracer (STM m) (TraceLabelPeer (ConnectionId p) (SizeInBytes, SlotNo, Time))

Tracer which updates fetched metrics (fetchyness) and inserts new peers into peerRegistry.

Metrics reporters

data ReportPeerMetrics (m :: Type -> Type) peerAddr

Constructors

ReportPeerMetrics

Fields

reportHeader :: Tracer (STM m) (TraceLabelPeer peerAddr (SlotNo, Time))

reportFetch :: Tracer (STM m) (TraceLabelPeer peerAddr (SizeInBytes, SlotNo, Time))

nullMetric :: MonadSTM m => ReportPeerMetrics m p

reportMetric :: forall m p. (MonadSTM m, Ord p) => PeerMetricsConfiguration -> PeerMetrics m p -> ReportPeerMetrics m (ConnectionId p)

Internals

type SlotMetric p = IntPSQ SlotNo (p, Time)

Integer based metric ordered by SlotNo which holds the peer and time.

The p parameter is truly polymorphic. For upstreamyness and we use peer address, and for fetchyness it is a pair of peer id and bytes downloaded.

newPeerMetric' :: (MonadLabelledSTM m, NoThunks p, NFData p) => SlotMetric p -> SlotMetric (p, SizeInBytes) -> PeerMetricsConfiguration -> m (PeerMetrics m p)