Ouroboros.Network.PeerSelection.PeerSelectionActions

Documentation

withPeerSelectionActions

Arguments

:: forall peeraddr peerconn resolver exception m a. (Alternative (STM m), MonadAsync m, MonadDelay m, MonadThrow m, MonadMVar m, Ord peeraddr, Exception exception)

=> Tracer m (TraceLocalRootPeers peeraddr exception)

-> Tracer m TracePublicRootPeers

-> Tracer m TraceLedgerPeers

-> (IP -> PortNumber -> peeraddr)

-> DNSActions resolver exception m

-> STM m PeerSelectionTargets

-> STM m [(HotValency, WarmValency, Map RelayAccessPoint PeerAdvertise)]
local root peers

-> STM m (Map RelayAccessPoint PeerAdvertise)
public root peers

-> PeerSharing
peer sharing configured value

-> (peerconn -> PeerSharing)
Extract peer sharing information from peerconn

-> STM m (Map peeraddr (PeerSharingController peeraddr m))
peer sharing registry

-> STM m (peeraddr, PeerSharing)
Read New Inbound Connections

-> PeerStateActions peeraddr peerconn m

-> StdGen
Random generator for picking ledger peers

-> LedgerPeersConsensusInterface m
Get Ledger Peers comes from here

-> STM m UseLedgerAfter
Get Use Ledger After value

-> ((Async m Void, Async m Void) -> PeerSelectionActions peeraddr peerconn m -> m a)
continuation, receives a handle to the local roots peer provider thread (only if local root peers were non-empty).

-> m a

Re-exports

data PeerSelectionTargets

Adjustable targets for the peer selection mechanism.

These are used by the peer selection governor as targets. They are used by the peer churn governor loop as knobs to adjust, to influence the peer selection governor.

The known, established and active peer targets are targets both from below and from above: the governor will attempt to grow or shrink the sets to hit these targets.

Unlike the other targets, the root peer target is "one sided", it is only a target from below. The governor does not try to shrink the root set to hit it, it simply stops looking for more.

There is also an implicit target that enough local root peers are selected as active. This comes from the configuration for local roots, and is not an independently adjustable target.

Constructors

PeerSelectionTargets

Fields

targetNumberOfRootPeers :: !Int

targetNumberOfKnownPeers :: !Int

The target number of all known peers. This includes ledger, big ledger peers.

targetNumberOfEstablishedPeers :: !Int

The target number of established peers (does not include big ledger peers).

The target includes root peers, local root peers, ledger peers and big ledger peers.

targetNumberOfActivePeers :: !Int

The target number of active peers (does not include big ledger peers).

The

targetNumberOfKnownBigLedgerPeers :: !Int

Target number of known big ledger peers.

This target is independent of targetNumberOfKnownPeers. The total number of known peers will be sum of the two targets.

targetNumberOfEstablishedBigLedgerPeers :: !Int

Target number of established big ledger peers.

This target is independent of targetNumberOfEstablishedPeers. The total number of established peers will be sum of the two targets and local root peers.

targetNumberOfActiveBigLedgerPeers :: !Int

Target number of active big ledger peers.

This target is independent of targetNumberOfActivePeers. The total number of active peers will be sum of the two targets and active local root peers.

data PeerAdvertise

Constructors

DoNotAdvertisePeer

DoAdvertisePeer