Ouroboros.Network.PeerSelection.RootPeersDNS.DNSSemaphore

DNS semaphore

data DNSSemaphore m

A semaphore used to limit concurrency of dns names resolution.

newLedgerAndPublicRootDNSSemaphore :: MonadSTM m => m (DNSSemaphore m)

Create a DNSSemaphore for root and ledger peers.

newDNSLocalRootSemaphore :: MonadSTM m => STM m (DNSSemaphore m)

Create a DNSSemaphore for local root peers.

withDNSSemaphore :: (MonadSTM m, MonadThrow m) => DNSSemaphore m -> m a -> m a

Run a computation by attempting to acquire the semaphore first. On termination or failure free the semaphore