Ouroboros.Network.PeerSelection.RootPeersDNS.LocalRootPeers

DNS based provider for local root peers

localRootPeersProvider

Arguments

:: forall m peerAddr resolver exception. (Alternative (STM m), MonadAsync m, MonadDelay m, MonadThrow m, Ord peerAddr)

=> Tracer m (TraceLocalRootPeers peerAddr exception)

-> (IP -> PortNumber -> peerAddr)

-> ResolvConf

-> DNSActions resolver exception m

-> STM m [(HotValency, WarmValency, Map RelayAccessPoint PeerAdvertise)]
input

-> StrictTVar m [(HotValency, WarmValency, Map peerAddr PeerAdvertise)]
output TVar

-> m Void

Resolve RelayAddress-es of local root peers using dns if needed. Local roots are provided wrapped in a StrictTVar, which value might change (re-read form a config file). The resolved dns names are available through the output StrictTVar.

data TraceLocalRootPeers peerAddr exception

Constructors

TraceLocalRootDomains [(HotValency, WarmValency, Map RelayAccessPoint PeerAdvertise)]
Int is the configured valency for the local producer groups

TraceLocalRootWaiting DomainAccessPoint DiffTime

TraceLocalRootResult DomainAccessPoint [(IP, TTL)]

TraceLocalRootGroups [(HotValency, WarmValency, Map peerAddr PeerAdvertise)]
This traces the results of the local root peer provider

TraceLocalRootDNSMap (Map DomainAccessPoint [peerAddr])
This traces the results of the domain name resolution

TraceLocalRootReconfigured

Fields

[(HotValency, WarmValency, Map RelayAccessPoint PeerAdvertise)]

Old value

[(HotValency, WarmValency, Map RelayAccessPoint PeerAdvertise)]

New value

TraceLocalRootFailure DomainAccessPoint (DNSorIOError exception)

TraceLocalRootError DomainAccessPoint SomeException