Ouroboros.Network.PeerSelection.State.EstablishedPeers

Documentation

data EstablishedPeers peeraddr peerconn

The set of established peers. To a first approximation it can be thought of as a Set of peeraddr.

It has one special feature:

It tracks which peers we are permitted to ask for peers now, or for peers we cannot issue share requests with now the time at which we would next be allowed to do so.

empty :: EstablishedPeers peeraddr perconn

toMap :: EstablishedPeers peeraddr peerconn -> Map peeraddr peerconn

O(1)

toSet :: EstablishedPeers peeraddr peerconn -> Set peeraddr

O(n)

readyPeers :: Ord peeraddr => EstablishedPeers peeraddr peerconn -> Set peeraddr

Map of established peers that are either active or ready to be promoted to active.

O(n log m), for n not-ready peers, and m established peers

size :: EstablishedPeers peeraddr peerconn -> Int

The number of established peers. The size of allPeers

O(1)

sizeReady :: EstablishedPeers peeraddr peerconn -> Int

The number of ready peers. The size of readyPeers

O(1)

member :: Ord peeraddr => peeraddr -> EstablishedPeers peeraddr peerconn -> Bool

insert :: Ord peeraddr => peeraddr -> peerconn -> Time -> EstablishedPeers peeraddr peerconn -> EstablishedPeers peeraddr peerconn

Insert a peer into EstablishedPeers.

delete :: Ord peeraddr => peeraddr -> EstablishedPeers peeraddr peerconn -> EstablishedPeers peeraddr peerconn

deletePeers :: Ord peeraddr => Set peeraddr -> EstablishedPeers peeraddr peerconn -> EstablishedPeers peeraddr peerconn

Bulk delete of peers from 'EstablishedPeers.

Special operations

setCurrentTime :: Ord peeraddr => Time -> EstablishedPeers peeraddr peerconn -> EstablishedPeers peeraddr peerconn

setActivateTimes :: Ord peeraddr => Map peeraddr Time -> EstablishedPeers peeraddr peerconn -> EstablishedPeers peeraddr peerconn

Tracking when we can (re)activate

minActivateTime

Arguments

:: Ord peeraddr

=> EstablishedPeers peeraddr peerconn

-> (peeraddr -> Bool)
a predicate which describes the peers to take into account

-> Maybe Time

Find smallest activation time for a peer belonging to a given set.

Tracking when we can peer share

minPeerShareTime :: Ord peeraddr => EstablishedPeers peeraddr peercon -> Maybe Time

The first time that a peer will become available for peer sharing. If peers are already available for peer share, or there are no peers at all then the result is Nothing.

setPeerShareTime :: Ord peeraddr => Set peeraddr -> Time -> EstablishedPeers peeraddr peercon -> EstablishedPeers peeraddr peercon

availableForPeerShare :: EstablishedPeers peeraddr peerconn -> Set peeraddr

The subset of established peers that we would be allowed to peer share with now. This is because we have not peer shared with them recently.

NOTE that this is the set of available peers one would be able to perform peer sharing _now_, it doesn't mean they are 100% eligible. This will depend on other factors like the peer's PeerSharing value.

invariant :: Ord peeraddr => EstablishedPeers peeraddr peerconn -> Bool