Ouroboros.Network.PeerSelection.State.KnownPeers

Types

data KnownPeers peeraddr

The set of known peers. To a first approximation it can be thought of as a Set of peeraddr.

It has one special feature:

It tracks the subset of peers that we are happy to publish in reply to peer share requests to our node. It supports random sampling from this set.

invariant :: Ord peeraddr => KnownPeers peeraddr -> Bool

KnownPeerInfo operations

alterKnownPeerInfo :: (Maybe PeerSharing, Maybe PeerAdvertise, Maybe IsLedgerPeer) -> Maybe KnownPeerInfo -> Maybe KnownPeerInfo

Basic container operations

empty :: KnownPeers peeraddr

size :: KnownPeers peeraddr -> Int

insert :: Ord peeraddr => Map peeraddr (Maybe PeerSharing, Maybe PeerAdvertise, Maybe IsLedgerPeer) -> KnownPeers peeraddr -> KnownPeers peeraddr

This inserts a map of peers with its respective peer sharing, peer advertise and ledger flags into the known peers set.

Please note that if in the map there's an entry for a peer already present in the known peers set, then its values will only be overwritten if they are a Just. Otherwise the current information will be preserved. On the other hand if there's an entry for a peer that isn't a member of the known peer set, the Nothing values will default to NoPeerSharing, DoNotAdvertisePeer and IsNotLedgerPeer, respectively, unless a Just value is used.

alter :: Ord peeraddr => (Maybe KnownPeerInfo -> Maybe KnownPeerInfo) -> Set peeraddr -> KnownPeers peeraddr -> KnownPeers peeraddr

delete :: Ord peeraddr => Set peeraddr -> KnownPeers peeraddr -> KnownPeers peeraddr

toSet :: KnownPeers peeraddr -> Set peeraddr

O(n)

member :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Bool

Special operations

setCurrentTime :: Ord peeraddr => Time -> KnownPeers peeraddr -> KnownPeers peeraddr

incrementFailCount :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> (Int, KnownPeers peeraddr)

resetFailCount :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> KnownPeers peeraddr

lookupFailCount :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Maybe Int

lookupTepidFlag :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Maybe Bool

setTepidFlag :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> KnownPeers peeraddr

clearTepidFlag :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> KnownPeers peeraddr

setSuccessfulConnectionFlag :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> KnownPeers peeraddr

Tracking when we can (re)connect

minConnectTime

Arguments

:: Ord peeraddr

=> KnownPeers peeraddr

-> (peeraddr -> Bool)
a predicate which describes the peers to take into account

-> Maybe Time

setConnectTimes :: Ord peeraddr => Map peeraddr Time -> KnownPeers peeraddr -> KnownPeers peeraddr

availableToConnect :: KnownPeers peeraddr -> Set peeraddr

The subset of known peers that we would be allowed to try to establish a connection to now. This is because we have not connected with them before or because any failure backoff time has expired.

Selecting peers to ask

canPeerShareRequest :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Bool

getPeerSharingRequestPeers :: Ord peeraddr => Set peeraddr -> KnownPeers peeraddr -> Set peeraddr

Filter peers available for Peer Sharing requests, according to their PeerSharing information

Selecting peers to share

canSharePeers :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Bool

getPeerSharingResponsePeers :: KnownPeers peeraddr -> Set peeraddr

Filter peers available for Peer Sharing replies, according to their PeerAdvertise information

Filtering ledger peers

isKnownLedgerPeer :: Ord peeraddr => peeraddr -> KnownPeers peeraddr -> Bool

Select a random subset of the known peers that are available to publish.

The selection is done in such a way that when the same initial PRNG state is used, the selected set does not significantly vary with small perturbations in the set of published peers.

The intention of this selection method is that the selection should give approximately the same replies to the same peers over the course of multiple requests from the same peer. This is to deliberately slow the rate at which peers can discover and map out the entire network.

Checks the KnownPeers Set for known ledger peers.

This is used in Peer Selection Governor to filter out the known-to-te ledger peers from the share result set.