Ouroboros.Network.PeerSelection.State.LocalRootPeers

Types

data LocalRootPeers peeraddr

Constructors

LocalRootPeers (Map peeraddr PeerAdvertise) [(HotValency, WarmValency, Set peeraddr)]

newtype HotValency

Newtype wrapper representing hot valency value from local root group configuration

Constructors

HotValency

Fields

getHotValency :: Int

newtype WarmValency

Newtype wrapper representing warm valency value from local root group configuration

Constructors

WarmValency

Fields

getWarmValency :: Int

invariant :: Ord peeraddr => LocalRootPeers peeraddr -> Bool

Basic operations

empty :: LocalRootPeers peeraddr

null :: LocalRootPeers peeraddr -> Bool

size :: LocalRootPeers peeraddr -> Int

member :: Ord peeraddr => peeraddr -> LocalRootPeers peeraddr -> Bool

hotTarget :: LocalRootPeers peeraddr -> HotValency

warmTarget :: LocalRootPeers peeraddr -> WarmValency

fromGroups :: Ord peeraddr => [(HotValency, WarmValency, Map peeraddr PeerAdvertise)] -> LocalRootPeers peeraddr

The local root peers info has some invariants that are not directly enforced in the types, and the config comes from an external source. Of course it's good to validate that at source, but here we need to not fail if we're given imperfect data.

So what we do is bash it until it is valid. We don't need to be too careful about how we do it, it's ok to be brutal. We should however make sure we trace a warning about dodgy config.

toGroups :: Ord peeraddr => LocalRootPeers peeraddr -> [(HotValency, WarmValency, Map peeraddr PeerAdvertise)]

Inverse of fromGroups, for the subset of inputs to fromGroups that satisfy the invariant.

toGroupSets :: LocalRootPeers peeraddr -> [(HotValency, WarmValency, Set peeraddr)]

toMap :: LocalRootPeers peeraddr -> Map peeraddr PeerAdvertise

keysSet :: LocalRootPeers peeraddr -> Set peeraddr

Special operations

clampToLimit

Arguments

:: Ord peeraddr

=> Int
The limit on the total number of local peers

-> LocalRootPeers peeraddr

-> LocalRootPeers peeraddr

Limit the size of the root peers collection to fit within given bounds.

The governor needs to be able to do this to enforce its invariant that:

LocalRootPeers.size localRootPeers <= targetNumberOfKnownPeers

It needs to be able to establish that invariant given arbitrary configuration for local root peers. It makes sense to do it this way rather than just enforce that local root peers config fits the invariant because the invariant depends on both the targets and the local root peers config and these can both vary dynamically and independently.

It is unlikely in practice that there are so many local root peers configured that it goes over this targets, so it's ok to resolve it pretty arbitrarily. We just take the local roots in left to right order up to the limit. So we have the property that

LocalRootPeers.size (LocalRootPeers.clampToLimit sz lrps)
 == min sz (LocalRootPeers.size lrps)