Ouroboros.Network.Protocol.BlockFetch.Client

Documentation

newtype BlockFetchClient block m a

Block fetch client type for requesting ranges of blocks and handling responses.

Constructors

BlockFetchClient

Fields

runBlockFetchClient ∷ m (BlockFetchRequest block m a)

data BlockFetchRequest block m a where

Constructors

SendMsgRequestRange ∷ ChainRange block → BlockFetchResponse block m a → BlockFetchClient block m a → BlockFetchRequest block m a
Request a chain range, supply handler for incoming blocks and a continuation.

SendMsgClientDone ∷ a → BlockFetchRequest block m a
Client terminating the block-fetch protocol.

data BlockFetchResponse block m a

Constructors

BlockFetchResponse

Fields

handleStartBatch ∷ m (BlockFetchReceiver block m)

handleNoBlocks ∷ m ()

data BlockFetchReceiver block m

Block are streamed and block receiver will handle each one when it comes, it also needs to handle errors send back from the server.

Constructors

BlockFetchReceiver

Fields

handleBlock ∷ block → m (BlockFetchReceiver block m)

handleBatchDone ∷ m ()

blockFetchClientPeer ∷ ∀ block m a. (StandardHash block, Monad m) ⇒ BlockFetchClient block m a → Peer (BlockFetch block) AsClient BFIdle m a

data BlockFetchClientPipelined block m a where

A BlockFetch client designed for running the protcol in a pipelined way.

Constructors

BlockFetchClientPipelined ∷ BlockFetchSender Z c block m a → BlockFetchClientPipelined block m a
A BlockFetchSender, but starting with zero outstanding pipelined responses, and for any internal collect type c.

data BlockFetchSender n c block m a where

A BlockFetchSender with n outstanding stream of block bodies.

Constructors

SendMsgRequestRangePipelined ∷ ChainRange block → c → (Maybe block → c → m c) → BlockFetchSender (S n) c block m a → BlockFetchSender n c block m a
Send a MsgRequestRange but do not wait for response. Supply a monadic action which runs on each received block and which updates the internal received value c. c could be a Monoid, though it's more genral this way.

CollectBlocksPipelined ∷ Maybe (BlockFetchSender (S n) c block m a) → (c → BlockFetchSender n c block m a) → BlockFetchSender (S n) c block m a
Collect the result of a previous pipelined receive action

SendMsgDonePipelined ∷ a → BlockFetchSender Z c block m a
Termination of the block-fetch protocol.

blockFetchClientPeerPipelined ∷ ∀ block m a. Monad m ⇒ BlockFetchClientPipelined block m a → PeerPipelined (BlockFetch block) AsClient BFIdle m a

blockFetchClientPeerSender ∷ ∀ n block c m a. Monad m ⇒ BlockFetchSender n c block m a → PeerSender (BlockFetch block) AsClient BFIdle n c m a