Ouroboros.Network.Protocol.ChainSync.Client

Description

A view of the chain synchronisation protocol from the point of view of the client.

This provides a view that uses less complex types and should be easier to use than the underlying typed protocol itself.

For execution, a conversion into the typed protocol is provided.

Protocol type for the client

The protocol states from the point of view of the client.

newtype ChainSyncClient header tip m a

A chain sync protocol client, on top of some effect m. The first choice of request is within that m.

Constructors

ChainSyncClient

Fields

runChainSyncClient ∷ m (ClientStIdle header tip m a)

data ClientStIdle header tip m a where

In the StIdle protocol state, the server does not have agency and can choose to send a request next, or a find intersection message.

Constructors

SendMsgRequestNext ∷ ClientStNext header tip m a → m (ClientStNext header tip m a) → ClientStIdle header tip m a
Send the MsgRequestNext, with handlers for the replies.

The handlers for this message are more complicated than most RPCs because the server can either send us a reply immediately or it can send us a MsgAwaitReply to indicate that the server itself has to block for a state change before it can send us the reply.

In the waiting case, the client gets the chance to take a local action.

SendMsgFindIntersect ∷ [Point header] → ClientStIntersect header tip m a → ClientStIdle header tip m a
Send the MsgFindIntersect, with handlers for the replies.

SendMsgDone ∷ a → ClientStIdle header tip m a
The client decided to end the protocol.

data ClientStNext header tip m a

In the StNext protocol state, the client does not have agency and is waiting to receive either

a roll forward,

roll back message,

It must be prepared to handle any of these.

Constructors

ClientStNext

Fields

recvMsgRollForward ∷ header → tip → ChainSyncClient header tip m a

recvMsgRollBackward ∷ Point header → tip → ChainSyncClient header tip m a

data ClientStIntersect header tip m a

In the StIntersect protocol state, the client does not have agency and is waiting to receive:

an intersection improved,

unchanged message,

the termination message.

It must be prepared to handle any of these.

Constructors

ClientStIntersect

Fields

recvMsgIntersectFound ∷ Point header → tip → ChainSyncClient header tip m a

recvMsgIntersectNotFound ∷ tip → ChainSyncClient header tip m a

Execution as a typed protocol

chainSyncClientPeer ∷ ∀ header tip m a. Monad m ⇒ ChainSyncClient header tip m a → Peer (ChainSync header tip) AsClient StIdle m a

Interpret a ChainSyncClient action sequence as a Peer on the client side of the ChainSyncProtocol.

Null chain sync client

chainSyncClientNull ∷ MonadTimer m ⇒ ChainSyncClient header tip m a

A chain sync client which never sends any message.