Ouroboros.Network.Protocol.ChainSync.Server

Description

A view of the chain synchronisation protocol from the point of view of the server.

This provides a view that uses less complex types and should be easier to use than the underlying typed protocol itself.

For execution, a conversion into the typed protocol is provided.

Protocol type for the server

The protocol states from the point of view of the server.

newtype ChainSyncServer header tip m a

A chain sync protocol server, on top of some effect m.

Constructors

ChainSyncServer

Fields

runChainSyncServer ∷ m (ServerStIdle header tip m a)

data ServerStIdle header tip m a

In the StIdle protocol state, the server does not have agency. Instead it is waiting for:

a next update request

a find intersection request

a termination messge

It must be prepared to handle either.

Constructors

ServerStIdle

Fields

recvMsgRequestNext ∷ m (Either (ServerStNext header tip m a) (m (ServerStNext header tip m a)))

recvMsgFindIntersect ∷ [Point header] → m (ServerStIntersect header tip m a)

recvMsgDoneClient ∷ m a

data ServerStNext header tip m a where

In the StNext protocol state, the server has agency and must send either:

a roll forward

a roll back message

a termination message

Constructors

SendMsgRollForward ∷ header → tip → ChainSyncServer header tip m a → ServerStNext header tip m a

SendMsgRollBackward ∷ Point header → tip → ChainSyncServer header tip m a → ServerStNext header tip m a

data ServerStIntersect header tip m a where

In the StIntersect protocol state, the server has agency and must send either:

an intersection improved,

unchanged message,

termination message

Constructors

SendMsgIntersectFound ∷ Point header → tip → ChainSyncServer header tip m a → ServerStIntersect header tip m a

SendMsgIntersectNotFound ∷ tip → ChainSyncServer header tip m a → ServerStIntersect header tip m a

Execution as a typed protocol

chainSyncServerPeer ∷ ∀ header tip m a. Monad m ⇒ ChainSyncServer header tip m a → Peer (ChainSync header tip) AsServer StIdle m a

Interpret a ChainSyncServer action sequence as a Peer on the server side of the ChainSyncProtocol.