Ouroboros.Network.Protocol.KeepAlive.Type

Description

The type of the keep alive protocol.

The keep alive protocol is used for

sending keep alive messages

making round trip measuremnets

Each side will run its own version of the keep alive protocol. It should be configured so that any intermediate state (such as in customer premise equipment or in a carrier grade NAT) is kept alive. This has to be a per-node configuration element as this is about the properties of that nodes network connectivity.

For making round trip measurements its in the interest of the other side to reply promptly.

Documentation

newtype Cookie

A 16bit value used to match responses to requests.

Constructors

Cookie

Fields

unCookie ∷ Word16

data KeepAliveProtocolFailure

Constructors

KeepAliveCookieMissmatch Cookie Cookie

data KeepAlive where

A kind to identify our protocol, and the types of the states in the state transition diagram of the protocol.

Constructors

StClient ∷ KeepAlive
The client can send a request and the server is waiting for a request.

StServer ∷ KeepAlive
The server is responsible for sending response back.

StDone ∷ KeepAlive
Both the client and server are in the terminal state. They're done.