Ouroboros.Network.Protocol.LocalStateQuery.Server

Protocol type for the server

The protocol states from the point of view of the server.

newtype LocalStateQueryServer block (query ∷ Type → Type) m a

Constructors

LocalStateQueryServer

Fields

runLocalStateQueryServer ∷ m (ServerStIdle block query m a)

data ServerStIdle block query m a

In the StIdle protocol state, the server does not have agency. Instead it is waiting for:

a request to acquire a state

a termination messge

It must be prepared to handle either.

Constructors

ServerStIdle

Fields

recvMsgAcquire ∷ Point block → m (ServerStAcquiring block query m a)

recvMsgDone ∷ m a

data ServerStAcquiring block query m a where

In the StAcquiring protocol state, the server has agency and must send either:

acquired

failure to acquire

Constructors

SendMsgAcquired ∷ ServerStAcquired block query m a → ServerStAcquiring block query m a

SendMsgFailure ∷ AcquireFailure → ServerStIdle block query m a → ServerStAcquiring block query m a

data ServerStAcquired block query m a

In the StAcquired protocol state, the server does not have agency. Instead it is waiting for:

a query

a request to (re)acquire another state

a release of the current state

It must be prepared to handle either.

Constructors

ServerStAcquired

Fields

recvMsgQuery ∷ ∀ result. query result → m (ServerStQuerying block query m a result)

recvMsgReAcquire ∷ Point block → m (ServerStAcquiring block query m a)

recvMsgRelease ∷ m (ServerStIdle block query m a)

data ServerStQuerying block query m a result where

In the StQuerying protocol state, the server has agency and must send:

a result

Constructors

SendMsgResult ∷ result → ServerStAcquired block query m a → ServerStQuerying block query m a result

Execution as a typed protocol

localStateQueryServerPeer ∷ ∀ block (query ∷ Type → Type) m a. Monad m ⇒ LocalStateQueryServer block query m a → Peer (LocalStateQuery block query) AsServer StIdle m a

Interpret a LocalStateQueryServer action sequence as a Peer on the server side of the LocalStateQuery protocol.