Ouroboros.Network.Protocol.LocalStateQuery.Type

Description

The type of the local ledger state query protocol.

This is used by local clients (like wallets and CLI tools) to query the ledger state of a local node.

Documentation

data LocalStateQuery block query where

The kind of the local state query protocol, and the types of the states in the protocol state machine.

It is parametrised over the type of block (for points), the type of queries and query results.

Constructors

StIdle ∷ LocalStateQuery block query
The client has agency. It can ask to acquire a state or terminate.

There is no timeout in this state.

StAcquiring ∷ LocalStateQuery block query
The server has agency. it must acquire the state at the requested point or report a failure.

There is a timeout in this state.

StAcquired ∷ LocalStateQuery block query
The client has agency. It can request queries against the current state, or it can release the state.

StQuerying ∷ result → LocalStateQuery block query
The server has agency. It must respond with the query result.

StDone ∷ LocalStateQuery block query
Nobody has agency. The terminal state.

data AcquireFailure

Constructors

AcquireFailurePointTooOld

AcquireFailurePointNotOnChain

class (∀ result. Show (query result)) ⇒ ShowQuery query where

To implement Show for:

('Message' ('LocalStateQuery' block query) st st')

we need a way to print the query GADT and its type index, result. This class contain the method we need to provide this Show instance.

We use a type class for this, as this Show constraint propagates to a lot of places.

Methods

showResult ∷ ∀ result. query result → result → String