Ouroboros.Network.Protocol.LocalTxMonitor.Type

Description

The type of the local transaction monitoring protocol.

This is used by local clients (like wallets, explorers and CLI tools) to monitor the transactions passing through the mempool of a local node.

The protocol is stateful such that the server keeps track of the transactions already sent to the client.

                   START
                     ⇓
                   ┌───────────────┐
           ┌──────▶│     Idle      │⇒ DONE
           │       └───┬───────────┘
           │           │
           │   Acquire │
           │           ▼
           │       ┌───────────────┐
   Release │       │   Acquiring   │
           │       └───┬───────────┘
           │           │       ▲
           │  Acquired │       │ AwaitAcquire
           │           ▼       │
           │       ┌───────────┴───┐
           └───────┤   Acquired    │
                   └───┬───────────┘
                       │       ▲
 HasTx|NextTx|GetSizes │       │ Reply (HasTx|NextTx|GetSizes)
                       ▼       │
                   ┌───────────┴───┐
                   │      Busy     │
                   └───────────────┘

Documentation

data LocalTxMonitor txid tx slot where

The kind of the local transaction monitoring protocol, and the types of the states in the protocol state machine.

It is parametrised over the type of transactions.

Constructors

StIdle ∷ LocalTxMonitor txid tx slot
The client has agency; it can request a transaction or terminate.

There is no timeout in this state.

StAcquiring ∷ LocalTxMonitor txid tx slot
The server has agency; it is capturing the latest mempool snapshot.

StAcquired ∷ LocalTxMonitor txid tx slot
The client has agency; The server is locked on a particular mempool snapshot. The client can now perform various requests on that snapshot, or acquire a new one, more recent.

StBusy ∷ StBusyKind → LocalTxMonitor txid tx slot
The server has agency; It must respond, there's no timeout.

StDone ∷ LocalTxMonitor txid tx slot
Nobody has agency. The terminal state.

data StBusyKind where

Constructors

NextTx ∷ StBusyKind
The server is busy fetching the next transaction from the mempool

HasTx ∷ StBusyKind
The server is busy looking for the presence of a specific transaction in the mempool

GetSizes ∷ StBusyKind
The server is busy looking for the current size and max capacity of the mempool

data MempoolSizeAndCapacity

Describes the MemPool sizes and capacity for a given snapshot.

Constructors

MempoolSizeAndCapacity

Fields

capacityInBytes ∷ !Word32

The maximum capacity of the mempool. Note that this may dynamically change when the ledger state is updated.

sizeInBytes ∷ !Word32

The summed byte size of all the transactions in the mempool.

numberOfTxs ∷ !Word32

The number of transactions in the mempool

data TokBusyKind (k ∷ StBusyKind) where

Constructors

TokNextTx ∷ TokBusyKind NextTx

TokHasTx ∷ TokBusyKind HasTx

TokGetSizes ∷ TokBusyKind GetSizes