Ouroboros.Network.Protocol.LocalTxSubmission.Type

Description

The type of the local transaction submission protocol.

This is used by local clients (like wallets and CLI tools) to submit transactions to a local node.

Documentation

data LocalTxSubmission tx reject where

The kind of the local transaction-submission protocol, and the types of the states in the protocol state machine.

It is parametrised over the type of transactions and the type of reasons used when rejecting a transaction.

Constructors

StIdle ∷ LocalTxSubmission tx reject
The client has agency; it can submit a transaction or terminate.

There is no timeout in this state.

StBusy ∷ LocalTxSubmission tx reject
The server has agency; it must process the submitted transaction and either accept or reject it (with a reason).

There is a timeout in this state. If the mempool is full and remains so for a period then the transaction should be rejected with a suitable temporary failure reason.

StDone ∷ LocalTxSubmission tx reject
Nobody has agency. The terminal state.