Ouroboros.Network.Protocol.TipSample.Server

Protocol type for the server

newtype TipSampleServer tip m a

A TipSample protocol server .

Constructors

TipSampleServer

Fields

runTipSampleServer ∷ m (ServerStIdle tip m a)

RunTipSampleServer must not block for a long time.

data ServerStIdle tip m a

In StIdle protocol state the server does not have agency, it is await for one of:

MsgFollowTip

MsgDone

The ServerStIdle contains handlers for all these messages.

Constructors

ServerStIdle

Fields

handleFollowTip ∷ ∀ (n ∷ N). Nat (S n) → SlotNo → m (SendTips (S n) tip m a)

The server is requested to return the first tip after SlotNo (at or after to be precise); handleTipAfterSlotNo might block.

handleDone ∷ m a

data SendTips (n ∷ N) tip m a where

Constructors

SendNextTip ∷ Nat (S (S n)) → tip → m (SendTips (S n) tip m a) → SendTips (S (S n)) tip m a

SendLastTip ∷ Nat (S Z) → tip → ServerStIdle tip m a → SendTips (S Z) tip m a

Execution as a typed protocol

tipSampleServerPeer ∷ ∀ tip m a. Monad m ⇒ TipSampleServer tip m a → Peer (TipSample tip) AsServer StIdle m a

Interpret TipSampleServer action sequence as a Peer on the server side of the TipSample protocol.