Ouroboros.Network.Protocol.TipSample.Type

Documentation

data TipRequestKind where

There are three of blocking requests: awiat until slot, await until a tip will change; the third one has a dedicated type: StFollowTip :: TipSample tip.

Constructors

BlockUntilSlot ∷ TipRequestKind

BlockUntilTip ∷ TipRequestKind

data TokTipRequest (tipRequestKind ∷ TipRequestKind) where

Constructors

TokBlockUntilSlot ∷ TokTipRequest BlockUntilSlot

TokBlockUntilTip ∷ TokTipRequest BlockUntilTip

data TipSample tip where

`tip-sample` protocol, desigined for established peers to sample tip from upstream peers.

Constructors

StIdle ∷ TipSample tip

StFollowTip ∷ N → TipSample tip

StDone ∷ TipSample tip