Ouroboros.Network.Protocol.Trans.Hello.Codec

Documentation

codecHello

Arguments

∷ ∀ ps (stIdle ∷ ps) m. (MonadST m, ShowProxy ps, ShowProxy stIdle, ∀ (st ∷ ps). Show (ClientHasAgency st), ∀ (st ∷ ps). Show (ServerHasAgency st))

⇒ Word
tag for MsgHello

→ (∀ (pr ∷ PeerRole) (st ∷ ps) (st' ∷ ps). PeerHasAgency pr st → Message ps st st' → Encoding)

→ (∀ (pr ∷ PeerRole) (st ∷ ps) s. PeerHasAgency pr st → Int → Word → Decoder s (SomeMessage st))
continuation which receives agency, tag, and length. Last two are decoded values.

→ Codec (Hello ps stIdle) DeserialiseFailure m ByteString

codecHello provides a flat encoding of the original codec and MsgHello. MsgTalk is invisible. It assumes that the top level encoding of ps protocol is a list of at least one element (the tag of a message).

byteLimitsHello ∷ ∀ ps (stIdle ∷ ps) bytes. ProtocolSizeLimits ps bytes → ProtocolSizeLimits (Hello ps stIdle) bytes

timeLimitsHello ∷ ∀ ps (stIdle ∷ ps). ProtocolTimeLimits ps → ProtocolTimeLimits (Hello ps stIdle)

codecHelloId ∷ ∀ ps (stIdle ∷ ps) m. (Monad m, ∀ (st ∷ ps). Show (ClientHasAgency st), ∀ (st ∷ ps). Show (ServerHasAgency st)) ⇒ Codec ps CodecFailure m (AnyMessage ps) → Codec (Hello ps stIdle) CodecFailure m (AnyMessage (Hello ps stIdle))