Ouroboros.Network.Protocol.Trans.Hello.Type

Documentation

data Hello ps (stIdle ∷ ps) where

The Hello protocol transformer reverses the initial agency of the protocol, from one in which the server has the initial agency to one in which the client has the initial agency.

It extends the underlying protocol with an initial extra MsgHello message, which reverses the agency between the two peers.

Constructors

StHello ∷ Hello ps stIdle
StHello state is the initial state of the Hello protocol.

StTalk ∷ ps → Hello ps stIdle
StTalk embeds states of the inner protocol.