Ouroboros.Network.Protocol.Trans.Hello.Util

Description

Utility module which provides wrappers which map a Peer or PiplinedPeer of an original protocol ps into Peer / PiplinedPeer of Hello ps stIdle protocol.

The module should be imported qualified.

Wrap Peer in Hello protocol

wrapClientPeer ∷ ∀ ps (stIdle ∷ ps) m a. Functor m ⇒ Peer ps AsClient stIdle m a → Peer (Hello ps stIdle) AsClient StHello m a

wrapServerPeer ∷ ∀ ps (stIdle ∷ ps) m a. Functor m ⇒ Peer ps AsServer stIdle m a → Peer (Hello ps stIdle) AsServer (StTalk stIdle) m a

Wrap PipelinedPeer in Hello protcol.

wrapClientPeerPipelined ∷ ∀ ps (stIdle ∷ ps) m a. Functor m ⇒ PeerPipelined ps AsClient stIdle m a → PeerPipelined (Hello ps stIdle) AsClient StHello m a

wrapServerPeerPipelined ∷ ∀ ps (stIdle ∷ ps) m a. Functor m ⇒ PeerPipelined ps AsServer stIdle m a → PeerPipelined (Hello ps stIdle) AsServer StHello m a