Ouroboros.Network.Protocol.TxSubmission.Client

Description

A view of the transaction submission protocol from the point of view of the client.

This provides a view that uses less complex types and should be easier to use than the underlying typed protocol itself.

For execution, txSubmissionClientPeer is provided for conversion into the typed protocol.

Protocol type for the client

The protocol states from the point of view of the client.

newtype TxSubmissionClient txid tx m a

The client side of the transaction submission protocol.

The peer in the client role submits transactions to the peer in the server role.

Constructors

TxSubmissionClient

Fields

runTxSubmissionClient ∷ m (ClientStIdle txid tx m a)

data ClientStIdle txid tx m a

In the StIdle protocol state, the client does not have agency. Instead it is waiting for:

a request for transaction ids (blocking or non-blocking)

a request for a given list of transactions

a termination message

It must be prepared to handle any of these.

Constructors

ClientStIdle

Fields

recvMsgRequestTxIds ∷ ∀ blocking. TokBlockingStyle blocking → Word16 → Word16 → m (ClientStTxIds blocking txid tx m a)

recvMsgRequestTxs ∷ [txid] → m (ClientStTxs txid tx m a)

data ClientStTxIds blocking txid tx m a where

Constructors

SendMsgReplyTxIds ∷ BlockingReplyList blocking (txid, TxSizeInBytes) → ClientStIdle txid tx m a → ClientStTxIds blocking txid tx m a

SendMsgDone ∷ a → ClientStTxIds StBlocking txid tx m a
In the blocking case, the client can terminate the protocol. This could be used when the client knows there will be no more transactions to submit.

data ClientStTxs txid tx m a where

Constructors

SendMsgReplyTxs ∷ [tx] → ClientStIdle txid tx m a → ClientStTxs txid tx m a

type TxSizeInBytes = Word32

Transactions are typically not big, but in principle in future we could have ones over 64k large.

data TokBlockingStyle (k ∷ StBlockingStyle) where

The value level equivalent of StBlockingStyle.

This is also used in MsgRequestTxIds where it is interpreted (and can be encoded) as a Bool with True for blocking, and False for non-blocking.

Constructors

TokBlocking ∷ TokBlockingStyle StBlocking

TokNonBlocking ∷ TokBlockingStyle StNonBlocking

data BlockingReplyList (blocking ∷ StBlockingStyle) a where

We have requests for lists of things. In the blocking case the corresponding reply must be non-empty, whereas in the non-blocking case and empty reply is fine.

Constructors

BlockingReply ∷ NonEmpty a → BlockingReplyList StBlocking a

NonBlockingReply ∷ [a] → BlockingReplyList StNonBlocking a

Execution as a typed protocol

txSubmissionClientPeer ∷ ∀ txid tx m a. Monad m ⇒ TxSubmissionClient txid tx m a → Peer (TxSubmission txid tx) AsClient StIdle m a

A non-pipelined Peer representing the TxSubmissionClient.