Ouroboros.Network.Testing.ConcreteBlock

Description

Concrete block

The network library should not export a concrete block type at all, except that it might need one in its tests (but not exported). Right now this module serves to isolate this in a specific module so we can identify easily where it is used; eventually it should be simplified and then moved to the network layer tests; the more sophiscated block abstraction (abstracted over an Ouroboros protocol) will live in the consensus layer.

Documentation

data Block

Our highly-simplified version of a block. It retains the separation between a block header and body, which is a detail needed for the protocols.

Constructors

Block

Fields

blockHeader ∷ BlockHeader

blockBody ∷ BlockBody

data BlockHeader

A block header. It retains simplified versions of all the essential elements.

Constructors

BlockHeader

Fields

headerHash ∷ HeaderHash BlockHeader

The cached ConcreteHeaderHash of this header.

headerPrevHash ∷ ChainHash BlockHeader

The headerHash of the previous block header

headerSlot ∷ SlotNo

The Ouroboros time slot index of this block

headerBlockNo ∷ BlockNo

The block index from the Genesis

headerBodyHash ∷ BodyHash

The hash of the corresponding block body

newtype BlockBody

Constructors

BlockBody String

hashHeader ∷ BlockHeader → ConcreteHeaderHash

Compute the ConcreteHeaderHash of the BlockHeader.

newtype BodyHash

The hash of all the information in a BlockBody.

Constructors

BodyHash Int

newtype ConcreteHeaderHash

The hash of all the information in a BlockHeader.

Constructors

HeaderHash Int

hashBody ∷ BlockBody → BodyHash

Creating sample chains

mkChain ∷ [(SlotNo, BlockBody)] → Chain Block

This takes the blocks in order from oldest to newest.

mkChainSimple ∷ [BlockBody] → Chain Block

mkChainFragment ∷ ChainHash Block → BlockNo → [(SlotNo, BlockBody)] → ChainFragment Block

mkChainFragmentSimple ∷ [BlockBody] → ChainFragment Block

mkAnchoredFragment ∷ Point Block → BlockNo → [(SlotNo, BlockBody)] → AnchoredFragment Block

mkAnchoredFragmentSimple ∷ [BlockBody] → AnchoredFragment Block

Generator utilities

mkPartialBlock ∷ SlotNo → BlockBody → Block

mkPartialBlockHeader ∷ SlotNo → BlockBody → BlockHeader

fixupBlock ∷ HeaderHash block ~ HeaderHash BlockHeader ⇒ ChainHash block → BlockNo → Block → Block

Fixup block so to fit it on top of a chain. Only block number, previous hash and block hash are updated; slot number and signers are kept intact.

fixupBlockHeader ∷ HeaderHash block ~ HeaderHash BlockHeader ⇒ ChainHash block → BlockNo → BlockHeader → BlockHeader

Fixup block header to fit it on top of a chain. Only block number and previous hash are updated; the slot and signer are kept unchanged.

fixupBlockAfterBlock ∷ Block → Block → Block

Fixup a block so to fit it on top of a given previous block.

fixupChain ∷ HasHeader b ⇒ (ChainHash b → BlockNo → b → b) → [b] → Chain b

Fix up the block number and hashes of a Chain. This also fixes up the first block to chain-on from genesis, since by construction the Chain type starts from genesis.

fixupChainFragmentFrom ∷ HasHeader b ⇒ ChainHash b → BlockNo → (ChainHash b → BlockNo → b → b) → [b] → ChainFragment b

fixupChainFragmentFromGenesis ∷ HasHeader b ⇒ (ChainHash b → BlockNo → b → b) → [b] → ChainFragment b

fixupChainFragmentFromSame ∷ HasHeader b ⇒ (ChainHash b → BlockNo → b → b) → [b] → ChainFragment b

fixupAnchoredFragmentFrom ∷ HasHeader b ⇒ Point b → BlockNo → (ChainHash b → BlockNo → b → b) → [b] → AnchoredFragment b

Orphan instances

Hashable SlotNo

Instance details

Methods

hashWithSalt ∷ Int → SlotNo → Int

hash ∷ SlotNo → Int

Hashable BlockNo

Instance details

Methods

hashWithSalt ∷ Int → BlockNo → Int

hash ∷ BlockNo → Int

Hashable (HeaderHash b) ⇒ Hashable (ChainHash b)
Hashable instance for Hash

We don't insist that Hashable in StandardHash because Hashable is only used in the network layer tests.

This requires UndecidableInstances because Hashable (HeaderHash b) is no smaller than Hashable (ChainHash b).

Instance details

Methods

hashWithSalt ∷ Int → ChainHash b → Int

hash ∷ ChainHash b → Int