Ouroboros.Network.Tracers

Documentation

data NetworkSubscriptionTracers withIPList addr vNumber

IP subscription tracers.

Constructors

NetworkSubscriptionTracers

Fields

nsMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nsHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nsErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nsSubscriptionTracer ∷ Tracer IO (withIPList (SubscriptionTrace addr))

subscription tracers; it is infrequent it should not be nullTracer by default.

type NetworkIPSubscriptionTracers addr vNumber = NetworkSubscriptionTracers WithIPList addr vNumber

nullNetworkSubscriptionTracers ∷ NetworkSubscriptionTracers withIPList addr vNumber

data NetworkDNSSubscriptionTracers vNumber addr

DNS subscription tracers.

Constructors

NetworkDNSSubscriptionTracers

Fields

ndstMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

ndstHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

ndstErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

ndstSubscriptionTracer ∷ Tracer IO (WithDomainName (SubscriptionTrace addr))

subscription tracer; it is infrequent it should not be nullTracer by default.

ndstDnsTracer ∷ Tracer IO (WithDomainName DnsTrace)

dns resolver tracer; it is infrequent it should not be nullTracer by default.

nullNetworkDNSSubscriptionTracers ∷ NetworkDNSSubscriptionTracers vNumber peerid