Ouroboros.Network.AnchoredFragment

AnchoredFragment type and fundamental operations

data AnchoredFragment block where

An AnchoredFragment is a ChainFragment that is anchored to some Point: the point right before the first, leftmost block in the fragment. The anchor point can be thought of as a left exclusive bound.

For example, the following fragment is anchored at a and contains b1, b2, and b3, which is the head of the fragment.

a ] b1 >: b2 >: b3

The fact that it is an exclusive bound is particularly convenient when dealing with Genesis. Genesis is the start of the chain, but not an actual block, so we cannot use it an inclusive bound. However, there is a Point that refers to Genesis (genesisPoint), which can be used as the anchor point, acting as an exclusive bound.

An AnchoredFragment anchored at Genesis, can thus be converted to a Chain (toChain), containing all blocks starting from Genesis.

Without an anchor point, an empty ChainFragment doesn't give us much more information: is it empty because the whole chain is empty? Or, did we just get an empty fragment that was split off from some later part of the chain?

Furthermore, an important criterion during chain selection is the length of the chain. However, comparing the length of two ChainFragments doesn't make much sense, since they might start at different points. Hence the idea to "anchor" fragments at some point, and require that fragments have the same anchor when comparing their lengths.

Note: instead of comparing the lengths of two ChainFragments, we could try to compare the blockNo of their heads, which is also a measure of the total length of the chain. However, EBBs throw a(nother) spanner in the works: each EBB shares its blockNo with a regular block, so comparing a ChainFragment that only contains the EBB with a ChainFragment that only contains the regular block with the same blockNo as the EBB, will not give a conclusive preference to either fragment, while in reality one fragment actually corresponds to a longer chain.

Bundled Patterns

pattern Empty ∷ HasHeader block ⇒ Anchor block → AnchoredFragment block
\( O(1) \). Pattern for matching on or creating an empty AnchoredFragment. An empty fragment has/needs an anchor point.

pattern (:>) ∷ (HasHeader block, HasCallStack) ⇒ AnchoredFragment block → block → AnchoredFragment block infixl 5
\( O(1) \). Add a block to the right of the anchored fragment.

pattern (:<) ∷ HasHeader block ⇒ block → AnchoredFragment block → AnchoredFragment block infixl 5
\( O(1) \). View the first, leftmost block of the anchored fragment.

This is only a view, not a constructor, as adding a block to the left would change the anchor of the fragment.

anchor ∷ AnchoredFragment block → Anchor block

anchorPoint ∷ AnchoredFragment block → Point block

anchorBlockNo ∷ AnchoredFragment block → WithOrigin BlockNo

unanchorFragment ∷ AnchoredFragment block → ChainFragment block

mkAnchoredFragment ∷ HasHeader block ⇒ Anchor block → ChainFragment block → AnchoredFragment block

data Anchor block

Anchor of an AnchoredFragment

Constructors

AnchorGenesis
The fragment is anchored at genesis

Anchor !SlotNo !(HeaderHash block) !BlockNo
The fragment is anchored after genesis

We don't use the Point type directly as that has its own use of WithOrigin, and we want to enforce here that we have a block number if and only if the point is not Origin.

anchorFromBlock ∷ HasHeader block ⇒ block → Anchor block

Construct anchor from a block

In other words, this would be the block immediately before the other blocks in the fragment.

anchorFromPoint ∷ Point block → BlockNo → Anchor block

Construct an anchor from a point

In this case, we must also be given the BlockNo. This only makes sense for points that aren't genesis.

anchorToPoint ∷ Anchor block → Point block

Compute which Point this anchor corresponds to

anchorToSlotNo ∷ Anchor block → WithOrigin SlotNo

Extract the SlotNo from the anchor

anchorToBlockNo ∷ Anchor block → WithOrigin BlockNo

Extract the BlockNo from the anchor

NOTE: When the Anchor is AnchorGenesis, this returns Origin. It does not return genesisBlockNo, which is badly named, and is instead the block number of the first block on the chain (i.e., genesisPoint and genesisBlockNo don't go hand in hand!)

anchorToHash ∷ Anchor block → ChainHash block

Extract the hash from the anchor

Returns GenesisHash if the anchor is AnchorGenesis.

anchorIsGenesis ∷ Anchor block → Bool

Does this anchor represent genesis (i.e., empty chain)?

anchorToTip ∷ HeaderHash a ~ HeaderHash b ⇒ Anchor a → Tip b

Translate Anchor to Tip

Right now this is in fact an isomorphism, but these two types are logically independent.

castAnchor ∷ HeaderHash a ~ HeaderHash b ⇒ Anchor a → Anchor b

The equivalent of castPoint for Anchor

valid ∷ HasHeader block ⇒ AnchoredFragment block → Bool

\( O(n) \).

validExtension ∷ HasHeader block ⇒ AnchoredFragment block → block → Bool

\( O(1) \).

Block re-exports

class (StandardHash b, Measured BlockMeasure b, Typeable b) ⇒ HasHeader b where

Abstract over the shape of blocks (or indeed just block headers)

Methods

blockHash ∷ b → HeaderHash b

blockPrevHash ∷ b → ChainHash b

blockSlot ∷ b → SlotNo

blockNo ∷ b → BlockNo

blockInvariant ∷ b → Bool

newtype Point block

A point on the chain is identified by its Slot and HeaderHash.

The Slot tells us where to look and the HeaderHash either simply serves as a check, or in some contexts it disambiguates blocks from different forks that were in the same slot.

It's a newtype rather than a type synonym, because using a type synonym would lead to ambiguity, since HeaderHash is a non-injective type family.

Constructors

Point

Fields

getPoint ∷ WithOrigin (Block SlotNo (HeaderHash block))

castPoint ∷ HeaderHash a ~ HeaderHash b ⇒ Point a → Point b

blockPoint ∷ HasHeader block ⇒ block → Point block

AnchoredFragment construction and inspection

Head inspection

headPoint ∷ HasHeader block ⇒ AnchoredFragment block → Point block

\( O(1) \). When the fragment is empty, the anchor point is returned.

headAnchor ∷ HasHeader block ⇒ AnchoredFragment block → Anchor block

\( O(1) \). The anchor corresponding to the most recently added block (i.e., the anchor that would be needed for a fragment starting after this)

headSlot ∷ HasHeader block ⇒ AnchoredFragment block → WithOrigin SlotNo

\( O(1) \). When the fragment is empty, the slot of the anchor point is returned, which may be origin (no slot).

headHash ∷ HasHeader block ⇒ AnchoredFragment block → ChainHash block

\( O(1) \). When the fragment is empty, the hash of the anchor point is returned.

headBlockNo ∷ HasHeader block ⇒ AnchoredFragment block → WithOrigin BlockNo

\( O(1) \). When the fragment is empty, the block number of the anchor point is returned.

Basic operations

head ∷ HasHeader block ⇒ AnchoredFragment block → Either (Anchor block) block

\( O(1) \). When the fragment is empty, return the anchor point, otherwise the most recently added block.

last ∷ HasHeader block ⇒ AnchoredFragment block → Either (Anchor block) block

\( O(1) \). When the fragment is empty, return the anchor point, otherwise the leftmost block.

lastPoint ∷ HasHeader block ⇒ AnchoredFragment block → Point block

\( O(1) \). When the fragment is empty, the anchor point is returned.

lastSlot ∷ HasHeader block ⇒ AnchoredFragment block → WithOrigin SlotNo

\( O(1) \). When the fragment is empty, the slot of the anchor point is returned, which may be the origin and therefore have no slot.

toNewestFirst ∷ HasHeader block ⇒ AnchoredFragment block → [block]

TODO. Make a list of blocks from a AnchoredFragment, in newest-to-oldest order.

toOldestFirst ∷ HasHeader block ⇒ AnchoredFragment block → [block]

\( O(n) \). Make a list of blocks from a AnchoredFragment, in oldest-to-newest order.

fromNewestFirst

Arguments

∷ HasHeader block

⇒ Anchor block
Anchor

→ [block]

→ AnchoredFragment block

\( O(n) \). Make a AnchoredFragment from a list of blocks in newest-to-oldest order. The last block in the list must be the block following the given anchor point.

fromOldestFirst

Arguments

∷ HasHeader block

⇒ Anchor block
Anchor

→ [block]

→ AnchoredFragment block

\( O(n) \). Make a AnchoredFragment from a list of blocks in oldest-to-newest order. The first block in the list must be the block following the given anchor point.

dropNewest

Arguments

∷ HasHeader block

⇒ Int
n

→ AnchoredFragment block

→ AnchoredFragment block

\( O(\log(\min(i,n-i)) \). Drop the newest n blocks from the AnchoredFragment. The anchor point is not changed.

takeOldest

Arguments

∷ HasHeader block

⇒ Int
n

→ AnchoredFragment block

→ AnchoredFragment block

\( O(\log(\min(i,n-i)) \). Take the oldest n blocks from the AnchoredFragment. The anchor point is not changed.

dropWhileNewest ∷ HasHeader block ⇒ (block → Bool) → AnchoredFragment block → AnchoredFragment block

\( O(n) \). Drop the newest blocks that satisfy the predicate, keeping the remainder. The anchor point is not changed.

takeWhileOldest ∷ HasHeader block ⇒ (block → Bool) → AnchoredFragment block → AnchoredFragment block

\( O(n) \). Take the oldest blocks that satisfy the predicate. The anchor point is not changed.

length ∷ HasHeader block ⇒ AnchoredFragment block → Int

\( O(1) \). Return the number of blocks. The anchor point is not counted.

null ∷ AnchoredFragment block → Bool

\( O(1) \). The anchor point is not counted.

Update type and operations

data ChainUpdate block a

A representation of two actions to update a chain: add a block or roll back to a previous point.

The type parameter a is there to allow a Functor instance. Typically, it will be instantiated with block itself.

Constructors

AddBlock a

RollBack (Point block)

addBlock ∷ HasHeader block ⇒ block → AnchoredFragment block → AnchoredFragment block

\( O(1) \). Add a block to the right of the anchored fragment.

Synonym for :>.

rollback ∷ HasHeader block ⇒ Point block → AnchoredFragment block → Maybe (AnchoredFragment block)

\( O(\log(\min(i,n-i)) \). If the Point is within the bounds of the AnchoredFragment (see withinFragmentBounds), roll back the anchored fragment such that its head is the given point. In case the given point was the anchor point, the returned anchored fragment will be empty.

In other words, remove blocks from the end of the AnchoredFragment until the given Point is the head. If the given Point is not within the bounds of the AnchoredFragment, return Nothing.

applyChainUpdate ∷ HasHeader block ⇒ ChainUpdate block block → AnchoredFragment block → Maybe (AnchoredFragment block)

applyChainUpdates ∷ HasHeader block ⇒ [ChainUpdate block block] → AnchoredFragment block → Maybe (AnchoredFragment block)

Special operations

pointOnFragment ∷ HasHeader block ⇒ Point block → AnchoredFragment block → Bool

\( O(\log(\min(i,n-i)) \). Same as pointOnChainFragment: does the fragment contain a block with the given block? The anchor point is ignored.

withinFragmentBounds ∷ HasHeader block ⇒ Point block → AnchoredFragment block → Bool

\( O(\log(\min(i,n-i)) \). Is the point within the fragment bounds? Either the point is the anchor point, or it corresponds to a block "on" the fragment.

findFirstPoint ∷ HasHeader block ⇒ [Point block] → AnchoredFragment block → Maybe (Point block)

\( O(p \log(\min(i,n-i)) \). Find the first Point in the list of points that is within the fragment bounds. Return Nothing if none of them are.

Note: in contrast to findFirstPoint, this is based on withinFragmentBounds instead of pointOnFragment.

successorBlock ∷ HasHeader block ⇒ Point block → AnchoredFragment block → Maybe block

\( O(\log(\min(i,n-i)) \). Find the block after the given point. If the given point is the anchor point, then the first block is returned (if there is one).

selectPoints ∷ HasHeader block ⇒ [Int] → AnchoredFragment block → [Point block]

\( O(o \log(\min(i,n-i))) \). See selectPoints.

The list of offsets must be increasing monotonically.

Note: offset n, where n equals the length of the anchored fragment, corresponds to the anchor point. When the fragment is empty, offset 0 will thus correspond to the anchor point.

isPrefixOf ∷ (HasHeader block, Eq block) ⇒ AnchoredFragment block → AnchoredFragment block → Bool

\( O(\max(n_1, n_2)) \). Check whether the first anchored fragment is a prefix of the second.

The two AnchoredFragments must have the same anchor point, otherwise the first cannot be a prefix of the second.

splitAfterPoint ∷ ∀ block1 block2. (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ AnchoredFragment block1 → Point block2 → Maybe (AnchoredFragment block1, AnchoredFragment block1)

\( O(\log(\min(i,n-i)) \). Split the AnchoredFragment after the given Point. Return Nothing if given Point is not within the fragment bounds (withinFragmentBounds).

The given Point may be the anchor point of the fragment, in which case the empty fragment with the given anchor point and the original fragment are returned.

POSTCONDITION: when Just (before, after) = splitAfterPoint f pt, then: * anchorPoint before == anchorPoint f * headPoint before == pt * anchorPoint after == pt * headPoint after == headPoint f * join before after == Just f

splitBeforePoint ∷ ∀ block1 block2. (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ AnchoredFragment block1 → Point block2 → Maybe (AnchoredFragment block1, AnchoredFragment block1)

\( O(\log(\min(i,n-i)) \). Split the AnchoredFragment before the given Point. Return Nothing if given Point is not on the fragment (pointOnFragment).

This means that Nothing is returned if the given Point is the anchor point of the fragment.

POSTCONDITION: joining (join) the two fragments gives back the original fragment.

POSTCONDITION: the last block (oldest) on the second fragment corresponds to the given point.

join ∷ HasHeader block ⇒ AnchoredFragment block → AnchoredFragment block → Maybe (AnchoredFragment block)

\( O(\log(\min(n_1, n_2))) \). Join two anchored fragments if the anchor of the second fragment is the head (newest block) of the first fragment.

If the first fragment is empty, it can be joined if its anchor is the same as the second fragment's anchor.

The returned fragment will have the same anchor as the first fragment.

intersect ∷ ∀ block1 block2. (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ AnchoredFragment block1 → AnchoredFragment block2 → Maybe (AnchoredFragment block1, AnchoredFragment block2, AnchoredFragment block1, AnchoredFragment block2)

\( O(n_2 \log(n_1)) \). Look for the most recent intersection of two AnchoredFragments c1 and c2.

The fragments need not have the same anchor point.

If they intersect, i.e., share a common Point (possibly the anchor point), then return a tuple of:

p1: the prefix of the first fragment

p2: the prefix of the second fragment

s1: the suffix of the first fragment

s2: the suffix of the second fragment

p1 and p2 will have the same head (possibly an anchor point), namely the intersection point i. The original chain c1 can be obtained by putting s1 after p1, similarly for c2: by putting s2 after p2:

Just c1 = join p1 s1
Just c2 = join p2 s2

Take for example the following two fragments that share blocks 4 and 5. The two fragments are fragments of the same chain, but don't contain all blocks of the original chain. The anchor points of the fragments are indicated with an asterisk (*). The -A and -B suffixes denote that blocks are part of a fork of the chain.

    ┆ 1*┆
    ├───┤
    │ 2 │     ┆ 2*┆
    ├───┤     ├───┤
    │ 4 │     │ 4 │
    ├───┤     ├───┤
    │ 5 │     │ 5 │
────┼───┼─────┼───┼───
    │ 6A│     │ 6B│
    └───┘     ├───┤
              │ 8B│
              └───┘
      c1        c2

The intersection of c1 and c2 is block 5 (the last Point the two fragments have in common) and we return the following fragments:

    ┆ 1*┆
    ├───┤
    │ 2 │     ┆ 2*┆
    ├───┤     ├───┤
    │ 4 │     │ 4 │
    ├───┤     ├───┤
    │ 5 │     │ 5 │      ┆ 5*┆     ┆ 5*┆
────┴───┴─────┴───┴──────┼───┼─────┼───┼──
                         │ 6A│     │ 6B│
                         └───┘     ├───┤
                                   │ 8B│
                                   └───┘
Just (p1,       p2,        s1,       s2)

The intersection point will be the anchor point of fragments s1 and s2. Fragment p1 will have the same anchor as c1 and p2 will have the same anchor as c2.

Note that an empty fragment can still intersect another fragment, as its anchor point can still intersect the other fragment. In that case the respective prefix and suffix are both equal to original empty fragment. Additionally, two empty fragments intersect if their anchor points are equal, in which case all prefixes and suffixes are equal to the empty fragment with the anchor point in question.

intersectionPoint ∷ (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ AnchoredFragment block1 → AnchoredFragment block2 → Maybe (Point block1)

\( O(n_2 \log(n_1)) \). Look for the most recent intersection point of two AnchoredFragments

The fragments need not have the same anchor point.

Reusing the example in the docstring of intersect: this function will return the anchor point 5*.

mapAnchoredFragment ∷ (HasHeader block1, HasHeader block2, HeaderHash block1 ~ HeaderHash block2) ⇒ (block1 → block2) → AnchoredFragment block1 → AnchoredFragment block2

\( O(n) \). Maps over the chain blocks. This is not allowed to change the block Points, or it would create an invalid chain. The anchorPoint is not affected.

anchorNewest

Arguments

∷ HasHeader block

⇒ Word64
n

→ AnchoredFragment block

→ AnchoredFragment block

Take the n newest blocks from the fragment.

WARNING: this may change the anchor of the fragment!

When the fragment itself is shorter than n blocks, the fragment will be returned unmodified.

Helper functions

prettyPrint ∷ HasHeader block ⇒ String → (Point block → String) → (block → String) → AnchoredFragment block → String