Ouroboros.Network.BlockFetch

Description

Let's start with the big picture...

Key:  ┏━━━━━━━━━━━━┓  ╔═════════════╗  ┏━━━━━━━━━━━━━━┓   ╔════════════╗
      ┃ STM-based  ┃  ║active thread║  ┃state instance┃┓  ║ one thread ║╗
      ┃shared state┃  ║             ║  ┃   per peer   ┃┃  ║  per peer  ║║
      ┗━━━━━━━━━━━━┛  ╚═════════════╝  ┗━━━━━━━━━━━━━━┛┃  ╚════════════╝║
                                        ┗━━━━━━━━━━━━━━┛   ╚════════════╝

  ╔═════════════╗     ┏━━━━━━━━━━━━━┓
  ║ Chain sync  ║╗    ┃   Ledger    ┃
  ║  protocol   ║║◀───┨   state     ┃◀───────────╮
  ║(client side)║║    ┃             ┃            │
  ╚══════╤══════╝║    ┗━━━━━━━━━━━━━┛            │
   ╚═════╪═══════╝                               │
         ▼                                       │
  ┏━━━━━━━━━━━━━┓     ┏━━━━━━━━━━━━━┓     ╔══════╧══════╗
  ┃  Candidate  ┃     ┃   Set of    ┃     ║  Chain and  ║
  ┃  chains     ┃     ┃  downloaded ┠────▶║   ledger    ║
  ┃  (headers)  ┃     ┃   blocks    ┃     ║  validation ║
  ┗━━━━━┯━━━━━━━┛     ┗━━━━━┯━━━━━━━┛     ╚══════╤══════╝
        │                   │ ▲                  │
        │ ╭─────────────────╯ │                  │
░░░░░░░░▼░▼░░░░░░░░           │                  ▼
░░╔═════════════╗░░           │           ┏━━━━━━━━━━━━━┓     ╔═════════════╗
░░║    Block    ║░░           │           ┃   Current   ┃     ║ Block fetch ║╗
░░╢    fetch    ║◀────────────┼───────────┨    chain    ┠────▶║ protocol    ║║
░░║    logic    ║░░           │           ┃  (blocks)   ┃     ║(server side)║║
░░╚═════════════╝░░           │           ┠─────────────┨     ╚═════════════╝║
░░░░░░░░░▲░░░░░░░░░           │           ┃  Tentative  ┃      ╚═════════════╝
░░░░░░░░░▼░░░░░░░░░░░░░░░░░░░░│░░░░░░░░   ┃    chain    ┠──╮
░░┏━━━━━━━━━━━━━┓░░░░░╔═══════╧═════╗░░   ┃  (headers)  ┃  │  ╔═════════════╗
░░┃ Block fetch ┃┓░░░░║ block fetch ║╗░   ┗━━━━━━━━━━━━━┛  │  ║ Chain sync  ║╗
░░┃  state and  ┃┃◀──▶║  protocol   ║║░                    ╰─▶║ protocol    ║║
░░┃  requests   ┃┃░░░░║(client side)║║░                       ║(server side)║║
░░┗━━━━━━━━━━━━━┛┃░░░░╚═════════════╝║░                       ╚═════════════╝║
░░░┗━━━━━━━━━━━━━┛░░░░░╚═════════════╝░                        ╚═════════════╝
░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░

Notes:

Thread communication is via STM based state.

Outbound: threads update STM state.

Inbound: threads wait on STM state changing (using retry).

These are no queues: there is only the current state, not all change events.

We consider the block fetch logic and the policy for the block fetch protocol client together as one unit of functionality. This is the shaded area in the diagram.

Looking at the diagram we see that these two threads interact with each other and other threads via the following shared state

State 	 Interactions 	 Internal/External

Candidate chains (headers) 	 Read 	 External

Current chain (blocks) 	 Read 	 External

Set of downloaded blocks 	 Read & Write 	 External

Block fetch requests 	 Read & Write 	 Internal

The block fetch requests state is private between the block fetch logic and the block fetch protocol client, so it is implemented here.

The other state is managed by the consensus layer and is considered external here. So here we define interfaces for interacting with the external state. These have to be provided when instantiating the block fetch logic.

Documentation

blockFetchLogic ∷ ∀ peer header block m. (MonadSTM m, Ord peer, HasHeader header, HasHeader block, HeaderHash header ~ HeaderHash block) ⇒ Tracer m [TraceLabelPeer peer (FetchDecision [Point header])] → Tracer m (TraceLabelPeer peer (TraceFetchClientState header)) → BlockFetchConsensusInterface peer header block m → FetchClientRegistry peer header block m → BlockFetchConfiguration → m Void

Execute the block fetch logic. It monitors the current chain and candidate chains. It decided which block bodies to fetch and manages the process of fetching them, including making alternative decisions based on timeouts and failures.

This runs forever and should be shut down using mechanisms such as async.

data BlockFetchConfiguration

Configuration for FetchDecisionPolicy. Should be determined by external local node config.

Constructors

BlockFetchConfiguration

Fields

bfcMaxConcurrencyBulkSync ∷ !Word

Maximum concurrent downloads during bulk syncing.

bfcMaxConcurrencyDeadline ∷ !Word

Maximum concurrent downloads during deadline syncing.

bfcMaxRequestsInflight ∷ !Word

Maximum requests in flight per each peer.

data BlockFetchConsensusInterface peer header block m

The consensus layer functionality that the block fetch logic requires.

These are provided as input to the block fetch by the consensus layer.

Constructors

BlockFetchConsensusInterface

Fields

readCandidateChains ∷ STM m (Map peer (AnchoredFragment header))

Read the K-suffixes of the candidate chains.

Assumptions: * They must be already validated. * They may contain fewer than K blocks. * Their anchor does not have to intersect with the current chain.

readCurrentChain ∷ STM m (AnchoredFragment header)

Read the K-suffix of the current chain.

This must contain info on the last K blocks (unless we're near the chain genesis of course).

readFetchMode ∷ STM m FetchMode

Read the current fetch mode that the block fetch logic should use.

The fetch mode is a dynamic part of the block fetch policy. In FetchModeBulkSync it follows a policy that optimises for expected bandwidth over latency to fetch any particular block, whereas in FetchModeDeadline it follows a policy optimises for the latency to fetch blocks, at the expense of wasting bandwidth.

This mode should be set so that when the node's current chain is near to "now" it uses the deadline mode, and when it is far away it uses the bulk sync mode.

readFetchedBlocks ∷ STM m (Point block → Bool)

Recent, only within last K

addFetchedBlock ∷ Point block → block → m ()

This and readFetchedBlocks are required to be linked. Upon successful completion of BlockFetchConsensusInterface it must be the case that readFetchedBlocks reports the block.

readFetchedMaxSlotNo ∷ STM m MaxSlotNo

The highest stored/downloaded slot number.

This is used to optimise the filtering of fragments in the block fetch logic: when removing already downloaded blocks from a fragment, the filtering (with a linear cost) is stopped as soon as a block has a slot number higher than this slot number, as it cannot have been downloaded anyway.

plausibleCandidateChain ∷ AnchoredFragment header → AnchoredFragment header → Bool

Given the current chain, is the given chain plausible as a candidate chain. Classically for Ouroboros this would simply check if the candidate is strictly longer, but for Ouroboros with operational key certificates there are also cases where we would consider a chain of equal length to the current chain.

compareCandidateChains ∷ AnchoredFragment header → AnchoredFragment header → Ordering

Compare two candidate chains and return a preference ordering. This is used as part of selecting which chains to prioritise for downloading block bodies.

blockFetchSize ∷ header → SizeInBytes

Much of the logic for deciding which blocks to download from which peer depends on making estimates based on recent performance metrics. These estimates of course depend on the amount of data we will be downloading.

blockMatchesHeader ∷ header → block → Bool

Given a block header, validate the supposed corresponding block body.

Tracer types

type FetchDecision result = Either FetchDecline result

Throughout the decision making process we accumulate reasons to decline to fetch any blocks. This type is used to wrap intermediate and final results.

data TraceFetchClientState header

Tracing types for the various events that change the state (i.e. FetchClientStateVars) for a block fetch client.

Note that while these are all state changes, the AddedFetchRequest occurs in the decision thread while the other state changes occur in the block fetch client threads.

Constructors

AddedFetchRequest (FetchRequest header) (PeerFetchInFlight header) PeerFetchInFlightLimits (PeerFetchStatus header)
The block fetch decision thread has added a new fetch instruction consisting of one or more individual request ranges.

AcknowledgedFetchRequest (FetchRequest header)
Mark the point when the fetch client picks up the request added by the block fetch decision thread. Note that this event can happen fewer times than the AddedFetchRequest due to fetch request merging.

StartedFetchBatch (ChainRange header) (PeerFetchInFlight header) PeerFetchInFlightLimits (PeerFetchStatus header)
Mark the start of receiving a streaming batch of blocks. This will be followed by one or more CompletedBlockFetch and a final CompletedFetchBatch.

CompletedBlockFetch (Point header) (PeerFetchInFlight header) PeerFetchInFlightLimits (PeerFetchStatus header)
Mark the completion of of receiving a single block within a streaming batch of blocks.

CompletedFetchBatch (ChainRange header) (PeerFetchInFlight header) PeerFetchInFlightLimits (PeerFetchStatus header)
Mark the successful end of receiving a streaming batch of blocks

RejectedFetchBatch (ChainRange header) (PeerFetchInFlight header) PeerFetchInFlightLimits (PeerFetchStatus header)
If the other peer rejects our request then we have this event instead of StartedFetchBatch and CompletedFetchBatch.

data TraceLabelPeer peerid a

A peer label for use in Tracers. This annotates tracer output as being associated with a given peer identifier.

Constructors

TraceLabelPeer peerid a

The FetchClientRegistry

data FetchClientRegistry peer header block m

A registry for the threads that are executing the client side of the BlockFetch protocol to communicate with our peers.

The registry contains the shared variables we use to communicate with these threads, both to track their status and to provide instructions.

The threads add/remove themselves to/from this registry when they start up and shut down.

newFetchClientRegistry ∷ MonadSTM m ⇒ m (FetchClientRegistry peer header block m)

bracketFetchClient ∷ ∀ m a peer header block. (MonadThrow m, MonadSTM m, MonadFork m, Ord peer) ⇒ FetchClientRegistry peer header block m → peer → (FetchClientContext header block m → m a) → m a

This is needed to start a block fetch client. It provides the required FetchClientContext. It registers and unregisters the fetch client on start and end.

It also manages synchronisation with the corresponding chain sync client.

bracketSyncWithFetchClient ∷ ∀ m a peer header block. (MonadThrow m, MonadSTM m, MonadFork m, Ord peer) ⇒ FetchClientRegistry peer header block m → peer → m a → m a

The block fetch and chain sync clients for each peer need to synchronise their startup and shutdown. This bracket operation provides that synchronisation for the chain sync client.

This must be used for the chain sync client outside of its own state registration and deregistration.

Re-export types used by BlockFetchConsensusInterface

data FetchMode

Constructors

FetchModeBulkSync
Use this mode when we are catching up on the chain but are stil well behind. In this mode the fetch logic will optimise for throughput rather than latency.

FetchModeDeadline
Use this mode for block-producing nodes that have a known deadline to produce a block and need to get the best chain before that. In this mode the fetch logic will optimise for picking the best chain within the given deadline.

type SizeInBytes = Word32