Ouroboros.Network.BlockFetch.Decision

Deciding what to fetch

fetchDecisions ∷ (HasHeader header, HasHeader block, HeaderHash header ~ HeaderHash block) ⇒ FetchDecisionPolicy header → FetchMode → AnchoredFragment header → (Point block → Bool) → MaxSlotNo → [(AnchoredFragment header, PeerInfo header extra)] → [(FetchDecision (FetchRequest header), PeerInfo header extra)]

data FetchDecisionPolicy header

Constructors

FetchDecisionPolicy

Fields

maxInFlightReqsPerPeer ∷ Word

maxConcurrencyBulkSync ∷ Word

maxConcurrencyDeadline ∷ Word

plausibleCandidateChain ∷ AnchoredFragment header → AnchoredFragment header → Bool

compareCandidateChains ∷ AnchoredFragment header → AnchoredFragment header → Ordering

blockFetchSize ∷ header → SizeInBytes

data FetchMode

Constructors

FetchModeBulkSync
Use this mode when we are catching up on the chain but are stil well behind. In this mode the fetch logic will optimise for throughput rather than latency.

FetchModeDeadline
Use this mode for block-producing nodes that have a known deadline to produce a block and need to get the best chain before that. In this mode the fetch logic will optimise for picking the best chain within the given deadline.

type PeerInfo header extra = (PeerFetchStatus header, PeerFetchInFlight header, PeerGSV, extra)

type FetchDecision result = Either FetchDecline result

Throughout the decision making process we accumulate reasons to decline to fetch any blocks. This type is used to wrap intermediate and final results.

data FetchDecline

All the various reasons we can decide not to fetch blocks from a peer.

Constructors

FetchDeclineChainNotPlausible

FetchDeclineChainNoIntersection

FetchDeclineAlreadyFetched

FetchDeclineInFlightThisPeer

FetchDeclineInFlightOtherPeer

FetchDeclinePeerShutdown

FetchDeclinePeerSlow

FetchDeclineReqsInFlightLimit !Word

FetchDeclineBytesInFlightLimit !SizeInBytes !SizeInBytes !SizeInBytes

FetchDeclinePeerBusy !SizeInBytes !SizeInBytes !SizeInBytes

FetchDeclineConcurrencyLimit !FetchMode !Word

Components of the decision-making process

filterPlausibleCandidates

Arguments

∷ HasHeader header

⇒ (AnchoredFragment block → AnchoredFragment header → Bool)

→ AnchoredFragment block
The current chain

→ [(AnchoredFragment header, peerinfo)]

→ [(FetchDecision (AnchoredFragment header), peerinfo)]

Keep only those candidate chains that are preferred over the current chain. Typically, this means that their length is longer than the length of the current chain.

selectForkSuffixes ∷ (HasHeader header, HasHeader block, HeaderHash header ~ HeaderHash block) ⇒ AnchoredFragment block → [(FetchDecision (AnchoredFragment header), peerinfo)] → [(FetchDecision (ChainSuffix header), peerinfo)]

filterNotAlreadyFetched ∷ (HasHeader header, HeaderHash header ~ HeaderHash block) ⇒ (Point block → Bool) → MaxSlotNo → [(FetchDecision (ChainSuffix header), peerinfo)] → [(FetchDecision (CandidateFragments header), peerinfo)]

Find the fragments of the chain suffix that we still need to fetch, these are the fragments covering blocks that have not yet been fetched and are not currently in the process of being fetched from this peer.

Typically this is a single fragment forming a suffix of the chain, but in the general case we can get a bunch of discontiguous chain fragments.

filterNotAlreadyInFlightWithPeer ∷ HasHeader header ⇒ [(FetchDecision (CandidateFragments header), PeerFetchInFlight header, peerinfo)] → [(FetchDecision (CandidateFragments header), peerinfo)]

prioritisePeerChains ∷ ∀ header peer. HasHeader header ⇒ FetchMode → (AnchoredFragment header → AnchoredFragment header → Ordering) → (header → SizeInBytes) → [(FetchDecision (CandidateFragments header), PeerFetchInFlight header, PeerGSV, peer)] → [(FetchDecision [ChainFragment header], peer)]

filterNotAlreadyInFlightWithOtherPeers ∷ HasHeader header ⇒ FetchMode → [(FetchDecision [ChainFragment header], PeerFetchStatus header, PeerFetchInFlight header, peerinfo)] → [(FetchDecision [ChainFragment header], peerinfo)]

A penultimate step of filtering, but this time across peers, rather than individually for each peer. If we're following the parallel fetch mode then we filter out blocks that are already in-flight with other peers.

Note that this does not cover blocks that are proposed to be fetched in this round of decisions. That step is covered in fetchRequestDecisions.

fetchRequestDecisions ∷ ∀ header peer. HasHeader header ⇒ FetchDecisionPolicy header → FetchMode → [(FetchDecision [ChainFragment header], PeerFetchStatus header, PeerFetchInFlight header, PeerGSV, peer)] → [(FetchDecision (FetchRequest header), peer)]