Ouroboros.Network.Diffusion

Documentation

data DiffusionTracers

Constructors

DiffusionTracers

Fields

dtIpSubscriptionTracer ∷ Tracer IO (WithIPList (SubscriptionTrace SockAddr))

IP subscription tracer

dtDnsSubscriptionTracer ∷ Tracer IO (WithDomainName (SubscriptionTrace SockAddr))

DNS subscription tracer

dtDnsResolverTracer ∷ Tracer IO (WithDomainName DnsTrace)

DNS resolver tracer

dtMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId SockAddr) MuxTrace)

Mux tracer

dtMuxLocalTracer ∷ Tracer IO (WithMuxBearer (ConnectionId LocalAddress) MuxTrace)

Mux tracer for local clients

dtHandshakeTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer

dtHandshakeLocalTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer for local clients

dtErrorPolicyTracer ∷ Tracer IO (WithAddr SockAddr ErrorPolicyTrace)

dtLocalErrorPolicyTracer ∷ Tracer IO (WithAddr LocalAddress ErrorPolicyTrace)

dtAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

Trace rate limiting of accepted connections

dtDiffusionInitializationTracer ∷ Tracer IO DiffusionInitializationTracer

dtLedgerPeersTracer ∷ Tracer IO TraceLedgerPeers

data DiffusionArguments

Network Node argumets

Constructors

DiffusionArguments

Fields

daIPv4Address ∷ Maybe (Either Socket AddrInfo)

IPv4 socket ready to accept connections or diffusion addresses

daIPv6Address ∷ Maybe (Either Socket AddrInfo)

IPV4 socket ready to accept connections or diffusion addresses

daLocalAddress ∷ Maybe (Either Socket FilePath)

AF_UNIX socket ready to accept connections or address for local clients

daIpProducers ∷ IPSubscriptionTarget

ip subscription addresses

daDnsProducers ∷ [DnsSubscriptionTarget]

list of domain names to subscribe to

daAcceptedConnectionsLimit ∷ AcceptedConnectionsLimit

parameters for limiting number of accepted connections

daDiffusionMode ∷ DiffusionMode

run in initiator only mode

data AcceptedConnectionsLimit

Policy which governs how to limit the number of accepted connections.

Constructors

AcceptedConnectionsLimit

Fields

acceptedConnectionsHardLimit ∷ !Word32

Hard limit of accepted connections.

acceptedConnectionsSoftLimit ∷ !Word32

Soft limit of accepted connections. If we are above this threshold, we will start rate limiting.

acceptedConnectionsDelay ∷ !DiffTime

Max delay for limiting accepted connections. We use linear regression starting from 0 at the soft limit up to acceptedConnectionDelay at the hard limit.

data DiffusionApplications ntnAddr ntcAddr ntnVersionData ntcVersionData m

Constructors

DiffusionApplications

Fields

daResponderApplication ∷ Versions NodeToNodeVersion ntnVersionData (OuroborosApplication ResponderMode ntnAddr ByteString m Void ())

NodeToNode reposnder application (server role)

daInitiatorApplication ∷ Versions NodeToNodeVersion ntnVersionData (OuroborosApplication InitiatorMode ntnAddr ByteString m () Void)

NodeToNode initiator application (client role)

daLocalResponderApplication ∷ Versions NodeToClientVersion ntcVersionData (OuroborosApplication ResponderMode ntcAddr ByteString m Void ())

NodeToClient responder applicaton (server role)

daErrorPolicies ∷ ErrorPolicies

error policies

daLedgerPeersCtx ∷ LedgerPeersConsensusInterface m

Interface used to get peers from the current ledger.

newtype OuroborosApplication (mode ∷ MuxMode) addr bytes (m ∷ Type → Type) a b

Like MuxApplication but using a MuxPeer rather than a raw Channel -> m a action.

Constructors

OuroborosApplication (ConnectionId addr → ControlMessageSTM m → [MiniProtocol mode bytes m a b])

runDataDiffusion ∷ DiffusionTracers → DiffusionArguments → DiffusionApplications RemoteAddress LocalAddress NodeToNodeVersionData NodeToClientVersionData IO → IO ()

re-exports

simpleSingletonVersions ∷ vNum → vData → r → Versions vNum vData r

Singleton smart constructor for Versions.

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps ∷ ![SockAddr]

List of destinations to possibly connect to

ispValency ∷ !Int

Number of parallel connections to keep actice.

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain ∷ !Domain

dstPort ∷ !PortNumber

dstValency ∷ !Int

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr

data DiffusionInitializationTracer

Constructors

RunServer !SockAddr

RunLocalServer !LocalAddress

UsingSystemdSocket !FilePath

CreateSystemdSocketForSnocketPath !FilePath

CreatedLocalSocket !FilePath

ConfiguringLocalSocket !FilePath !FileDescriptor

ListeningLocalSocket !FilePath !FileDescriptor

LocalSocketUp !FilePath !FileDescriptor

CreatingServerSocket !SockAddr

ConfiguringServerSocket !SockAddr

ListeningServerSocket !SockAddr

ServerSocketUp !SockAddr

UnsupportedLocalSystemdSocket !SockAddr

UnsupportedReadySocketCase

DiffusionErrored SomeException

newtype LedgerPeersConsensusInterface m

Constructors

LedgerPeersConsensusInterface

Fields

lpGetPeers ∷ SlotNo → STM m (Maybe [(PoolStake, NonEmpty RelayAccessPoint)])