Ouroboros.Network.Diffusion

Common API

data P2P

Promoted data types.

Constructors

P2P

NonP2P

data InitializationTracer ntnAddr ntcAddr

Constructors

RunServer (NonEmpty ntnAddr)

RunLocalServer ntcAddr

UsingSystemdSocket ntcAddr

CreateSystemdSocketForSnocketPath ntcAddr

CreatedLocalSocket ntcAddr

ConfiguringLocalSocket ntcAddr FileDescriptor

ListeningLocalSocket ntcAddr FileDescriptor

LocalSocketUp ntcAddr FileDescriptor

CreatingServerSocket ntnAddr

ConfiguringServerSocket ntnAddr

ListeningServerSocket ntnAddr

ServerSocketUp ntnAddr

UnsupportedLocalSystemdSocket ntnAddr

UnsupportedReadySocketCase

DiffusionErrored SomeException

data Tracers ntnAddr ntnVersion ntcAddr ntcVersion m

Common DiffusionTracers interface between P2P and NonP2P

Constructors

Tracers

Fields

dtMuxTracer ∷ Tracer m (WithMuxBearer (ConnectionId ntnAddr) MuxTrace)

Mux tracer

dtHandshakeTracer ∷ Tracer m (HandshakeTr ntnAddr ntnVersion)

Handshake protocol tracer

dtLocalMuxTracer ∷ Tracer m (WithMuxBearer (ConnectionId ntcAddr) MuxTrace)

Mux tracer for local clients

dtLocalHandshakeTracer ∷ Tracer m (HandshakeTr ntcAddr ntcVersion)

Handshake protocol tracer for local clients

dtDiffusionInitializationTracer ∷ Tracer m (InitializationTracer ntnAddr ntcAddr)

Diffusion initialisation tracer

dtLedgerPeersTracer ∷ Tracer m TraceLedgerPeers

Ledger Peers tracer

nullTracers ∷ Applicative m ⇒ Tracers ntnAddr ntnVersion ntcAddr ntcVersion m

data ExtraTracers (p2p ∷ P2P) where

Tracers which depend on p2p mode.

Constructors

P2PTracers ∷ TracersExtra RemoteAddress NodeToNodeVersion NodeToNodeVersionData LocalAddress NodeToClientVersion NodeToClientVersionData IOException IO → ExtraTracers 'P2P

NonP2PTracers ∷ TracersExtra → ExtraTracers 'NonP2P

data Failure ntnAddr

data Arguments ntnFd ntnAddr ntcFd ntcAddr

Common DiffusionArguments interface between P2P and NonP2P

Constructors

Arguments

Fields

daIPv4Address ∷ Maybe (Either ntnFd ntnAddr)

an IPv4 socket ready to accept connections or an IPv4 addresses

daIPv6Address ∷ Maybe (Either ntnFd ntnAddr)

an IPV4 socket ready to accept connections or an IPv6 addresses

daLocalAddress ∷ Maybe (Either ntcFd ntcAddr)

an AF_UNIX socket ready to accept connections or an AF_UNIX socket path

daAcceptedConnectionsLimit ∷ AcceptedConnectionsLimit

parameters for limiting number of accepted connections

daMode ∷ DiffusionMode

run in initiator only mode

data ExtraArguments (p2p ∷ P2P) m where

Diffusion arguments which depend on p2p mode.

Constructors

P2PArguments ∷ ArgumentsExtra m → ExtraArguments 'P2P m

NonP2PArguments ∷ ArgumentsExtra → ExtraArguments 'NonP2P m

data Applications ntnAddr ntnVersion ntnVersionData ntcAddr ntcVersion ntcVersionData m

Common DiffusionArguments interface between P2P and NonP2P

TODO: we need initiator only mode for blockchain explorer or a similar application, there's no reason why one should run a node-to-node server for it.

Constructors

Applications

Fields

daApplicationInitiatorMode ∷ Versions ntnVersion ntnVersionData (OuroborosBundle InitiatorMode ntnAddr ByteString m () Void)

NodeToNode initiator applications for initiator only mode.

TODO: we should accept one or the other, but not both: daApplicationInitiatorMode, daApplicationInitiatorResponderMode.

daApplicationInitiatorResponderMode ∷ Versions ntnVersion ntnVersionData (OuroborosBundle InitiatorResponderMode ntnAddr ByteString m () ())

NodeToNode initiator & responder applications for bidirectional mode.

daLocalResponderApplication ∷ Versions ntcVersion ntcVersionData (OuroborosApplication ResponderMode ntcAddr ByteString m Void ())

NodeToClient responder application (server role)

daLedgerPeersCtx ∷ LedgerPeersConsensusInterface m

Interface used to get peers from the current ledger.

TODO: it should be in InterfaceExtra

data ExtraApplications (p2p ∷ P2P) ntnAddr m where

Application data which depend on p2p mode.

Constructors

P2PApplications ∷ ApplicationsExtra ntnAddr m → ExtraApplications 'P2P ntnAddr m

NonP2PApplications ∷ ApplicationsExtra → ExtraApplications 'NonP2P ntnAddr m

Run data diffusion

run ∷ ∀ (p2p ∷ P2P). Tracers RemoteAddress NodeToNodeVersion LocalAddress NodeToClientVersion IO → ExtraTracers p2p → Arguments Socket RemoteAddress LocalSocket LocalAddress → ExtraArguments p2p IO → Applications RemoteAddress NodeToNodeVersion NodeToNodeVersionData LocalAddress NodeToClientVersion NodeToClientVersionData IO → ExtraApplications p2p RemoteAddress IO → IO ()

Run data diffusion in either P2P or NonP2P mode.

Re-exports

type AbstractTransitionTrace peerAddr = TransitionTrace' peerAddr AbstractState