Ouroboros.Network.Diffusion

Documentation

data DiffusionTracers

Constructors

DiffusionTracers

Fields

dtIpSubscriptionTracer ∷ Tracer IO (WithIPList (SubscriptionTrace SockAddr))

IP subscription tracer

dtDnsSubscriptionTracer ∷ Tracer IO (WithDomainName (SubscriptionTrace SockAddr))

DNS subscription tracer

dtDnsResolverTracer ∷ Tracer IO (WithDomainName DnsTrace)

DNS resolver tracer

dtMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId SockAddr) MuxTrace)

Mux tracer

dtMuxLocalTracer ∷ Tracer IO (WithMuxBearer (ConnectionId LocalAddress) MuxTrace)

Mux tracer for local clients

dtHandshakeTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer

dtHandshakeLocalTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer for local clients

dtErrorPolicyTracer ∷ Tracer IO (WithAddr SockAddr ErrorPolicyTrace)

dtLocalErrorPolicyTracer ∷ Tracer IO (WithAddr LocalAddress ErrorPolicyTrace)

dtAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

Trace rate limiting of accepted connections

data DiffusionArguments

Network Node argumets

Constructors

DiffusionArguments

Fields

daAddresses ∷ [AddrInfo]

diffusion addresses

daLocalAddress ∷ FilePath

address for local clients

daIpProducers ∷ IPSubscriptionTarget

ip subscription addresses

daDnsProducers ∷ [DnsSubscriptionTarget]

list of domain names to subscribe to

daAcceptedConnectionsLimit ∷ AcceptedConnectionsLimit

parameters for limiting number of accepted connections

data AcceptedConnectionsLimit

Policy which governs how to limit the number of accepted connections.

Constructors

AcceptedConnectionsLimit

Fields

acceptedConnectionsHardLimit ∷ !Word32

Hard limit of accepted connections.

acceptedConnectionsSoftLimit ∷ !Word32

Soft limit of accepted connections. If we are above this threshold, we will start rate limiting.

acceptedConnectionsDelay ∷ !DiffTime

Max delay for limiting accepted connections. We use linear regression starting from 0 at the soft limit up to acceptedConnectionDelay at the hard limit.

data DiffusionApplications

Constructors

DiffusionApplications

Fields

daResponderApplication ∷ Versions NodeToNodeVersion DictVersion (ConnectionId SockAddr → OuroborosApplication ResponderApp ByteString IO Void ())

NodeToNode reposnder application (server role)

daInitiatorApplication ∷ Versions NodeToNodeVersion DictVersion (ConnectionId SockAddr → OuroborosApplication InitiatorApp ByteString IO () Void)

NodeToNode initiator application (client role)

daLocalResponderApplication ∷ Versions NodeToClientVersion DictVersion (ConnectionId LocalAddress → OuroborosApplication ResponderApp ByteString IO Void ())

NodeToClient responder applicaton (server role)

daErrorPolicies ∷ ErrorPolicies

error policies

newtype OuroborosApplication (appType ∷ AppType) bytes (m ∷ Type → Type) a b

Like MuxApplication but using a MuxPeer rather than a raw Channel -> m a action.

Constructors

OuroborosApplication [MiniProtocol appType bytes m a b]

runDataDiffusion ∷ DiffusionTracers → DiffusionArguments → DiffusionApplications → IO ()

re-exports

simpleSingletonVersions ∷ vNum → vData → extra vData → r → Versions vNum extra r

Singleton smart constructor for Versions.

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps ∷ ![SockAddr]

List of destinations to possibly connect to

ispValency ∷ !Int

Number of parallel connections to keep actice.

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain ∷ !Domain

dstPort ∷ !PortNumber

dstValency ∷ !Int

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr