Ouroboros.Network.MockChain.ProducerState

Documentation

data ChainProducerState block

Constructors

ChainProducerState

Fields

chainState ∷ Chain block

chainReaders ∷ ReaderStates block

nextReaderId ∷ ReaderId

type ReaderStates block = Map ReaderId (ReaderState block)

Readers are represented here as a relation.

type ReaderId = Int

data ReaderState block

Producer keeps track of consumer chain. The only information for a producer to know is * readerPoint: (some) intersection point of consumer's chain and producer's chain; * readerNext: information what to do on next instruction: either roll forward from the intersection point or roll back to it.

The second piece of information is needed to distinguish the following two cases:

consumer chain is a subchain of the producer chain

it is a fork.

Since consumer is following the producer chain, the producer has this information at its end. If producer updates its chain to use another fork it may happen that the reader pointer is not on the new chain. In this case the producer will set RollBackTo and find intersection of the two chains for readerPoint. And upon consumer's request will replay with MsgRollBackward readerPoint. After sending this message, the producer assumes that the the consumer is following the protocol (i.e. will rollback its chain) and will reset the readerNext field to ReaderForwardFrom. The second case: when the readerNext is ReaderForwardFrom, then when sending next instruction the producer will either:

take the next block (or header) on its chain imediatelly folowing the readerPoint, updtate readerPoint to the point of the new value and send MsgRollForward with the new block (or header).

if there is no block, which means that the consumer side and producer side are synchornized, the producer will send MsgAwaitResponse and will wait until its chain is updated: either by a fork or by a new block.

In this implementation a map from ReaderId to ReaderState is shared between all producers running on a single node; hence the unique identifier ReaderId for each reader: this is an implementation detail.

Constructors

ReaderState

Fields

readerPoint ∷ Point block

Where the chain of the consumer and producer intersect. If the consumer is on the chain then this is the consumer's chain head, but if the consumer's chain is off the producer's chain then this is the point the consumer will need to rollback to.

readerNext ∷ ReaderNext

Where the will go next, roll back to the reader point, or roll forward from the reader point.

data ReaderNext

Constructors

ReaderBackTo

ReaderForwardFrom

invChainProducerState ∷ HasHeader block ⇒ ChainProducerState block → Bool

invReaderStates ∷ HasHeader block ⇒ Chain block → ReaderStates block → Bool

initChainProducerState ∷ Chain block → ChainProducerState block

Initialise ChainProducerState with a given Chain and empty list of readers.

lookupReader ∷ ChainProducerState block → ReaderId → ReaderState block

Get the recorded state of a chain consumer. The ReaderId is assumed to exist.

readerExists ∷ ReaderId → ChainProducerState block → Bool

Return True when a reader with the given ReaderId exists.

producerChain ∷ ChainProducerState block → Chain block

Extract Chain from ChainProducerState.

findFirstPoint ∷ HasHeader block ⇒ [Point block] → ChainProducerState block → Maybe (Point block)

initReader ∷ HasHeader block ⇒ Point block → ChainProducerState block → (ChainProducerState block, ReaderId)

Add a new reader with the given intersection point and return the new ReaderId.

deleteReader ∷ ReaderId → ChainProducerState block → ChainProducerState block

Delete an existing reader. The ReaderId is assumed to exist.

updateReader

Arguments

∷ HasHeader block

⇒ ReaderId

→ Point block
new reader intersection point

→ ChainProducerState block

→ ChainProducerState block

Change the intersection point of a reader. This also puts it into the ReaderBackTo state.

switchFork ∷ HasHeader block ⇒ Chain block → ChainProducerState block → ChainProducerState block

Switch chains and update readers; if a reader point falls out of the chain, replace it with the intersection of both chains and put it in the ReaderBackTo state, otherwise preserve reader state.

readerInstruction ∷ HasHeader block ⇒ ReaderId → ChainProducerState block → Maybe (ChainUpdate block block, ChainProducerState block)

What a reader needs to do next. Should they move on to the next block or do they need to roll back to a previous point on their chain. It also updates the producer's state assuming that the reader follows its instruction.

addBlock ∷ HasHeader block ⇒ block → ChainProducerState block → ChainProducerState block

Add a block to the chain. It does not require any reader's state changes.

rollback ∷ (HasHeader block, HeaderHash block ~ HeaderHash block') ⇒ Point block' → ChainProducerState block → Maybe (ChainProducerState block)

Rollback producer chain. It requires to update reader states, since some readerPoints may not be on the new chain; in this case find intersection of the two chains and set readerNext to ReaderBackTo.

applyChainUpdate ∷ (HasHeader block, HeaderHash block ~ HeaderHash block') ⇒ ChainUpdate block' block → ChainProducerState block → Maybe (ChainProducerState block)

Convenient function which combines both addBlock and rollback.

applyChainUpdates ∷ (HasHeader block, HeaderHash block ~ HeaderHash block') ⇒ [ChainUpdate block' block] → ChainProducerState block → Maybe (ChainProducerState block)

Apply a list of ChainUpdates.