Ouroboros.Network.NodeToClient

Description

This is the starting point for a module that will bring together the overall node to client protocol, as a collection of mini-protocols.

Documentation

nodeToClientProtocols ∷ (ConnectionId addr → STM m ControlMessage → NodeToClientProtocols appType bytes m a b) → NodeToClientVersion → OuroborosApplication appType addr bytes m a b

Make an OuroborosApplication for the bundle of mini-protocols that make up the overall node-to-client protocol.

This function specifies the wire format protocol numbers as well as the protocols that run for each NodeToClientVersion.

They are chosen to not overlap with the node to node protocol numbers. This is not essential for correctness, but is helpful to allow a single shared implementation of tools that can analyse both protocols, e.g. wireshark plugins.

data NodeToClientProtocols appType bytes m a b

Recorod of node-to-client mini protocols.

Constructors

NodeToClientProtocols

Fields

localChainSyncProtocol ∷ RunMiniProtocol appType bytes m a b

local chain-sync mini-protocol

localTxSubmissionProtocol ∷ RunMiniProtocol appType bytes m a b

local tx-submission mini-protocol

localStateQueryProtocol ∷ RunMiniProtocol appType bytes m a b

local state-query mini-protocol

localTxMonitorProtocol ∷ RunMiniProtocol appType bytes m a b

local tx-monitor mini-protocol

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_1

NodeToClientV_2
added local-query mini-protocol

NodeToClientV_3
enabled CardanoNodeToClientVersion2

NodeToClientV_4
enabled CardanoNodeToClientVersion3, adding more queries

NodeToClientV_5
enabled CardanoNodeToClientVersion4, i.e., Allegra

NodeToClientV_6
enabled CardanoNodeToClientVersion5, i.e., Mary

NodeToClientV_7
enabled CardanoNodeToClientVersion6, adding a query

NodeToClientV_8
LocalStateQuery protocol codec change, allows to acquire tip point.

NodeToClientV_9
enabled CardanoNodeToClientVersion7, i.e., Alonzo

NodeToClientV_10
added GetChainBlockNo and GetChainPoint queries

NodeToClientV_11
added GetRewardInfoPools Block query

NodeToClientV_12
added LocalTxMonitor mini-protocol

newtype NodeToClientVersionData

Version data for NodeToClient protocol v1

Constructors

NodeToClientVersionData

Fields

networkMagic ∷ NetworkMagic

data NetworkConnectTracers addr vNumber

Tracer used by connectToNode (and derivatives, like connectTo or 'Ouroboros.Network.NodeToClient.connectTo).

Constructors

NetworkConnectTracers

Fields

nctMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nctHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nullNetworkConnectTracers ∷ NetworkConnectTracers addr vNumber

connectTo

Arguments

∷ LocalSnocket
callback constructed by withIOManager

→ NetworkConnectTracers LocalAddress NodeToClientVersion

→ Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplication InitiatorMode LocalAddress ByteString IO a b)
A dictionary of protocol versions & applications to run on an established connection. The application to run will be chosen by initial handshake protocol (the highest shared version will be chosen).

→ FilePath
path of the unix socket or named pipe

→ IO ()

A specialised version of connectToNode. It is a general purpose function which can connect using any version of the protocol. This is mostly useful for future enhancements.

data NetworkServerTracers addr vNumber

Tracers required by a server which handles inbound connections.

Constructors

NetworkServerTracers

Fields

nstMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nstHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nstErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nstAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

tracing rate limiting of accepting connections.

nullNetworkServerTracers ∷ NetworkServerTracers addr vNumber

data NetworkMutableState addr

Mutable state maintained by the network component.

Constructors

NetworkMutableState

Fields

nmsConnectionTable ∷ ConnectionTable IO addr

ConnectionTable which maintains information about current upstream and downstream connections.

nmsPeerStates ∷ StrictTVar IO (PeerStates IO addr)

PeerStates which maintains state of each downstream / upstream peer that errored, misbehaved or was not interesting to us.

newNetworkMutableState ∷ IO (NetworkMutableState addr)

newNetworkMutableStateSTM ∷ STM (NetworkMutableState addr)

cleanNetworkMutableState ∷ NetworkMutableState addr → IO ()

Clean PeerStates within NetworkMutableState every 200s

withServer ∷ LocalSnocket → NetworkServerTracers LocalAddress NodeToClientVersion → NetworkMutableState LocalAddress → LocalSocket → Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplication ResponderMode LocalAddress ByteString IO a b) → ErrorPolicies → IO Void

A specialised version of withServerNode.

Comments to withServer apply here as well.

type NetworkClientSubcriptionTracers = NetworkSubscriptionTracers Identity LocalAddress NodeToClientVersion

data NetworkSubscriptionTracers withIPList addr vNumber

IP subscription tracers.

Constructors

NetworkSubscriptionTracers

Fields

nsMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nsHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nsErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nsSubscriptionTracer ∷ Tracer IO (withIPList (SubscriptionTrace addr))

subscription tracers; it is infrequent it should not be nullTracer by default.

data ClientSubscriptionParams a

Constructors

ClientSubscriptionParams

Fields

cspAddress ∷ !LocalAddress

unix socket or named pipe address

cspConnectionAttemptDelay ∷ !(Maybe DiffTime)

delay between connection attempts

cspErrorPolicies ∷ !ErrorPolicies

error policies for subscription worker

ncSubscriptionWorker ∷ ∀ mode x y. HasInitiator mode ~ True ⇒ LocalSnocket → NetworkClientSubcriptionTracers → NetworkMutableState LocalAddress → ClientSubscriptionParams () → Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplication mode LocalAddress ByteString IO x y) → IO Void

ncSubscriptionWorker which starts given application versions on each established connection.

Null Protocol Peers

chainSyncPeerNull ∷ ∀ (header ∷ Type) (point ∷ Type) (tip ∷ Type) m a. MonadTimer m ⇒ Peer (ChainSync header point tip) AsClient StIdle m a

localStateQueryPeerNull ∷ ∀ (block ∷ Type) (point ∷ Type) (query ∷ Type → Type) m a. MonadTimer m ⇒ Peer (LocalStateQuery block point query) AsClient StIdle m a

localTxSubmissionPeerNull ∷ ∀ (tx ∷ Type) (reject ∷ Type) m a. MonadTimer m ⇒ Peer (LocalTxSubmission tx reject) AsClient StIdle m a

localTxMonitorPeerNull ∷ ∀ (txid ∷ Type) (tx ∷ Type) (slot ∷ Type) m a. MonadTimer m ⇒ Peer (LocalTxMonitor txid tx slot) AsClient StIdle m a

Re-exported network interface

newtype IOManager

Constructors

IOManager

Fields

associateWithIOManager ∷ ∀ hole. hole → IO ()

type AssociateWithIOCP = IOManager

withIOManager ∷ WithIOManager

type LocalSnocket = Snocket IO LocalSocket LocalAddress

System dependent LocalSnocket

localSnocket ∷ IOManager → LocalSnocket

Create a LocalSnocket.

On Windows, there is no way to get path associated to a named pipe. To go around this, the address passed to open via LocalFamily will be referenced by LocalSocket.

newtype LocalSocket

System dependent LocalSnocket type

Constructors

LocalSocket

Fields

getLocalHandle ∷ LocalHandle

newtype LocalAddress

Local address, on Unix is associated with AF_UNIX family, on

Windows with `named-pipes`.

Constructors

LocalAddress

Fields

getFilePath ∷ FilePath

Versions

newtype Versions vNum vData r

The version map supported by the local agent keyed on the version identifier.

Each Version contains a function which takes negotiated version data and returns negotiated application (the r type variable).

If one needs to combine multiple versions the simplest way is to use one of the combinators: foldMapVersions, combineVersions or the Semigroup instance directly:

fold $ (simpleSingletonVersions ...)
      :| [ (simpleSingletonVersions ...)
         , (simpleSingletonVersions ...)
         , ...
         ]

Constructors

Versions

Fields

getVersions ∷ Map vNum (Version vData r)

versionedNodeToClientProtocols ∷ NodeToClientVersion → NodeToClientVersionData → (ConnectionId LocalAddress → STM m ControlMessage → NodeToClientProtocols appType bytes m a b) → Versions NodeToClientVersion NodeToClientVersionData (OuroborosApplication appType LocalAddress bytes m a b)

Versions containing a single version of nodeToClientProtocols.

simpleSingletonVersions ∷ vNum → vData → r → Versions vNum vData r

Singleton smart constructor for Versions.

foldMapVersions ∷ (Ord vNum, Foldable f, HasCallStack) ⇒ (x → Versions vNum extra r) → f x → Versions vNum extra r

Useful for folding multiple Versions.

A foldMap restricted to the Versions Semigroup.

PRECONDITION: f x is non-empty.

combineVersions ∷ (Ord vNum, Foldable f, HasCallStack) ⇒ f (Versions vNum extra r) → Versions vNum extra r

Codecs

nodeToClientHandshakeCodec ∷ MonadST m ⇒ Codec (Handshake NodeToClientVersion Term) DeserialiseFailure m ByteString

nodeToClientVersionCodec ∷ CodecCBORTerm (Text, Maybe Int) NodeToClientVersion

We set 16ths bit to distinguish NodeToNodeVersion and NodeToClientVersion. This way connecting wrong protocol suite will fail during Handshake negotiation

This is done in backward compatible way, so NodeToClientV_1 encoding is not changed.

nodeToClientCodecCBORTerm ∷ NodeToClientVersion → CodecCBORTerm Text NodeToClientVersionData

Re-exports

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr

type LocalConnectionId = ConnectionId LocalAddress

data ErrorPolicies

List of error policies for exception handling and a policy for handing application return values.

Constructors

ErrorPolicies

Fields

epAppErrorPolicies ∷ [ErrorPolicy]

Application Error Policies

epConErrorPolicies ∷ [ErrorPolicy]

connect Error Policies

networkErrorPolicies ∷ ErrorPolicies

ErrorPolicies for client application. Additional rules can be added by means of a Semigroup instance of ErrorPolicies.

This error policies will try to preserve subscriptionWorker, e.g. if the connect function throws an IOException we will suspend it for a shortDelay, and try to re-connect.

This allows to recover from a situation where a node temporarily shutsdown, or running a client application which is subscribed two more than one node (possibly over network).

nullErrorPolicies ∷ ErrorPolicies

data ErrorPolicy where

Constructors

ErrorPolicy

Fields

∷ ∀ e. Exception e

⇒ (e → Maybe (SuspendDecision DiffTime))

Nothing means no decision. It is equivalent to not having the policy at all. In evalErrorPolicies this will fall-through and match against the remaining policies.

→ ErrorPolicy

data ErrorPolicyTrace

Trace data for error policies

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) DiffTime DiffTime
suspending peer with a given exception until

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) DiffTime
suspending consumer until

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)
caught a local exception

ErrorPolicyResumePeer
resume a peer (both consumer and producer)

ErrorPolicyKeepSuspended
consumer was suspended until producer will resume

ErrorPolicyResumeConsumer
resume consumer

ErrorPolicyResumeProducer
resume producer

ErrorPolicyUnhandledApplicationException SomeException
an application throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyUnhandledConnectionException SomeException
connect throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyAcceptException IOException
accept throwed an exception

data WithAddr addr a

Constructors

WithAddr

Fields

wiaAddr ∷ addr

wiaEvent ∷ a

data SuspendDecision t

Semigroup of commands which acts on PeerState. The t variable might be initiated to DiffTime or Time m.

This semigroup allows to either suspend both consumer and producer or just the consumer part.

Constructors

SuspendPeer !t !t
peer is suspend; The first t is the time until which a local producer is suspended, the second one is the time until which a local consumer is suspended.

SuspendConsumer !t
suspend local consumer / initiator side until t (this mean we are not allowing to communicate with the producer / responder of a remote peer).

Throw
throw an error from the main thread.

data TraceSendRecv ps where

Structured Tracer output for runPeer and derivitives.

Constructors

TraceSendMsg ∷ ∀ ps. AnyMessageAndAgency ps → TraceSendRecv ps

TraceRecvMsg ∷ ∀ ps. AnyMessageAndAgency ps → TraceSendRecv ps

data ProtocolLimitFailure

data Handshake (vNumber ∷ k) (vParams ∷ k1)

The handshake mini-protocol is used initially to agree the version and associated parameters of the protocol to use for all subsequent communication.

data LocalAddresses addr

Constructors

LocalAddresses

Fields

laIpv4 ∷ Maybe addr

Local IPv4 address to use, Nothing indicates don't use IPv4

laIpv6 ∷ Maybe addr

Local IPv6 address to use, Nothing indicates don't use IPv6

laUnix ∷ Maybe addr

Local Unix address to use, Nothing indicates don't use Unix sockets

data SubscriptionTrace addr

Constructors

SubscriptionTraceConnectStart addr

SubscriptionTraceConnectEnd addr ConnectResult

Exception e ⇒ SubscriptionTraceSocketAllocationException addr e

Exception e ⇒ SubscriptionTraceConnectException addr e

Exception e ⇒ SubscriptionTraceApplicationException addr e

SubscriptionTraceTryConnectToPeer addr

SubscriptionTraceSkippingPeer addr

SubscriptionTraceSubscriptionRunning

SubscriptionTraceSubscriptionWaiting Int

SubscriptionTraceSubscriptionFailed

SubscriptionTraceSubscriptionWaitingNewConnection DiffTime

SubscriptionTraceStart Int

SubscriptionTraceRestart DiffTime Int Int

SubscriptionTraceConnectionExist addr

SubscriptionTraceUnsupportedRemoteAddr addr

SubscriptionTraceMissingLocalAddress

SubscriptionTraceAllocateSocket addr

SubscriptionTraceCloseSocket addr

type HandshakeTr ntcAddr ntcVersion = WithMuxBearer (ConnectionId ntcAddr) (TraceSendRecv (Handshake ntcVersion Term))