Ouroboros.Network.NodeToClient

Description

This is the starting point for a module that will bring together the overall node to client protocol, as a collection of mini-protocols.

Documentation

nodeToClientProtocols ∷ NodeToClientProtocols appType bytes m a b → NodeToClientVersion → OuroborosApplication appType bytes m a b

Make an OuroborosApplication for the bundle of mini-protocols that make up the overall node-to-client protocol.

This function specifies the wire format protocol numbers as well as the protocols that run for each NodeToClientVersion.

They are chosen to not overlap with the node to node protocol numbers. This is not essential for correctness, but is helpful to allow a single shared implementation of tools that can analyse both protocols, e.g. wireshark plugins.

data NodeToClientProtocols appType bytes m a b

Recorod of node-to-client mini protocols.

Constructors

NodeToClientProtocols

Fields

localChainSyncProtocol ∷ RunMiniProtocol appType bytes m a b

local chain-sync mini-protocol

localTxSubmissionProtocol ∷ RunMiniProtocol appType bytes m a b

local tx-submission mini-protocol

localStateQueryProtocol ∷ RunMiniProtocol appType bytes m a b

local state-query mini-protocol

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_1

NodeToClientV_2
added local-query mini-protocol

newtype NodeToClientVersionData

Version data for NodeToClient protocol v1

Constructors

NodeToClientVersionData

Fields

networkMagic ∷ NetworkMagic

data DictVersion vData where

Constructors

DictVersion ∷ ∀ vData. (Typeable vData, Acceptable vData, Show vData) ⇒ CodecCBORTerm Text vData → DictVersion vData

data NetworkConnectTracers addr vNumber

Tracer used by connectToNode (and derivatives, like connectTo or 'Ouroboros.Network.NodeToClient.connectTo).

Constructors

NetworkConnectTracers

Fields

nctMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nctHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nullNetworkConnectTracers ∷ NetworkConnectTracers addr vNumber

connectTo_V1

Arguments

∷ LocalSnocket

→ NetworkConnectTracers LocalAddress NodeToClientVersion

→ NodeToClientVersionData
Client version data sent during initial handshake protocol. Client and server must agree on it.

→ (ConnectionId LocalAddress → OuroborosApplication InitiatorApp ByteString IO a b)
OuroborosInitiatorApplication which is run on an established connection using a multiplexer after the initial handshake protocol suceeds.

→ FilePath
path to unix socket or named pipe

→ IO ()

Deprecated: Use connectTo_V2

A version of connectToNode which connects using the NodeToClientV_1 version of the protocol.

connectTo_V2

Arguments

∷ LocalSnocket

→ NetworkConnectTracers LocalAddress NodeToClientVersion

→ NodeToClientVersionData
Client version data sent during initial handshake protocol. Client and server must agree on it.

→ (ConnectionId LocalAddress → OuroborosApplication InitiatorApp ByteString IO a b)
NodeToClientV_1 version of OuroborosInitiatorApplication which is run on an established connection using a multiplexer after the initial handshake protocol suceeds.

→ (ConnectionId LocalAddress → OuroborosApplication InitiatorApp ByteString IO a b)
NodeToClientV_2 version of OuroborosInitiatorApplication which is run on an established connection using a multiplexer after the initial handshake protocol suceeds. NodeToClientV_2 supports LocalStateQuery mini-protocol.

→ FilePath
path to unix socket or named pipe

→ IO ()

A version of connectToNode which connects using the NodeToClientV_1 version of the protocol.

connectTo

Arguments

∷ LocalSnocket
callback constructed by withIOManager

→ NetworkConnectTracers LocalAddress NodeToClientVersion

→ Versions NodeToClientVersion DictVersion (ConnectionId LocalAddress → OuroborosApplication InitiatorApp ByteString IO a b)
A dictionary of protocol versions & applications to run on an established connection. The application to run will be chosen by initial handshake protocol (the highest shared version will be chosen).

→ FilePath
path of the unix socket or named pipe

→ IO ()

A specialised version of connectToNode. It is a general purpose function which can connect using any version of the protocol. This is mostly useful for future enhancements.

data NetworkServerTracers addr vNumber

Tracers required by a server which handles inbound connections.

Constructors

NetworkServerTracers

Fields

nstMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nstHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nstErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nstAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

tracing rate limiting of accepting connections.

nullNetworkServerTracers ∷ NetworkServerTracers addr vNumber

data NetworkMutableState addr

Mutable state maintained by the network component.

Constructors

NetworkMutableState

Fields

nmsConnectionTable ∷ ConnectionTable IO addr

ConnectionTable which maintains information about current upstream and downstream connections.

nmsPeerStates ∷ StrictTVar IO (PeerStates IO addr)

PeerStates which maintains state of each downstream / upstream peer that errored, misbehaved or was not interesting to us.

newNetworkMutableState ∷ IO (NetworkMutableState addr)

newNetworkMutableStateSTM ∷ STM (NetworkMutableState addr)

cleanNetworkMutableState ∷ NetworkMutableState addr → IO ()

Clean PeerStates within NetworkMutableState every 200s

withServer_V1

Arguments

∷ HasResponder appType ~ True

⇒ LocalSnocket

→ NetworkServerTracers LocalAddress NodeToClientVersion

→ NetworkMutableState LocalAddress

→ LocalFD

→ NodeToClientVersionData
Client version data sent during initial handshake protocol. Client and server must agree on it.

→ (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO a b)
applications which has the reponder side, i.e. OuroborosResponderApplication or OuroborosInitiatorAndResponderApplication.

→ ErrorPolicies

→ IO Void

Deprecated: Use withServer_V2

A specialised version of withServer which can only communicate using NodeToClientV_1 version of the protocol.

withServer_V2

Arguments

∷ HasResponder appType ~ True

⇒ LocalSnocket

→ NetworkServerTracers LocalAddress NodeToClientVersion

→ NetworkMutableState LocalAddress

→ LocalFD

→ NodeToClientVersionData
Client version data sent during initial handshake protocol. Client and server must agree on it.

→ (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO a b)
NodeToClientV_1 version of applications which has the reponder side, i.e. OuroborosResponderApplication or OuroborosInitiatorAndResponderApplication.

→ (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO a b)
NodeToClientV_2 version of OuroborosApplication, which supports LocalStateQuery mini-protocol.

→ ErrorPolicies

→ IO Void

A specialised version of withServer which can only communicate using NodeToClientV_1 or NodeToClientV_2 version of the protocol.

withServer ∷ HasResponder appType ~ True ⇒ LocalSnocket → NetworkServerTracers LocalAddress NodeToClientVersion → NetworkMutableState LocalAddress → LocalFD → Versions NodeToClientVersion DictVersion (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO a b) → ErrorPolicies → IO Void

A specialised version of withServerNode; Use withServer_V1 instead of you would like to use NodeToCLientV_1 version of the protocols.

Comments to withServer apply here as well.

type NetworkClientSubcriptionTracers = NetworkSubscriptionTracers Identity LocalAddress NodeToClientVersion

data NetworkSubscriptionTracers withIPList addr vNumber

IP subscription tracers.

Constructors

NetworkSubscriptionTracers

Fields

nsMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nsHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nsErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nsSubscriptionTracer ∷ Tracer IO (withIPList (SubscriptionTrace addr))

subscription tracers; it is infrequent it should not be nullTracer by default.

data ClientSubscriptionParams a

Constructors

ClientSubscriptionParams

Fields

cspAddress ∷ !LocalAddress

unix socket or named pipe address

cspConnectionAttemptDelay ∷ !(Maybe DiffTime)

delay between connection attempts

cspErrorPolicies ∷ !ErrorPolicies

error policies for subscription worker

ncSubscriptionWorker ∷ ∀ appType x y. HasInitiator appType ~ True ⇒ LocalSnocket → NetworkClientSubcriptionTracers → NetworkMutableState LocalAddress → ClientSubscriptionParams () → Versions NodeToClientVersion DictVersion (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO x y) → IO Void

ncSubscriptionWorker which starts given application versions on each established connection.

ncSubscriptionWorker_V1 ∷ ∀ appType x y. HasInitiator appType ~ True ⇒ LocalSnocket → NetworkClientSubcriptionTracers → NetworkMutableState LocalAddress → ClientSubscriptionParams () → NodeToClientVersionData → (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO x y) → IO Void

Deprecated: Use ncSubscriptionWorker_V2

Like ncSubscriptionWorker but specific to NodeToClientV_1.

ncSubscriptionWorker_V2

Arguments

∷ HasInitiator appType ~ True

⇒ LocalSnocket

→ NetworkClientSubcriptionTracers

→ NetworkMutableState LocalAddress

→ ClientSubscriptionParams ()

→ NodeToClientVersionData

→ (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO x y)
NodeToClientV_1 version of OuroborosApplication

→ (ConnectionId LocalAddress → OuroborosApplication appType ByteString IO x y)
NodeToClientV_2 version of OuroboorsApplication, which supports LocalStateQuery mini-protocol.

→ IO Void

Like ncSubscriptionWorker but specific to NodeToClientV_2.

Null Protocol Peers

chainSyncPeerNull ∷ ∀ (header ∷ Type) (tip ∷ Type) m a. MonadTimer m ⇒ Peer (ChainSync header tip) AsClient StIdle m a

localStateQueryPeerNull ∷ ∀ (block ∷ Type) (query ∷ Type → Type) m a. MonadTimer m ⇒ Peer (LocalStateQuery block query) AsClient StIdle m a

localTxSubmissionPeerNull ∷ ∀ (tx ∷ Type) (reject ∷ Type) m a. MonadTimer m ⇒ Peer (LocalTxSubmission tx reject) AsClient StIdle m a

Re-exported network interface

newtype IOManager

This is public api to interact with the io manager; On Windows IOManager holds associateWithIOCompletionPort; on other platforms IOManager can run over any type, and thus is guaranteed to be no-op.

Constructors

IOManager

Fields

associateWithIOManager ∷ ∀ hole. hole → IO ()

type AssociateWithIOCP = IOManager

withIOManager ∷ WithIOManager

withIOManager allows to do asynchronous io on Windows hiding the differences between posix and Win32.

It starts an io manger thread, which should be only one running at a time, so the best place to call it is very close to the main function and last for duration of the application.

Async IO operatitions which are using the iocp port should not leak out-side of withIOManager. They will be silently cancelled when withIOManager exists. In particular one should not return IOManager from withIOManager.

type LocalSnocket = Snocket IO Socket LocalAddress

System dependent LocalSnocket type

localSnocket ∷ IOManager → FilePath → LocalSnocket

newtype LocalAddress

Local address, on Unix is associated with AF_UNIX family, on

Windows with `named-pipes`.

Constructors

LocalAddress

Fields

getFilePath ∷ FilePath

Versions

newtype Versions vNum (extra ∷ Type → Type) r

The set of versions supported by the local agent are described by a map keyed on the version identifier.

If one needs to combine multiple versions the simplest way is to use one of the combinators: foldMapVersions, combineVersions or the Semigroup instance directly:

fold $ (simpleSingletonVersions ...)
      :| [ (simpleSingletonVersions ...)
         , (simpleSingletonVersions ...)
         , ...
         ]

Constructors

Versions

Fields

getVersions ∷ Map vNum (Sigma (Version extra r))

versionedNodeToClientProtocols ∷ NodeToClientVersion → NodeToClientVersionData → NodeToClientProtocols appType bytes m a b → Versions NodeToClientVersion DictVersion (ConnectionId LocalAddress → OuroborosApplication appType bytes m a b)

Versions containing a single version of nodeToClientProtocols.

simpleSingletonVersions ∷ vNum → vData → extra vData → r → Versions vNum extra r

Singleton smart constructor for Versions.

foldMapVersions ∷ Ord vNum ⇒ (x → Versions vNum extra r) → NonEmpty x → Versions vNum extra r

Useful for folding a non-empty list of Versions.

A foldMap version for Semigroups which is restricted to Versions.

combineVersions ∷ Ord vNum ⇒ NonEmpty (Versions vNum extra r) → Versions vNum extra r

Codecs

nodeToClientHandshakeCodec ∷ (MonadST m, MonadThrow m) ⇒ Codec (Handshake NodeToClientVersion Term) DeserialiseFailure m ByteString

nodeToClientVersionCodec ∷ CodecCBORTerm (Text, Maybe Int) NodeToClientVersion

We set 16ths bit to distinguish NodeToNodeVersion and NodeToClientVersion. This way connectin wrong protocol suite will fail during Handshake negotation

This is done in backward compatible way, so NodeToClientV_1 encoding is not changed.

nodeToClientCodecCBORTerm ∷ CodecCBORTerm Text NodeToClientVersionData

Re-exports

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr

type LocalConnectionId = ConnectionId LocalAddress

data ErrorPolicies

List of error policies for exception handling and a policy for handing application return values.

Constructors

ErrorPolicies

Fields

epAppErrorPolicies ∷ [ErrorPolicy]

Application Error Policies

epConErrorPolicies ∷ [ErrorPolicy]

connect Error Policies

networkErrorPolicies ∷ ErrorPolicies

ErrorPolicies for client application. Additional rules can be added by means of a Semigroup instance of ErrorPolicies.

This error policies will try to preserve subscriptionWorker, e.g. if the connect function throws an IOException we will suspend it for a shortDelay, and try to re-connect.

This allows to recover from a situation where a node temporarily shutsdown, or running a client application which is subscribed two more than one node (possibly over network).

nullErrorPolicies ∷ ErrorPolicies

data ErrorPolicy where

Constructors

ErrorPolicy

Fields

∷ Exception e

⇒ (e → Maybe (SuspendDecision DiffTime))

Nothing means no decision. It is equivalent to not having the policy at all. In evalErrorPolicies this will fall-through and match against the remaining policies.

→ ErrorPolicy

data ErrorPolicyTrace

Trace data for error policies

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime !DiffTime
suspending peer with a given exception until

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime
suspending consumer until

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)
caught a local exception

ErrorPolicyResumePeer
resume a peer (both consumer and producer)

ErrorPolicyKeepSuspended
consumer was suspended until producer will resume

ErrorPolicyResumeConsumer
resume consumer

ErrorPolicyResumeProducer
resume producer

ErrorPolicyUnhandledApplicationException SomeException
an application throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyUnhandledConnectionException SomeException
connect throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyAcceptException IOException
accept throwed an exception

data WithAddr addr a

Constructors

WithAddr

Fields

wiaAddr ∷ !addr

wiaEvent ∷ !a

data SuspendDecision t

Semigroup of commands which acts on PeerState. The t variable might be initiated to DiffTime or Time m.

This semigroup allows to either suspend both consumer and producer or just the consumer part.

Constructors

SuspendPeer !t !t
peer is suspend; The first t is the time until which a local producer is suspended, the second one is the time until which a local consumer is suspended.

SuspendConsumer !t
suspend local consumer / initiator side until t (this mean we are not allowing to communicate with the producer / responder of a remote peer).

Throw
throw an error from the main thread.

data TraceSendRecv ps

Structured Tracer output for runPeer and derivitives.

Constructors

TraceSendMsg (AnyMessage ps)

TraceRecvMsg (AnyMessage ps)

data ProtocolLimitFailure

data Handshake (vNumber ∷ k) (vParams ∷ k1) ∷ ∀ k k1. k → k1 → Type

The handshake mini-protocol is used initially to agree the version and associated parameters of the protocol to use for all subsequent communication.

data LocalAddresses addr

Constructors

LocalAddresses

Fields

laIpv4 ∷ Maybe addr

Local IPv4 address to use, Nothing indicates don't use IPv4

laIpv6 ∷ Maybe addr

Local IPv6 address to use, Nothing indicates don't use IPv6

laUnix ∷ Maybe addr

Local Unix address to use, Nothing indicates don't use Unix sockets

data SubscriptionTrace addr where

Constructors

SubscriptionTraceConnectStart ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceConnectEnd ∷ ∀ addr. addr → ConnectResult → SubscriptionTrace addr

SubscriptionTraceSocketAllocationException ∷ ∀ addr e. Exception e ⇒ addr → e → SubscriptionTrace addr

SubscriptionTraceConnectException ∷ ∀ addr e. Exception e ⇒ addr → e → SubscriptionTrace addr

SubscriptionTraceApplicationException ∷ ∀ addr e. Exception e ⇒ addr → e → SubscriptionTrace addr

SubscriptionTraceTryConnectToPeer ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceSkippingPeer ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceSubscriptionRunning ∷ ∀ addr. SubscriptionTrace addr

SubscriptionTraceSubscriptionWaiting ∷ ∀ addr. Int → SubscriptionTrace addr

SubscriptionTraceSubscriptionFailed ∷ ∀ addr. SubscriptionTrace addr

SubscriptionTraceSubscriptionWaitingNewConnection ∷ ∀ addr. DiffTime → SubscriptionTrace addr

SubscriptionTraceStart ∷ ∀ addr. Int → SubscriptionTrace addr

SubscriptionTraceRestart ∷ ∀ addr. DiffTime → Int → Int → SubscriptionTrace addr

SubscriptionTraceConnectionExist ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceUnsupportedRemoteAddr ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceMissingLocalAddress ∷ ∀ addr. SubscriptionTrace addr

SubscriptionTraceAllocateSocket ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceCloseSocket ∷ ∀ addr. addr → SubscriptionTrace addr

type HandshakeTr = WithMuxBearer (ConnectionId LocalAddress) (TraceSendRecv (Handshake NodeToClientVersion Term))