Ouroboros.Network.NodeToNode

Description

This is the starting point for a module that will bring together the overall node to node protocol, as a collection of mini-protocols.

Documentation

nodeToNodeProtocols ∷ MiniProtocolParameters → NodeToNodeProtocols appType bytes m a b → OuroborosApplication appType bytes m a b

Make an OuroborosApplication for the bundle of mini-protocols that make up the overall node-to-node protocol.

This function specifies the wire format protocol numbers.

The application specific protocol numbers start from 2. The MiniProtocolNum 0 is reserved for the Handshake protocol, while MiniProtocolNum 1 is reserved for DeltaQ messages. Handshake protocol is not included in NodeToNodeProtocols as it runs before mux is started but it reusing MuxBearer to send and receive messages. Only when the handshake protocol suceedes, we will know which protocols to run / multiplex.

These are chosen to not overlap with the node to client protocol numbers (and the handshake protocol number). This is not essential for correctness, but is helpful to allow a single shared implementation of tools that can analyse both protocols, e.g. wireshark plugins.

data NodeToNodeProtocols appType bytes m a b

Constructors

NodeToNodeProtocols

Fields

chainSyncProtocol ∷ RunMiniProtocol appType bytes m a b

chain-sync mini-protocol

blockFetchProtocol ∷ RunMiniProtocol appType bytes m a b

block-fetch mini-protocol

txSubmissionProtocol ∷ RunMiniProtocol appType bytes m a b

tx-submission mini-protocol

data MiniProtocolParameters

Constructors

MiniProtocolParameters

Fields

chainSyncPipelineingHighMark ∷ !Word32

high threshold for pipelining (we will never exceed that many messages pipelined).

chainSyncPipelineingLowMark ∷ !Word32

low threshold: if we hit the chainSyncPipelineingHighMark we will listen for responses until there are at most chainSyncPipelineingLowMark pipelined message

Must be smaller than chainSyncPipelineingHighMark.

Note: chainSyncPipelineingLowMark and chainSyncPipelineingLowMark are passed to pipelineDecisionLowHighMark.

blockFetchPipelineingMax ∷ !Word

maximal number of pipelined messages in 'block-fetch' mini-protocol.

txSubmissionMaxUnacked ∷ !Word16

maximal number of unacked tx (pipelining is bounded by twice this number)

defaultMiniProtocolParameters ∷ MiniProtocolParameters

data NodeToNodeVersion

Enumeration of node to node protocol versions.

Constructors

NodeToNodeV_1

newtype NodeToNodeVersionData

Version data for NodeToNode protocol v1

Constructors

NodeToNodeVersionData

Fields

networkMagic ∷ NetworkMagic

data DictVersion vData where

Constructors

DictVersion ∷ ∀ vData. (Typeable vData, Acceptable vData, Show vData) ⇒ CodecCBORTerm Text vData → DictVersion vData

nodeToNodeCodecCBORTerm ∷ CodecCBORTerm Text NodeToNodeVersionData

data NetworkConnectTracers addr vNumber

Tracer used by connectToNode (and derivatives, like connectTo or 'Ouroboros.Network.NodeToClient.connectTo).

Constructors

NetworkConnectTracers

Fields

nctMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nctHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nullNetworkConnectTracers ∷ NetworkConnectTracers addr vNumber

connectTo ∷ Snocket IO Socket SockAddr → NetworkConnectTracers SockAddr NodeToNodeVersion → Versions NodeToNodeVersion DictVersion (ConnectionId SockAddr → OuroborosApplication InitiatorApp ByteString IO a b) → Maybe SockAddr → SockAddr → IO ()

A specialised version of connectToNode.

connectTo_V1 ∷ SocketSnocket → NetworkConnectTracers SockAddr NodeToNodeVersion → NodeToNodeVersionData → (ConnectionId SockAddr → OuroborosApplication InitiatorApp ByteString IO a b) → Maybe SockAddr → SockAddr → IO ()

Like connectTo but specific to NodeToNodeV_1.

data NetworkServerTracers addr vNumber

Tracers required by a server which handles inbound connections.

Constructors

NetworkServerTracers

Fields

nstMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nstHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nstErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nstAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

tracing rate limiting of accepting connections.

nullNetworkServerTracers ∷ NetworkServerTracers addr vNumber

data NetworkMutableState addr

Mutable state maintained by the network component.

Constructors

NetworkMutableState

Fields

nmsConnectionTable ∷ ConnectionTable IO addr

ConnectionTable which maintains information about current upstream and downstream connections.

nmsPeerStates ∷ StrictTVar IO (PeerStates IO addr)

PeerStates which maintains state of each downstream / upstream peer that errored, misbehaved or was not interesting to us.

data AcceptedConnectionsLimit

Policy which governs how to limit the number of accepted connections.

Constructors

AcceptedConnectionsLimit

Fields

acceptedConnectionsHardLimit ∷ !Word32

Hard limit of accepted connections.

acceptedConnectionsSoftLimit ∷ !Word32

Soft limit of accepted connections. If we are above this threshold, we will start rate limiting.

acceptedConnectionsDelay ∷ !DiffTime

Max delay for limiting accepted connections. We use linear regression starting from 0 at the soft limit up to acceptedConnectionDelay at the hard limit.

newNetworkMutableState ∷ IO (NetworkMutableState addr)

newNetworkMutableStateSTM ∷ STM (NetworkMutableState addr)

cleanNetworkMutableState ∷ NetworkMutableState addr → IO ()

Clean PeerStates within NetworkMutableState every 200s

withServer ∷ HasResponder appType ~ True ⇒ SocketSnocket → NetworkServerTracers SockAddr NodeToNodeVersion → NetworkMutableState SockAddr → AcceptedConnectionsLimit → SockAddr → Versions NodeToNodeVersion DictVersion (ConnectionId SockAddr → OuroborosApplication appType ByteString IO a b) → ErrorPolicies → IO Void

A specialised version of withServerNode. It forks a thread which runs an accept loop (server thread):

when the server thread throws an exception the main thread rethrows it (by wait)

when an async exception is thrown to kill the main thread the server thread will be cancelled as well (by withAsync)

withServer_V1 ∷ HasResponder appType ~ True ⇒ SocketSnocket → NetworkServerTracers SockAddr NodeToNodeVersion → NetworkMutableState SockAddr → AcceptedConnectionsLimit → SockAddr → NodeToNodeVersionData → (ConnectionId SockAddr → OuroborosApplication appType ByteString IO x y) → ErrorPolicies → IO Void

Like withServer but specific to NodeToNodeV_1.

Subscription Workers

IP subscriptin worker

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps ∷ ![SockAddr]

List of destinations to possibly connect to

ispValency ∷ !Int

Number of parallel connections to keep actice.

type NetworkIPSubscriptionTracers addr vNumber = NetworkSubscriptionTracers WithIPList addr vNumber

data NetworkSubscriptionTracers withIPList addr vNumber

IP subscription tracers.

Constructors

NetworkSubscriptionTracers

Fields

nsMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

nsHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

nsErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

nsSubscriptionTracer ∷ Tracer IO (withIPList (SubscriptionTrace addr))

subscription tracers; it is infrequent it should not be nullTracer by default.

nullNetworkSubscriptionTracers ∷ NetworkSubscriptionTracers withIPList addr vNumber

data SubscriptionParams a target

ipSubscriptionWorker and dnsSubscriptionWorker parameters

Constructors

SubscriptionParams

Fields

spLocalAddresses ∷ LocalAddresses SockAddr

spConnectionAttemptDelay ∷ SockAddr → Maybe DiffTime

should return expected delay for the given address

spErrorPolicies ∷ ErrorPolicies

spSubscriptionTarget ∷ target

type IPSubscriptionParams a = SubscriptionParams a IPSubscriptionTarget

ipSubscriptionWorker ∷ ∀ appType x y. HasInitiator appType ~ True ⇒ SocketSnocket → NetworkIPSubscriptionTracers SockAddr NodeToNodeVersion → NetworkMutableState SockAddr → IPSubscriptionParams () → Versions NodeToNodeVersion DictVersion (ConnectionId SockAddr → OuroborosApplication appType ByteString IO x y) → IO Void

ipSubscriptionWorker which starts given application versions on each established connection.

ipSubscriptionWorker_V1 ∷ ∀ appType x y. HasInitiator appType ~ True ⇒ SocketSnocket → NetworkIPSubscriptionTracers SockAddr NodeToNodeVersion → NetworkMutableState SockAddr → IPSubscriptionParams () → NodeToNodeVersionData → (ConnectionId SockAddr → OuroborosApplication appType ByteString IO x y) → IO Void

Like ipSubscriptionWorker but specific to NodeToNodeV_1.

DNS subscription worker

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain ∷ !Domain

dstPort ∷ !PortNumber

dstValency ∷ !Int

type DnsSubscriptionParams a = SubscriptionParams a DnsSubscriptionTarget

data NetworkDNSSubscriptionTracers vNumber addr

DNS subscription tracers.

Constructors

NetworkDNSSubscriptionTracers

Fields

ndstMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) MuxTrace)

low level mux-network tracer, which logs mux sdu (send and received) and other low level multiplexing events.

ndstHandshakeTracer ∷ Tracer IO (WithMuxBearer (ConnectionId addr) (TraceSendRecv (Handshake vNumber Term)))

handshake protocol tracer; it is important for analysing version negotation mismatches.

ndstErrorPolicyTracer ∷ Tracer IO (WithAddr addr ErrorPolicyTrace)

error policy tracer; must not be nullTracer, otherwise all the exceptions which are not matched by any error policy will be caught and not logged or rethrown.

ndstSubscriptionTracer ∷ Tracer IO (WithDomainName (SubscriptionTrace addr))

subscription tracer; it is infrequent it should not be nullTracer by default.

ndstDnsTracer ∷ Tracer IO (WithDomainName DnsTrace)

dns resolver tracer; it is infrequent it should not be nullTracer by default.

nullNetworkDNSSubscriptionTracers ∷ NetworkDNSSubscriptionTracers vNumber peerid

dnsSubscriptionWorker ∷ ∀ appType x y. HasInitiator appType ~ True ⇒ SocketSnocket → NetworkDNSSubscriptionTracers NodeToNodeVersion SockAddr → NetworkMutableState SockAddr → DnsSubscriptionParams () → Versions NodeToNodeVersion DictVersion (ConnectionId SockAddr → OuroborosApplication appType ByteString IO x y) → IO Void

dnsSubscriptionWorker which starts given application versions on each established connection.

dnsSubscriptionWorker_V1 ∷ ∀ appType x y. HasInitiator appType ~ True ⇒ SocketSnocket → NetworkDNSSubscriptionTracers NodeToNodeVersion SockAddr → NetworkMutableState SockAddr → DnsSubscriptionParams () → NodeToNodeVersionData → (ConnectionId SockAddr → OuroborosApplication appType ByteString IO x y) → IO Void

Like dnsSubscriptionWorker but specific to NodeToNodeV_1.

Re-exports

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr

type RemoteConnectionId = ConnectionId SockAddr

data ProtocolLimitFailure

data Handshake (vNumber ∷ k) (vParams ∷ k1) ∷ ∀ k k1. k → k1 → Type

The handshake mini-protocol is used initially to agree the version and associated parameters of the protocol to use for all subsequent communication.

data LocalAddresses addr

Constructors

LocalAddresses

Fields

laIpv4 ∷ Maybe addr

Local IPv4 address to use, Nothing indicates don't use IPv4

laIpv6 ∷ Maybe addr

Local IPv6 address to use, Nothing indicates don't use IPv6

laUnix ∷ Maybe addr

Local Unix address to use, Nothing indicates don't use Unix sockets

Error Policies and Peer state

data ErrorPolicies

List of error policies for exception handling and a policy for handing application return values.

Constructors

ErrorPolicies

Fields

epAppErrorPolicies ∷ [ErrorPolicy]

Application Error Policies

epConErrorPolicies ∷ [ErrorPolicy]

connect Error Policies

remoteNetworkErrorPolicy ∷ ErrorPolicies

A minimal error policy for remote peers, which only handles exceptions raised by `ouroboros-network`.

localNetworkErrorPolicy ∷ ErrorPolicies

Error policy for local clients. This is equivalent to nullErrorPolicies, but explicit in the errors which can be catched.

We are very permissive here, and very strict in the networkErrorPolicy. After any failure the client will be killed and not penalised by this policy. This allows to restart the local client without a delay.

nullErrorPolicies ∷ ErrorPolicies

data ErrorPolicy where

Constructors

ErrorPolicy

Fields

∷ Exception e

⇒ (e → Maybe (SuspendDecision DiffTime))

Nothing means no decision. It is equivalent to not having the policy at all. In evalErrorPolicies this will fall-through and match against the remaining policies.

→ ErrorPolicy

data SuspendDecision t

Semigroup of commands which acts on PeerState. The t variable might be initiated to DiffTime or Time m.

This semigroup allows to either suspend both consumer and producer or just the consumer part.

Constructors

SuspendPeer !t !t
peer is suspend; The first t is the time until which a local producer is suspended, the second one is the time until which a local consumer is suspended.

SuspendConsumer !t
suspend local consumer / initiator side until t (this mean we are not allowing to communicate with the producer / responder of a remote peer).

Throw
throw an error from the main thread.

Traces

data AcceptConnectionsPolicyTrace

Trace for the AcceptConnectionsLimit policy.

Constructors

ServerTraceAcceptConnectionRateLimiting DiffTime Int

ServerTraceAcceptConnectionHardLimit Word32

data TraceSendRecv ps

Structured Tracer output for runPeer and derivitives.

Constructors

TraceSendMsg (AnyMessage ps)

TraceRecvMsg (AnyMessage ps)

data SubscriptionTrace addr where

Constructors

SubscriptionTraceConnectStart ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceConnectEnd ∷ ∀ addr. addr → ConnectResult → SubscriptionTrace addr

SubscriptionTraceSocketAllocationException ∷ ∀ addr e. Exception e ⇒ addr → e → SubscriptionTrace addr

SubscriptionTraceConnectException ∷ ∀ addr e. Exception e ⇒ addr → e → SubscriptionTrace addr

SubscriptionTraceApplicationException ∷ ∀ addr e. Exception e ⇒ addr → e → SubscriptionTrace addr

SubscriptionTraceTryConnectToPeer ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceSkippingPeer ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceSubscriptionRunning ∷ ∀ addr. SubscriptionTrace addr

SubscriptionTraceSubscriptionWaiting ∷ ∀ addr. Int → SubscriptionTrace addr

SubscriptionTraceSubscriptionFailed ∷ ∀ addr. SubscriptionTrace addr

SubscriptionTraceSubscriptionWaitingNewConnection ∷ ∀ addr. DiffTime → SubscriptionTrace addr

SubscriptionTraceStart ∷ ∀ addr. Int → SubscriptionTrace addr

SubscriptionTraceRestart ∷ ∀ addr. DiffTime → Int → Int → SubscriptionTrace addr

SubscriptionTraceConnectionExist ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceUnsupportedRemoteAddr ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceMissingLocalAddress ∷ ∀ addr. SubscriptionTrace addr

SubscriptionTraceAllocateSocket ∷ ∀ addr. addr → SubscriptionTrace addr

SubscriptionTraceCloseSocket ∷ ∀ addr. addr → SubscriptionTrace addr

data DnsTrace

Constructors

DnsTraceLookupException SomeException

DnsTraceLookupAError DNSError

DnsTraceLookupAAAAError DNSError

DnsTraceLookupIPv6First

DnsTraceLookupIPv4First

DnsTraceLookupAResult [SockAddr]

DnsTraceLookupAAAAResult [SockAddr]

data ErrorPolicyTrace

Trace data for error policies

Constructors

ErrorPolicySuspendPeer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime !DiffTime
suspending peer with a given exception until

ErrorPolicySuspendConsumer (Maybe (ConnectionOrApplicationExceptionTrace SomeException)) !DiffTime
suspending consumer until

ErrorPolicyLocalNodeError (ConnectionOrApplicationExceptionTrace SomeException)
caught a local exception

ErrorPolicyResumePeer
resume a peer (both consumer and producer)

ErrorPolicyKeepSuspended
consumer was suspended until producer will resume

ErrorPolicyResumeConsumer
resume consumer

ErrorPolicyResumeProducer
resume producer

ErrorPolicyUnhandledApplicationException SomeException
an application throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyUnhandledConnectionException SomeException
connect throwed an exception, which was not handled by any ErrorPolicy.

ErrorPolicyAcceptException IOException
accept throwed an exception

data WithIPList a

Constructors

WithIPList

Fields

wilSrc ∷ !(LocalAddresses SockAddr)

wilDsts ∷ ![SockAddr]

wilEvent ∷ !a

data WithDomainName a

Constructors

WithDomainName

Fields

wdnDomain ∷ !Domain

wdnEvent ∷ !a

data WithAddr addr a

Constructors

WithAddr

Fields

wiaAddr ∷ !addr

wiaEvent ∷ !a

type HandshakeTr = WithMuxBearer (ConnectionId SockAddr) (TraceSendRecv (Handshake NodeToNodeVersion Term))