Ouroboros.Network.PeerSelection.Governor.ActivePeers

Documentation

belowTarget ∷ ∀ peeraddr peerconn m. (MonadSTM m, Ord peeraddr) ⇒ PeerSelectionActions peeraddr peerconn m → MkGuardedDecision peeraddr peerconn m

If we are below the target of hot peers we promote some of the /warm peers/ according to policyPickWarmPeersToPromote policy.

aboveTarget ∷ ∀ peeraddr peerconn m. (MonadSTM m, Ord peeraddr) ⇒ PeerSelectionActions peeraddr peerconn m → MkGuardedDecision peeraddr peerconn m

If we are above the target of hot peers we demote some of the /warm peers/, according to policyPickHotPeersToDemote.

jobDemoteActivePeer ∷ ∀ peeraddr peerconn m. (Monad m, Ord peeraddr) ⇒ PeerSelectionActions peeraddr peerconn m → peeraddr → peerconn → Job m (Completion m peeraddr peerconn)