Ouroboros.Network.PeerSelection.Governor.Monitor

Description

This module contains governor decisions for monitoring tasks:

monitoring target peer changes

monitoring results governor's job results

monitoring connections

Documentation

targetPeers ∷ MonadSTM m ⇒ PeerSelectionActions peeraddr peerconn m → PeerSelectionState peeraddr peerconn → Guarded (STM m) (Decision m peeraddr peerconn)

Monitor PeerSelectionTargets, if they change, we just need to update PeerSelectionState, since we return it in a Decision action it will be picked by the governor's peerSelectionGovernorLoop.

jobs ∷ MonadSTM m ⇒ JobPool m (Completion m peeraddr peerconn) → PeerSelectionState peeraddr peerconn → Time → Guarded (STM m) (Decision m peeraddr peerconn)

Await for the first result from JobPool and return its Decision.

connections ∷ ∀ m peeraddr peerconn. (MonadSTM m, Ord peeraddr) ⇒ PeerSelectionActions peeraddr peerconn m → PeerSelectionState peeraddr peerconn → Guarded (STM m) (Decision m peeraddr peerconn)

Monitor connections.

localRoots ∷ ∀ peeraddr peerconn m. (MonadSTM m, Ord peeraddr) ⇒ PeerSelectionActions peeraddr peerconn m → PeerSelectionState peeraddr peerconn → Guarded (STM m) (Decision m peeraddr peerconn)

Monitor local roots using readLocalRootPeers STM action.