Ouroboros.Network.PeerSelection.Governor.Types

Documentation

type PickPolicy peeraddr m = Map peeraddr KnownPeerInfo → Int → STM m (Set peeraddr)

A peer pick policy is an action that picks a subset of elements from a map of peers.

The pre-condition is that the map of available choices will be non-empty, and the requested number to pick will be strictly positive.

The post-condition is that the picked set is non-empty but must not be bigger than the requested number.

pickPeers ∷ (Ord peeraddr, Functor m, HasCallStack) ⇒ (Map peeraddr a → Int → m (Set peeraddr)) → Map peeraddr a → Int → m (Set peeraddr)

Check pre-conditions and post-conditions on the pick policies

data PeerSelectionPolicy peeraddr m

Constructors

PeerSelectionPolicy

Fields

policyPickKnownPeersForGossip ∷ PickPolicy peeraddr m

policyPickColdPeersToPromote ∷ PickPolicy peeraddr m

policyPickWarmPeersToPromote ∷ PickPolicy peeraddr m

policyPickHotPeersToDemote ∷ PickPolicy peeraddr m

policyPickWarmPeersToDemote ∷ PickPolicy peeraddr m

policyPickColdPeersToForget ∷ PickPolicy peeraddr m

policyFindPublicRootTimeout ∷ !DiffTime

policyMaxInProgressGossipReqs ∷ !Int

policyGossipRetryTime ∷ !DiffTime

policyGossipBatchWaitTime ∷ !DiffTime

policyGossipOverallTimeout ∷ !DiffTime

data PeerSelectionTargets

Adjustable targets for the peer selection mechanism.

These are used by the peer selection governor as targets. They are used by the peer churn governor loop as knobs to adjust, to influence the peer selection governor.

The known, established and active peer targets are targets both from below and from above: the governor will attempt to grow or shrink the sets to hit these targets.

Unlike the other targets, the root peer target is "one sided", it is only a target from below. The governor does not try to shrink the root set to hit it, it simply stops looking for more.

Constructors

PeerSelectionTargets

Fields

targetNumberOfRootPeers ∷ !Int

targetNumberOfKnownPeers ∷ !Int

targetNumberOfEstablishedPeers ∷ !Int

targetNumberOfActivePeers ∷ !Int

nullPeerSelectionTargets ∷ PeerSelectionTargets

sanePeerSelectionTargets ∷ PeerSelectionTargets → Bool

data PeerSelectionActions peeraddr peerconn m

Actions performed by the peer selection governor.

These being pluggable allows:

choice of known peer root sets

running both in simulation and for real

Constructors

PeerSelectionActions

Fields

readPeerSelectionTargets ∷ STM m PeerSelectionTargets

readLocalRootPeers ∷ STM m (Map peeraddr PeerAdvertise)

Read the current set of locally or privately known root peers.

In general this is expected to be updated asynchronously by some other thread. It is intended to cover the use case of peers from local configuration. It could be dynamic due to DNS resolution, or due to dynamic configuration updates.

requestPublicRootPeers ∷ Int → m (Set peeraddr, DiffTime)

Request a sample of public root peers.

It is intended to cover use cases including:

federated relays from a DNS pool

stake pool relays published in the blockchain

a pre-distributed snapshot of stake pool relays from the blockchain

requestPeerGossip ∷ peeraddr → m [peeraddr]

The action to contact a known peer and request a sample of its known peers.

This is synchronous, but it should expect to be interrupted by a timeout asynchronous exception. Failures are throw as exceptions.

establishPeerConnection ∷ peeraddr → m peerconn

monitorPeerConnection ∷ peerconn → STM m PeerStatus

activatePeerConnection ∷ peerconn → m ()

deactivatePeerConnection ∷ peerconn → m ()

closePeerConnection ∷ peerconn → m ()

data PeerSelectionState peeraddr peerconn

The internal state used by the peerSelectionGovernor.

The local and public root sets are disjoint, and their union is the overall root set.

Constructors

PeerSelectionState

Fields

targets ∷ !PeerSelectionTargets

localRootPeers ∷ !(Map peeraddr PeerAdvertise)

The current set of local root peers.

publicRootPeers ∷ !(Set peeraddr)

knownPeers ∷ !(KnownPeers peeraddr)

establishedPeers ∷ !(Map peeraddr peerconn)

establishedStatus ∷ !(Map peeraddr PeerStatus)

activePeers ∷ !(Set peeraddr)

publicRootBackoffs ∷ !Int

A counter to manage the exponential backoff strategy for when to retry querying for more public root peers. It is negative for retry counts after failure, and positive for retry counts that are successful but make no progress.

publicRootRetryTime ∷ !Time

The earliest time we would be prepared to request more public root peers. This is used with the publicRootBackoffs to manage the exponential backoff.

inProgressPublicRootsReq ∷ !Bool

inProgressGossipReqs ∷ !Int

inProgressPromoteCold ∷ !(Set peeraddr)

inProgressPromoteWarm ∷ !(Set peeraddr)

inProgressDemoteWarm ∷ !(Set peeraddr)

inProgressDemoteHot ∷ !(Set peeraddr)

emptyPeerSelectionState ∷ PeerSelectionState peeraddr peerconn

invariantPeerSelectionState ∷ Ord peeraddr ⇒ PeerSelectionState peeraddr peerconn → Bool

data Guarded m a

The governor is using Guarded m (Decision m peeraddr peerconn) where m is an STM monad, to drive its progress.

Constructors

GuardedSkip !(Maybe (Min Time))
GuardedSkip is used to instruct that there is no action to be made by the governor.

Let us note that the combined value which is computed by guardedDecisions term in peerSelectionGovernorLoop will never return it: this is bacause there are monitoring decisions which never return this constructor, e.g. targetPeers, jobs, connections, and thus the governor has always something to do.

Guarded !(Maybe (Min Time)) (m a)
Guarded is used to provide an action, possibly with a timeout, to the governor main loop.

data Decision m peeraddr peerconn

Constructors

Decision

Fields

decisionTrace ∷ TracePeerSelection peeraddr

A trace event to classify the decision and action

decisionState ∷ PeerSelectionState peeraddr peerconn

An updated state to use immediately

decisionJobs ∷ [Job m (Completion m peeraddr peerconn)]

An optional Job to execute asynchronously. This job leads to a further Decision. This gives a state update to apply upon completion, but also allows chaining further job actions.

type MkGuardedDecision peeraddr peerconn m = PeerSelectionPolicy peeraddr m → PeerSelectionState peeraddr peerconn → Guarded (STM m) (Decision m peeraddr peerconn)

Type alias for function types which are used to create governor decisions. Allmost all decisions are following this pattern.

newtype Completion m peeraddr peerconn

Constructors

Completion (PeerSelectionState peeraddr peerconn → Time → Decision m peeraddr peerconn)

data TracePeerSelection peeraddr

Constructors

TraceLocalRootPeersChanged (Map peeraddr PeerAdvertise) (Map peeraddr PeerAdvertise)

TraceTargetsChanged PeerSelectionTargets PeerSelectionTargets

TracePublicRootsRequest Int Int

TracePublicRootsResults (Set peeraddr) Int DiffTime

TracePublicRootsFailure SomeException Int DiffTime

TraceGossipRequests Int Int (Set peeraddr) (Set peeraddr)

TraceGossipResults [(peeraddr, Either SomeException [peeraddr])]

TraceForgetColdPeers Int Int (Set peeraddr)

TracePromoteColdPeers Int Int (Set peeraddr)

TracePromoteColdFailed peeraddr SomeException

TracePromoteColdDone peeraddr

TracePromoteWarmPeers Int Int (Set peeraddr)

TracePromoteWarmFailed peeraddr SomeException

TracePromoteWarmDone peeraddr

TraceDemoteWarmPeers Int Int (Set peeraddr)

TraceDemoteWarmFailed peeraddr SomeException

TraceDemoteWarmDone peeraddr

TraceDemoteHotPeers Int Int (Set peeraddr)

TraceDemoteHotFailed peeraddr SomeException

TraceDemoteHotDone peeraddr

TraceDemoteAsynchronous (Map peeraddr PeerStatus)

TraceGovernorWakeup

data DebugPeerSelection peeraddr peerconn

Constructors

TraceGovernorState (PeerSelectionState peeraddr peerconn) (Maybe DiffTime)