Ouroboros.Network.PeerSelection.LedgerPeers

Documentation

data DomainAccessPoint

Constructors

DomainAccessPoint

Fields

dapDomain :: !Domain

dapPortNumber :: !PortNumber

data IP

Constructors

IPv4

Fields

ipv4 :: IPv4

IPv6

Fields

ipv6 :: IPv6

newtype LedgerPeersConsensusInterface m

Constructors

LedgerPeersConsensusInterface

Fields

lpGetPeers :: SlotNo -> STM m (Maybe [(PoolStake, NonEmpty RelayAccessPoint)])

data RelayAccessPoint

Constructors

RelayAccessDomain !Domain !PortNumber

RelayAccessAddress !IP !PortNumber

Bundled Patterns

pattern RelayDomainAccessPoint :: DomainAccessPoint -> RelayAccessPoint

newtype PoolStake

Constructors

PoolStake

Fields

unPoolStake :: Rational

newtype AccPoolStake

Constructors

AccPoolStake

Fields

unAccPoolStake :: Rational

data TraceLedgerPeers

Trace LedgerPeers events.

Constructors

PickedBigLedgerPeer RelayAccessPoint AccPoolStake PoolStake
Trace for a significant ledger peer picked with accumulated and relative stake of its pool.

PickedLedgerPeer RelayAccessPoint AccPoolStake PoolStake
Trace for a ledger peer picked with accumulated and relative stake of its pool.

PickedBigLedgerPeers NumberOfPeers [RelayAccessPoint]

PickedLedgerPeers NumberOfPeers [RelayAccessPoint]
Trace for the number of peers and we wanted to pick and the list of peers picked.

FetchingNewLedgerState Int Int
Trace for fetching a new list of peers from the ledger. The first Int is the number of ledger peers returned the latter is the number of big ledger peers.

DisabledLedgerPeers
Trace for when getting peers from the ledger is disabled, that is DontUseLedger.

TraceUseLedgerAfter UseLedgerAfter
Trace UseLedgerAfter value

WaitingOnRequest

RequestForPeers NumberOfPeers

ReusingLedgerState Int DiffTime

FallingBackToPublicRootPeers

NotEnoughBigLedgerPeers NumberOfPeers Int

NotEnoughLedgerPeers NumberOfPeers Int

newtype NumberOfPeers

Constructors

NumberOfPeers

Fields

getNumberOfPeers :: Word16

data LedgerPeersKind

Which ledger peers to pick.

Constructors

AllLedgerPeers

BigLedgerPeers

accPoolStake :: [(PoolStake, NonEmpty RelayAccessPoint)] -> Map AccPoolStake (PoolStake, NonEmpty RelayAccessPoint)

Convert a list of pools with stake to a Map keyed on the accumulated stake. Consensus provides a list of pairs of relative stake and corresponding relays for all usable registered pools. By creating a Map keyed on the AccPoolStake that is the sum of the pool's relative stake and the stake of all preceding pools we can support weighted random selection in O(log n) time by taking advantage of Map.lookupGE (returns the smallest key greater or equal to the provided value).

accBigPoolStake :: [(PoolStake, NonEmpty RelayAccessPoint)] -> Map AccPoolStake (PoolStake, NonEmpty RelayAccessPoint)

Convert a list of pools with stake to a Map keyed on the accumulated stake which only contains big ledger peers, e.g. largest ledger peers which cumulatively control 90% of stake.

withLedgerPeers :: forall peerAddr m a. (MonadAsync m, MonadMonotonicTime m, Ord peerAddr) => StdGen -> (IP -> PortNumber -> peerAddr) -> Tracer m TraceLedgerPeers -> STM m UseLedgerAfter -> LedgerPeersConsensusInterface m -> ([DomainAccessPoint] -> m (Map DomainAccessPoint (Set peerAddr))) -> ((NumberOfPeers -> LedgerPeersKind -> m (Maybe (Set peerAddr, DiffTime))) -> Async m Void -> m a) -> m a

For a LedgerPeers worker thread and submit request and receive responses.

data UseLedgerAfter

Only use the ledger after the given slot number.

Constructors

DontUseLedger

UseLedgerAfter SlotNo

data IsLedgerPeer

Identifies a peer as coming from ledger or not

Constructors

IsLedgerPeer
a ledger peer.

IsNotLedgerPeer

data IsBigLedgerPeer

Constructors

IsBigLedgerPeer

IsNotBigLedgerPeer

data PortNumber

bigLedgerPeerQuota :: AccPoolStake

The total accumulated stake of big ledger peers.