Ouroboros.Network.PeerSelection.LedgerPeers

Documentation

data DomainAddress

A product of a Domain and PortNumber. After resolving the domain we will use the PortNumber to form SockAddr.

Constructors

DomainAddress

Fields

daDomain ∷ !Domain

daPortNumber ∷ !PortNumber

data IP

Constructors

IPv4

Fields

ipv4 ∷ IPv4

IPv6

Fields

ipv6 ∷ IPv6

newtype LedgerPeersConsensusInterface m

Constructors

LedgerPeersConsensusInterface

Fields

lpGetPeers ∷ SlotNo → STM m (Maybe [(PoolStake, NonEmpty RelayAddress)])

data RelayAddress

Constructors

RelayAddressDomain DomainAddress

RelayAddressAddr IP PortNumber

newtype PoolStake

The relative stake of a stakepool in relation to the total amount staked. A value in the [0, 1] range.

Constructors

PoolStake

Fields

unPoolStake ∷ Rational

newtype AccPoolStake

The accumulated relative stake of a stake pool, like PoolStake but it also includes the relative stake of all preceding pools. A value in the range [0, 1].

Constructors

AccPoolStake

Fields

unAccPoolStake ∷ Rational

data TraceLedgerPeers

Trace LedgerPeers events.

Constructors

PickedPeer !RelayAddress !AccPoolStake !PoolStake
Trace for a peer picked with accumulated and relative stake of its pool.

PickedPeers !Word16 ![RelayAddress]
Trace for the number of peers we wanted to pick and the list of peers picked.

FetchingNewLedgerState !Int
Trace for fetching a new list of peers from the ledger. Int is the number of peers returned.

pickPeers ∷ ∀ m. Monad m ⇒ StdGen → Tracer m TraceLedgerPeers → Map AccPoolStake (PoolStake, NonEmpty RelayAddress) → Word16 → m (StdGen, [RelayAddress])

Try to pick n random peers.

accPoolStake ∷ [(PoolStake, NonEmpty RelayAddress)] → Map AccPoolStake (PoolStake, NonEmpty RelayAddress)

Convert a list of pools with stake to a Map keyed on the accumulated stake. Consensus provides a list of pairs of relative stake and corresponding relays for all usable registered pools. By creating a Map keyed on the AccPoolStake that is the sum of the pool's relative stake and the stake of all preceding pools we can support weighted random selection in O(log n) time by taking advantage of Map.lookupGE (returns the smallest key greater or equal to the provided value).

data PortNumber