Ouroboros.Network.PeerSelection.LedgerPeers

Documentation

data DomainAccessPoint

A product of a Domain and PortNumber. After resolving the domain we will use the PortNumber to form SockAddr.

Constructors

DomainAccessPoint

Fields

dapDomain ∷ !Domain

dapPortNumber ∷ !PortNumber

data IP

Constructors

IPv4

Fields

ipv4 ∷ IPv4

IPv6

Fields

ipv6 ∷ IPv6

newtype LedgerPeersConsensusInterface m

Constructors

LedgerPeersConsensusInterface

Fields

lpGetPeers ∷ SlotNo → STM m (Maybe [(PoolStake, NonEmpty RelayAccessPoint)])

data RelayAccessPoint

A relay can have either an IP address and a port number or a domain with a port number

Constructors

RelayAccessDomain !Domain !PortNumber

RelayAccessAddress !IP !PortNumber

Bundled Patterns

pattern RelayDomainAccessPoint ∷ DomainAccessPoint → RelayAccessPoint
RelayDomainAccessPoint a bidirectional pattern which links RelayAccessDomain and DomainAccessPoint.

newtype PoolStake

The relative stake of a stakepool in relation to the total amount staked. A value in the [0, 1] range.

Constructors

PoolStake

Fields

unPoolStake ∷ Rational

newtype AccPoolStake

The accumulated relative stake of a stake pool, like PoolStake but it also includes the relative stake of all preceding pools. A value in the range [0, 1].

Constructors

AccPoolStake

Fields

unAccPoolStake ∷ Rational

data TraceLedgerPeers

Trace LedgerPeers events.

Constructors

PickedPeer RelayAccessPoint AccPoolStake PoolStake
Trace for a peer picked with accumulated and relative stake of its pool.

PickedPeers NumberOfPeers [RelayAccessPoint]
Trace for the number of peers we wanted to pick and the list of peers picked.

FetchingNewLedgerState Int
Trace for fetching a new list of peers from the ledger. Int is the number of peers returned.

DisabledLedgerPeers
Trace for when getting peers from the ledger is disabled, that is DontUseLedger.

TraceUseLedgerAfter UseLedgerAfter
Trace UseLedgerAfter value

WaitingOnRequest

RequestForPeers NumberOfPeers

ReusingLedgerState Int DiffTime

FallingBackToBootstrapPeers

newtype NumberOfPeers

Constructors

NumberOfPeers Word16

accPoolStake ∷ [(PoolStake, NonEmpty RelayAccessPoint)] → Map AccPoolStake (PoolStake, NonEmpty RelayAccessPoint)

Convert a list of pools with stake to a Map keyed on the accumulated stake. Consensus provides a list of pairs of relative stake and corresponding relays for all usable registered pools. By creating a Map keyed on the AccPoolStake that is the sum of the pool's relative stake and the stake of all preceding pools we can support weighted random selection in O(log n) time by taking advantage of Map.lookupGE (returns the smallest key greater or equal to the provided value).

withLedgerPeers ∷ ∀ peerAddr m a. (MonadAsync m, MonadTime m, Ord peerAddr) ⇒ StdGen → (IP → PortNumber → peerAddr) → Tracer m TraceLedgerPeers → STM m UseLedgerAfter → LedgerPeersConsensusInterface m → ([DomainAccessPoint] → m (Map DomainAccessPoint (Set peerAddr))) → ((NumberOfPeers → m (Maybe (Set peerAddr, DiffTime))) → Async m Void → m a) → m a

For a LederPeers worker thread and submit request and receive responses.

data UseLedgerAfter

Only use the ledger after the given slot number.

Constructors

DontUseLedger

UseLedgerAfter SlotNo

data PortNumber