Ouroboros.Network.PeerSelection.RootPeersDNS

DNS based provider for local root peers

localRootPeersProvider ∷ Tracer IO TraceLocalRootPeers → TimeoutFn IO → ResolvConf → StrictTVar IO (Map DomainAddress (Map SockAddr PeerAdvertise)) → [(DomainAddress, PeerAdvertise)] → IO ()

This action typically runs indefinitely, but can terminate successfully in corner cases where there is nothing to do.

data DomainAddress

A product of a Domain and PortNumber. After resolving the domain we will use the PortNumber to form SockAddr.

Constructors

DomainAddress

Fields

daDomain ∷ !Domain

daPortNumber ∷ !PortNumber

data RelayAddress

A relay can have either an IP address and a port number or a domain with a port number. TODO: move to a Ledger Peer file.

Constructors

RelayAddressDomain DomainAddress

RelayAddressAddr IP PortNumber

data IP

Constructors

IPv4

Fields

ipv4 ∷ IPv4

IPv6

Fields

ipv6 ∷ IPv6

data TraceLocalRootPeers

Constructors

TraceLocalRootDomains [(DomainAddress, PeerAdvertise)]

TraceLocalRootWaiting DomainAddress DiffTime

TraceLocalRootResult DomainAddress [(IPv4, TTL)]

TraceLocalRootFailure DomainAddress DNSorIOError

DNS based provider for public root peers

publicRootPeersProvider ∷ Tracer IO TracePublicRootPeers → TimeoutFn IO → ResolvConf → [DomainAddress] → ((Int → IO (Set SockAddr, DiffTime)) → IO a) → IO a

data TracePublicRootPeers

Constructors

TracePublicRootDomains [DomainAddress]

TracePublicRootResult Domain [(IPv4, TTL)]

TracePublicRootFailure Domain DNSError

DNS type re-exports

data ResolvConf

type Domain = ByteString

type TTL = Word32

Socket type re-exports

data PortNumber