Ouroboros.Network.PeerSelection.RootPeersDNS

DNS based actions for local and public root providers

data DNSActions resolver exception m

Dictionary of DNS actions vocabulary

Constructors

DNSActions

Fields

dnsResolverResource ∷ ResolvConf → m (Resource m (DNSorIOError exception) resolver)

TODO: it could be useful for publicRootPeersProvider.

dnsAsyncResolverResource ∷ ResolvConf → m (Resource m (DNSorIOError exception) resolver)

Resource which passes the Resolver (or abstract resolver type) through a StrictTVar. Better than resolverResource when using in multiple threads.

On Windows we use a different implementation which always returns a newly initialised Resolver at each step. This is because on Windows we don't have a way to check that the network configuration has changed. The dns library is using 'GetNetworkParams@ win32 api call to get the list of default dns servers.

dnsLookupWithTTL ∷ ResolvConf → resolver → Domain → m ([DNSError], [(IP, TTL)])

Like lookupA but also return the TTL for the results.

DNS library timeouts do not work reliably on Windows (#1873), hence the additional timeout.

DNS resolver IO auxiliar functions

constantResource ∷ Applicative m ⇒ a → Resource m err a

DNSActions IO

ioDNSActions ∷ LookupReqs → DNSActions Resolver IOException IO

Bundle of DNS Actions that runs in IO The IPv4 and IPv6 addresses the node will be using should determine the LookupReqs so that we can avoid lookups for address types that wont be used.

data LookupReqs

Constructors

LookupReqAOnly

LookupReqAAAAOnly

LookupReqAAndAAAA

DNS based provider for local root peers

localRootPeersProvider

Arguments

∷ ∀ m peerAddr resolver exception. (MonadAsync m, MonadDelay m, Eq (Async m Void), Ord peerAddr)

⇒ Tracer m (TraceLocalRootPeers peerAddr exception)

→ (IP → PortNumber → peerAddr)

→ ResolvConf

→ DNSActions resolver exception m

→ STM m [(Int, Map RelayAccessPoint PeerAdvertise)]
input

→ StrictTVar m (Seq (Int, Map peerAddr PeerAdvertise))
output TVar

→ m Void

Resolve RelayAddress-es of local root peers using dns if needed. Local roots are provided wrapped in a StrictTVar, which value might change (re-read form a config file). The resolved dns names are available through the output StrictTVar.

data DomainAccessPoint

A product of a Domain and PortNumber. After resolving the domain we will use the PortNumber to form SockAddr.

Constructors

DomainAccessPoint

Fields

dapDomain ∷ !Domain

dapPortNumber ∷ !PortNumber

data RelayAccessPoint

A relay can have either an IP address and a port number or a domain with a port number

Constructors

RelayAccessDomain !Domain !PortNumber

RelayAccessAddress !IP !PortNumber

Bundled Patterns

pattern RelayDomainAccessPoint ∷ DomainAccessPoint → RelayAccessPoint
RelayDomainAccessPoint a bidirectional pattern which links RelayAccessDomain and DomainAccessPoint.

data IP

Constructors

IPv4

Fields

ipv4 ∷ IPv4

IPv6

Fields

ipv6 ∷ IPv6

data TraceLocalRootPeers peerAddr exception

Constructors

TraceLocalRootDomains [(Int, Map RelayAccessPoint PeerAdvertise)]
Int is the configured valency for the local producer groups

TraceLocalRootWaiting DomainAccessPoint DiffTime

TraceLocalRootResult DomainAccessPoint [(IP, TTL)]

TraceLocalRootGroups (Seq (Int, Map peerAddr PeerAdvertise))
This traces the results of the local root peer provider

TraceLocalRootFailure DomainAccessPoint (DNSorIOError exception)

TraceLocalRootError DomainAccessPoint SomeException

DNS based provider for public root peers

publicRootPeersProvider ∷ ∀ peerAddr resolver exception a m. (MonadThrow m, MonadAsync m, Exception exception, Ord peerAddr) ⇒ Tracer m TracePublicRootPeers → (IP → PortNumber → peerAddr) → ResolvConf → STM m [RelayAccessPoint] → DNSActions resolver exception m → ((Int → m (Set peerAddr, DiffTime)) → m a) → m a

TODO track PeerAdvertise

data TracePublicRootPeers

Constructors

TracePublicRootRelayAccessPoint [RelayAccessPoint]

TracePublicRootDomains [DomainAccessPoint]

TracePublicRootResult Domain [(IP, TTL)]

TracePublicRootFailure Domain DNSError

resolveDomainAccessPoint ∷ ∀ exception resolver m. (MonadThrow m, MonadAsync m, Exception exception) ⇒ Tracer m TracePublicRootPeers → ResolvConf → DNSActions resolver exception m → [DomainAccessPoint] → m (Map DomainAccessPoint (Set SockAddr))

Provides DNS resolution functionality.

DNS type re-exports

data ResolvConf

type Domain = ByteString

type TTL = Word32

Socket type re-exports

data PortNumber