Ouroboros.Network.PeerSelection.RootPeersDNS

DNS based provider for local root peers

localRootPeersProvider ∷ Tracer IO TraceLocalRootPeers → ResolvConf → TVar IO (Map Domain (Map IPv4 PeerAdvertise)) → [(Domain, PeerAdvertise)] → IO ()

This action typically runs indefinitely, but can terminate successfully in corner cases where there is nothing to do.

data TraceLocalRootPeers

Constructors

TraceLocalRootDomains [(Domain, PeerAdvertise)]

TraceLocalRootWaiting Domain DiffTime

TraceLocalRootResult Domain [(IPv4, TTL)]

TraceLocalRootFailure Domain DNSError

DNS based provider for public root peers

publicRootPeersProvider ∷ Tracer IO TracePublicRootPeers → ResolvConf → [Domain] → ((Int → IO (Set IPv4, DiffTime)) → IO a) → IO a

data TracePublicRootPeers

Constructors

TracePublicRootDomains [Domain]

TracePublicRootResult Domain [(IPv4, TTL)]

TracePublicRootFailure Domain DNSError

DNS type re-exports

data ResolvConf

type Domain = ByteString

type TTL = Word32

data IPv4

The abstract data type to express an IPv4 address. To create this, use toIPv4. Or use read "192.0.2.1" :: IP, for example. Also, "192.0.2.1" can be used as literal with OverloadedStrings.

>>> read "192.0.2.1" :: IPv4
192.0.2.1