Ouroboros.Network.PeerSelection.Types

Documentation

data PeerSource

Where did this peer come from? Policy functions can choose to treat peers differently depending on where we found them from.

Constructors

PeerSourceLocalRoot

PeerSourcePublicRoot

PeerSourceGossip

PeerSourceStaleRoot

data PeerAdvertise

Should this peer be advertised to other peers asking for known peers? For certain peers specified by configuration it would be an appropriate policy to keep them private.

Constructors

DoAdvertisePeer

DoNotAdvertisePeer

data PeerStatus

Constructors

PeerCold

PeerWarm

PeerHot