Ouroboros.Network.Protocol.ChainSync.Examples

Documentation

chainSyncClientExample ∷ ∀ header tip m a. (HasHeader header, MonadSTM m) ⇒ StrictTVar m (Chain header) → Client header tip m a → ChainSyncClient header tip m a

An instance of the client side of the chain sync protocol that consumes into a Chain stored in a StrictTVar.

This is of course only useful in tests and reference implementations since this is not a realistic chain representation.

data Client header tip m t

Constructors

Client

Fields

rollbackward ∷ Point header → tip → m (Either t (Client header tip m t))

rollforward ∷ header → m (Either t (Client header tip m t))

points ∷ [Point header] → m (Client header tip m t)

pureClient ∷ Applicative m ⇒ Client header tip m void

A client which doesn't do anything and never ends. Used with chainSyncClientExample, the StrictTVar m (Chain header) will be updated but nothing further will happen.

data Tip b

Used in chain-sync protocol to advertise the tip of the server's chain.

Constructors

TipGenesis
The tip is genesis

Tip !SlotNo !(HeaderHash b) !BlockNo
The tip is not genesis

chainSyncServerExample ∷ ∀ blk header m a. (HasHeader header, MonadSTM m, HeaderHash header ~ HeaderHash blk) ⇒ a → StrictTVar m (ChainProducerState header) → ChainSyncServer header (Tip blk) m a

An instance of the server side of the chain sync protocol that reads from a pure ChainProducerState stored in a StrictTVar.

This is of course only useful in tests and reference implementations since this is not a realistic chain representation.