Ouroboros.Network.Protocol.LocalStateQuery.Client

Protocol type for the client

The protocol states from the point of view of the client.

newtype LocalStateQueryClient block query m a

Constructors

LocalStateQueryClient

Fields

runLocalStateQueryClient ∷ ClientStIdle block query m a

data ClientStIdle block query (m ∷ ★ → ★) a where

In the StIdle protocol state, the client has agency and must send:

a request to acquire a state

a termination messge

Constructors

SendMsgAcquire ∷ Point block → ClientStAcquiring block query m a → ClientStIdle block query m a

SendMsgDone ∷ a → ClientStIdle block query m a

data ClientStAcquiring block query m a

In the StAcquiring protocol state, the client does not have agency. Instead it is waiting for:

acquired

failure to acquire

It must be prepared to handle either.

Constructors

ClientStAcquiring

Fields

recvMsgAcquired ∷ ClientStAcquired block query m a

recvMsgFailure ∷ AcquireFailure → m (ClientStIdle block query m a)

data ClientStAcquired block query m a where

In the StAcquired protocol state, the client has agency and must send:

a query

a request to (re)acquire another state

a release of the current state

Constructors

SendMsgQuery ∷ query result → ClientStQuerying block query m a result → ClientStAcquired block query m a

SendMsgReAcquire ∷ Point block → ClientStAcquiring block query m a → ClientStAcquired block query m a

SendMsgRelease ∷ ClientStIdle block query m a → ClientStAcquired block query m a

data ClientStQuerying block query m a result

In the StQuerying protocol state, the client does not have agency. Instead it is waiting for:

a result

Constructors

ClientStQuerying

Fields

recvMsgResult ∷ result → m (ClientStAcquired block query m a)

Execution as a typed protocol

localStateQueryClientPeer ∷ ∀ block query m a. Monad m ⇒ LocalStateQueryClient block query m a → Peer (LocalStateQuery block query) AsClient StIdle m a

Interpret a LocalStateQueryClient action sequence as a Peer on the client side of the LocalStateQuery protocol.