Ouroboros.Network.Protocol.LocalStateQuery.Examples

Documentation

localStateQueryClient ∷ ∀ block query result m. Applicative m ⇒ [(Point block, query result)] → LocalStateQueryClient block query m [(Point block, Either AcquireFailure result)]

An example LocalStateQueryClient, which, for each point in the given list, acquires the state for that point, and if that succeeds, returns the result for the corresponding query. When the state could not be acquired, the AcquireFailure is returned instead of the query results.

localStateQueryServer ∷ ∀ block query m state. Applicative m ⇒ (Point block → Either AcquireFailure state) → (∀ result. state → query result → result) → LocalStateQueryServer block query m ()

An example LocalStateQueryServer. The first function is called to acquire a state, after which the second will be used to query the state.