Ouroboros.Network.Protocol.LocalTxMonitor.Type

Description

The type of the local transaction monitoring protocol.

This is used by local clients (like wallets, explorers and CLI tools) to monitor the transactions passing through the mempool of a local node.

Documentation

data LocalTxMonitor tx where

The kind of the local transaction monitoring protocol, and the types of the states in the protocol state machine.

It is parametrised over the type of transactions.

Constructors

StIdle ∷ LocalTxMonitor tx
The client has agency; it can request a transaction or terminate.

There is no timeout in this state.

StBusy ∷ LocalTxMonitor tx
The server has agency; it must (eventually) return the next transaction that has not yet been sent to the client.

There is no timeout in this state.

StDone ∷ LocalTxMonitor tx
Nobody has agency. The terminal state.