Ouroboros.Network.Protocol.LocalTxSubmission.Client

Description

A view of the transaction submission protocol from the point of view of the client.

This provides a view that uses less complex types and should be easier to use than the underlying typed protocol itself.

For execution, localTxSubmissionClientPeer is provided for conversion into the typed protocol.

Protocol type for the client

The protocol states from the point of view of the client.

newtype LocalTxSubmissionClient tx reject m a

Constructors

LocalTxSubmissionClient

Fields

runLocalTxSubmissionClient ∷ m (LocalTxClientStIdle tx reject m a)

data LocalTxClientStIdle tx reject m a where

The client side of the local transaction submission protocol.

The peer in the client role submits transactions to the peer in the server role.

Constructors

SendMsgSubmitTx ∷ tx → (Maybe reject → m (LocalTxClientStIdle tx reject m a)) → LocalTxClientStIdle tx reject m a
The client submits a single transaction and waits a reply.

The server replies to inform the client that it has either accepted the transaction or rejected it. In the rejection case a reason for the rejection is included.

SendMsgDone ∷ a → LocalTxClientStIdle tx reject m a
The client can terminate the protocol.

Execution as a typed protocol

localTxSubmissionClientPeer ∷ ∀ tx reject m a. Monad m ⇒ LocalTxSubmissionClient tx reject m a → Peer (LocalTxSubmission tx reject) AsClient StIdle m a

A non-pipelined Peer representing the LocalTxSubmissionClient.

Null local tx submission client

localTxSubmissionClientNull ∷ MonadTimer m ⇒ LocalTxSubmissionClient tx reject m a

A local tx submission client which never sends any message.